% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisApi.r
\name{analysis$delete}
\alias{analysis$delete}
\title{Delete an Analysis.}
\arguments{
\item{webId}{The ID of the Analysis to delete.}
}
\value{
The Analysis was deleted.
}
\description{
Delete an Analysis.
}
