% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisApi.r
\name{analysis$update}
\alias{analysis$update}
\title{Update an Analysis.}
\arguments{
\item{webId}{The ID of the Analysis to update.}

\item{PIAnalysis}{A partial Analysis containing the desired changes.}
}
\value{
The Analysis was updated.
}
\description{
Update an Analysis.
}
