% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisCategoryApi.r
\name{analysisCategory$delete}
\alias{analysisCategory$delete}
\title{Delete an analysis category.}
\arguments{
\item{webId}{The ID of the analysis category to delete.}
}
\value{
The analysis category was deleted.
}
\description{
Delete an analysis category.
}
