% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisCategoryApi.r
\name{analysisCategory$update}
\alias{analysisCategory$update}
\title{Update an analysis category by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the analysis category to update.}

\item{PIAnalysisCategory}{A partial analysis category containing the desired changes.}
}
\value{
The analysis category was updated.
}
\description{
Update an analysis category by replacing items in its definition.
}
