% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRuleApi.r
\name{analysisRule$delete}
\alias{analysisRule$delete}
\title{Delete an Analysis Rule.}
\arguments{
\item{webId}{The ID of the Analysis Rule.}

\item{column}{the value of the user label column}
}
\value{
The Analysis Rule was deleted.
}
\description{
Delete an Analysis Rule.
}
