% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRuleApi.r
\name{analysisRule$get}
\alias{analysisRule$get}
\title{Retrieve an Analysis Rule.}
\arguments{
\item{webId}{The ID of the Analysis Rule.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested Analysis Rule.
}
\description{
Retrieve an Analysis Rule.
}
