% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRuleApi.r
\name{analysisRule$update}
\alias{analysisRule$update}
\title{Update an Analysis Rule by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the Analysis Rule.}

\item{analysisRule}{A partial Analysis Rule containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The Analysis Rule was updated.
}
\description{
Update an Analysis Rule by replacing items in its definition.
}
