% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRulePlugInApi.r
\name{analysisRulePlugIn$getByPath}
\alias{analysisRulePlugIn$getByPath}
\title{Retrieve an Analysis Rule Plug-in by path.}
\arguments{
\item{path}{The path to the Analysis Rule Plug-in.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified Analysis Rule Plug-in.
}
\description{
Retrieve an Analysis Rule Plug-in by path.
}
