% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisTemplateApi.r
\name{analysisTemplate$delete}
\alias{analysisTemplate$delete}
\title{Delete an analysis template.}
\arguments{
\item{webId}{The ID of the analysis template to update.}
}
\value{
The analysis template was deleted.
}
\description{
Delete an analysis template.
}
