% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisTemplateApi.r
\name{analysisTemplate$get}
\alias{analysisTemplate$get}
\title{Retrieve an analysis template.}
\arguments{
\item{webId}{The ID of the analysis template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified analysis template.
}
\description{
Retrieve an analysis template.
}
