% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisTemplateApi.r
\name{analysisTemplate$getCategories}
\alias{analysisTemplate$getCategories}
\title{Get an analysis template's categories.}
\arguments{
\item{webId}{The ID of the analysis template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of categories.
}
\description{
Get an analysis template's categories.
}
