% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisTemplateApi.r
\name{analysisTemplate$update}
\alias{analysisTemplate$update}
\title{Update an analysis template by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the analysis template to update.}

\item{PIAnalysisTemplate}{A partial analysis template containing the desired changes.}
}
\value{
The analysis template was updated.
}
\description{
Update an analysis template by replacing items in its definition.
}
