% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$createEventFrame}
\alias{assetDatabase$createEventFrame}
\title{Create an event frame.}
\arguments{
\item{webId}{The ID of the database on which to create the event frame.}

\item{PIEventFrame}{The new event frame definition.}
}
\value{
The event frame was created. The response's Location header is a link to the event frame.
}
\description{
Create an event frame.
}
