% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$createTable}
\alias{assetDatabase$createTable}
\title{Create a table on the Asset Database.}
\arguments{
\item{webId}{The ID of the database in which to create the table.}

\item{PITable}{The new table definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The table was created. The response's Location header is a link to the created table.
}
\description{
Create a table on the Asset Database.
}
