% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$delete}
\alias{assetDatabase$delete}
\title{Delete an asset database.}
\arguments{
\item{webId}{The ID of the database.}
}
\value{
The database was deleted.
}
\description{
Delete an asset database.
}
