% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$deleteSecurityEntry}
\alias{assetDatabase$deleteSecurityEntry}
\title{Delete a security entry owned by the asset database.}
\arguments{
\item{name}{The name of the security entry. For every backslash character (\\) in the security entry name, replace with asterisk (*). As an example, use domain*username instead of domain\\username.}

\item{webId}{The ID of the asset database where the security entry will be deleted.}

\item{applyToChildren}{If false, the new access permissions are only applied to the associated object. If true, the access permissions of children with any parent-child reference types will change when the permissions on the primary parent change.}

\item{securityItem}{The security item of the desired security entries to be deleted. If the parameter is not specified, security entries of the 'Default' security item will be deleted.}
}
\value{
The security entry was deleted.
}
\description{
Delete a security entry owned by the asset database.
}
