% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$export}
\alias{assetDatabase$export}
\title{Export the asset database.}
\arguments{
\item{webId}{The ID of the database.}

\item{endTime}{The latest ending time for AFEventFrame, AFTransfer, and AFCase objects that may be part of the export. The default is '*'.}

\item{exportMode}{Indicates the type of export to perform. The default is 'StrongReferences'. Multiple export modes may be specified by using multiple instances of exportMode.}

\item{startTime}{The earliest starting time for AFEventFrame, AFTransfer, and AFCase objects that may be part of the export. The default is '*-30d'.}
}
\value{
Database exported. The response body contains the serialized database.
}
\description{
Export the asset database.
}
