% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$findElementAttributes}
\alias{assetDatabase$findElementAttributes}
\title{Retrieves a list of element attributes matching the specified filters from the specified asset database.}
\arguments{
\item{webId}{The ID of the asset database to use as the root of the search.}

\item{attributeCategory}{Specify that returned attributes must have this category. The default is no filter.}

\item{attributeDescriptionFilter}{The attribute description filter string used for finding objects. Only the first 440 characters of the description will be searched. For Asset Servers older than 2.7, a 400 status code (Bad Request) will be returned if this parameter is specified. The default is no filter.}

\item{attributeNameFilter}{The attribute name filter string used for finding objects. The default is no filter.}

\item{attributeType}{Specify that returned attributes' value type must be this value type. The default is no filter.}

\item{elementCategory}{Specify that the owner of the returned attributes must have this category. The default is no filter.}

\item{elementDescriptionFilter}{The element description filter string used for finding objects. Only the first 440 characters of the description will be searched. For Asset Servers older than 2.7, a 400 status code (Bad Request) will be returned if this parameter is specified. The default is no filter.}

\item{elementNameFilter}{The element name filter string used for finding objects. The default is no filter.}

\item{elementTemplate}{Specify that the owner of the returned attributes must have this template or a template derived from this template. The default is no filter.}

\item{elementType}{Specify that the element of the returned attributes must have this AFElementType. The default is no filter.}

\item{maxCount}{The maximum number of objects to be returned (the page size). The default is 1000.}

\item{searchFullHierarchy}{Specifies if the search should include objects nested further than immediate children of the given resource. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
a list of matching attributes.
}
\description{
Retrieves a list of element attributes matching the specified filters from the specified asset database.
}
