% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$findEventFrameAttributes}
\alias{assetDatabase$findEventFrameAttributes}
\title{Retrieves a list of event frame attributes matching the specified filters from the specified asset database.}
\arguments{
\item{webId}{The ID of the asset database to use as the root of the search.}

\item{attributeCategory}{Specify that returned attributes must have this category. The default is no filter.}

\item{attributeDescriptionFilter}{The attribute description filter string used for finding objects. Only the first 440 characters of the description will be searched. For Asset Servers older than 2.7, a 400 status code (Bad Request) will be returned if this parameter is specified. The default is no filter.}

\item{attributeNameFilter}{The attribute name filter string used for finding objects. The default is no filter.}

\item{attributeType}{Specify that returned attributes' value type must be this value type. The default is no filter.}

\item{endTime}{A string representing the latest ending time for the event frames to be matched. The endTime must be greater than or equal to the startTime. The default is '*'.}

\item{eventFrameCategory}{Specify that the owner of the returned attributes must have this category. The default is no filter.}

\item{eventFrameDescriptionFilter}{The event frame description filter string used for finding objects. Only the first 440 characters of the description will be searched. For Asset Servers older than 2.7, a 400 status code (Bad Request) will be returned if this parameter is specified. The default is no filter.}

\item{eventFrameNameFilter}{The event frame name filter string used for finding objects. The default is no filter.}

\item{eventFrameTemplate}{Specify that the owner of the returned attributes must have this template or a template derived from this template. The default is no filter.}

\item{maxCount}{The maximum number of objects to be returned (the page size). The default is 1000.}

\item{referencedElementNameFilter}{The name query string which must match the name of a referenced element. The default is no filter.}

\item{searchFullHierarchy}{Specifies if the search should include objects nested further than immediate children of the given resource. The default is 'false'.}

\item{searchMode}{Determines how the startTime and endTime parameters are treated when searching for event frames. The default is 'Overlapped'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{startTime}{A string representing the earliest starting time for the event frames to be matched. startTime must be less than or equal to the endTime. The default is '*-8h'.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
a list of matching attributes.
}
\description{
Retrieves a list of event frame attributes matching the specified filters from the specified asset database.
}
