% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$getAttributeCategories}
\alias{assetDatabase$getAttributeCategories}
\title{Retrieve attribute categories for a given Asset Database.}
\arguments{
\item{webId}{The ID of the database.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The attribute categories that exist in the database.
}
\description{
Retrieve attribute categories for a given Asset Database.
}
