% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$getEnumerationSets}
\alias{assetDatabase$getEnumerationSets}
\title{Retrieve enumeration sets for given asset database.}
\arguments{
\item{webId}{The ID of the database.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified enumeration sets.
}
\description{
Retrieve enumeration sets for given asset database.
}
