% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$getTableCategories}
\alias{assetDatabase$getTableCategories}
\title{Retrieve table categories for a given Asset Database.}
\arguments{
\item{webId}{The ID of the database.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The table categories on the specified database.
}
\description{
Retrieve table categories for a given Asset Database.
}
