% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$getTables}
\alias{assetDatabase$getTables}
\title{Retrieve tables for given Asset Database.}
\arguments{
\item{webId}{The ID of the database.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The tables on the specified database.
}
\description{
Retrieve tables for given Asset Database.
}
