% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$import}
\alias{assetDatabase$import}
\title{Import an asset database.}
\arguments{
\item{webId}{The ID of the asset database.}

\item{importMode}{Indicates the type of import to perform. The default is 'AllowCreate | AllowUpdate | AutoCheckIn'. Multiple import modes may be specified by using multiple instances of importMode.}
}
\value{
Database imported.
}
\description{
Import an asset database.
}
