% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$update}
\alias{assetDatabase$update}
\title{Update an asset database by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the database.}

\item{PIAssetDatabase}{A partial database containing the desired changes.}
}
\value{
The database was updated.
}
\description{
Update an asset database by replacing items in its definition.
}
