% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$get}
\alias{assetServer$get}
\title{Retrieve an Asset Server.}
\arguments{
\item{webId}{The ID of the server.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested server.
}
\description{
Retrieve an Asset Server.
}
