% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getAnalysisRulePlugIns}
\alias{assetServer$getAnalysisRulePlugIns}
\title{Retrieve a list of all Analysis Rule Plug-in's.}
\arguments{
\item{webId}{The ID of the asset server, where the Analysis Rule Plug-in's are installed.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of Analysis Rule Plug-in's.
}
\description{
Retrieve a list of all Analysis Rule Plug-in's.
}
