% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getByPath}
\alias{assetServer$getByPath}
\title{Retrieve an Asset Server by path.}
\arguments{
\item{path}{The path to the server.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested server.
}
\description{
Retrieve an Asset Server by path.
}
