% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getDatabases}
\alias{assetServer$getDatabases}
\title{Retrieve a list of all Asset Databases on the specified Asset Server.}
\arguments{
\item{webId}{The ID of the server.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of databases.
}
\description{
Retrieve a list of all Asset Databases on the specified Asset Server.
}
