% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getSecurity}
\alias{assetServer$getSecurity}
\title{Get the security information of the specified security item associated with the asset server for a specified user.}
\arguments{
\item{webId}{The ID of the asset server for the security to be checked.}

\item{securityItem}{The security item of the desired security information to be returned. Multiple security items may be specified with multiple instances of the parameter. If the parameter is not specified, only 'Default' security item of the security information will be returned.}

\item{userIdentity}{The user identity for the security information to be checked. Multiple security identities may be specified with multiple instances of the parameter. If the parameter is not specified, only the current user's security rights will be returned.}

\item{forceRefresh}{Indicates if the security cache should be refreshed before getting security information. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Security rights.
}
\description{
Get the security information of the specified security item associated with the asset server for a specified user.
}
