% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getSecurityIdentities}
\alias{assetServer$getSecurityIdentities}
\title{Retrieve security identities based on the specified criteria. By default, all security identities in the specified Asset Server are returned.}
\arguments{
\item{webId}{The ID of the asset server to search.}

\item{field}{Specifies which of the object's properties are searched. The default is 'Name'.}

\item{maxCount}{The maximum number of objects to be returned. The default is 1000.}

\item{query}{The query string used for finding objects. The default is no query string.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of security identities matching the specified condition.
}
\description{
Retrieve security identities based on the specified criteria. By default, all security identities in the specified Asset Server are returned.
}
