% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getSecurityIdentitiesForUser}
\alias{assetServer$getSecurityIdentitiesForUser}
\title{Retrieve security identities for a specific user.}
\arguments{
\item{webId}{The ID of the server.}

\item{userIdentity}{The user identity to search for.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of security identities for the specified user.
}
\description{
Retrieve security identities for a specific user.
}
