% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$list}
\alias{assetServer$list}
\title{Retrieve a list of all Asset Servers known to this service.}
\arguments{
\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of servers.
}
\description{
Retrieve a list of all Asset Servers known to this service.
}
