% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$updateSecurityEntry}
\alias{assetServer$updateSecurityEntry}
\title{Update a security entry owned by the asset server.}
\arguments{
\item{name}{The name of the security entry.}

\item{webId}{The ID of the asset server where the security entry will be updated.}

\item{PISecurityEntry}{The new security entry definition. The full list of allow and deny rights must be supplied or they will be removed.}

\item{applyToChildren}{If false, the new access permissions are only applied to the associated object. If true, the access permissions of children with any parent-child reference types will change when the permissions on the primary parent change.}

\item{securityItem}{The security item of the desired security entries to be updated. If the parameter is not specified, security entries of the 'Default' security item will be updated.}
}
\value{
The security entry was updated.
}
\description{
Update a security entry owned by the asset server.
}
