% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeApi.r
\name{attribute$createAttribute}
\alias{attribute$createAttribute}
\title{Create a new attribute as a child of the specified attribute.}
\arguments{
\item{webId}{The ID of the parent attribute on which to create the attribute.}

\item{PIAttribute}{The definition of the new attribute.}
}
\value{
The attribute was created. The response's Location header is a link to the created attribute.
}
\description{
Create a new attribute as a child of the specified attribute.
}
