% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeApi.r
\name{attribute$createConfig}
\alias{attribute$createConfig}
\title{Create or update an attribute's DataReference configuration (Create/Update PI point for PI Point DataReference).}
\arguments{
\item{webId}{The ID of the attribute.}
}
\value{
The attribute's DataReference configuration was created or updated.
}
\description{
Create or update an attribute's DataReference configuration (Create/Update PI point for PI Point DataReference).
}
