% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeApi.r
\name{attribute$get}
\alias{attribute$get}
\title{Retrieve an attribute.}
\arguments{
\item{webId}{The ID of the attribute.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested attribute.
}
\description{
Retrieve an attribute.
}
