% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeApi.r
\name{attribute$update}
\alias{attribute$update}
\title{Update an attribute by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the attribute.}

\item{attribute}{A partial attribute containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The attribute was updated.
}
\description{
Update an attribute by replacing items in its definition.
}
