% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeCategoryApi.r
\name{attributeCategory$delete}
\alias{attributeCategory$delete}
\title{Delete an attribute category.}
\arguments{
\item{webId}{The ID of the attribute category to delete.}
}
\value{
The attribute category was deleted.
}
\description{
Delete an attribute category.
}
