% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeCategoryApi.r
\name{attributeCategory$getByPath}
\alias{attributeCategory$getByPath}
\title{Retrieve an attribute category by path.}
\arguments{
\item{path}{The path to the target attribute category.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified attribute category.
}
\description{
Retrieve an attribute category by path.
}
