% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeCategoryApi.r
\name{attributeCategory$getSecurityEntries}
\alias{attributeCategory$getSecurityEntries}
\title{Retrieve the security entries associated with the attribute category based on the specified criteria. By default, all security entries for this attribute category are returned.}
\arguments{
\item{webId}{The ID of the attribute category.}

\item{nameFilter}{The name query string used for filtering security entries. The default is no filter.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of security entries matching the specified condition.
}
\description{
Retrieve the security entries associated with the attribute category based on the specified criteria. By default, all security entries for this attribute category are returned.
}
