% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeCategoryApi.r
\name{attributeCategory$getSecurityEntryByName}
\alias{attributeCategory$getSecurityEntryByName}
\title{Retrieve the security entry associated with the attribute category with the specified name.}
\arguments{
\item{name}{The name of the security entry. For every backslash character (\\) in the security entry name, replace with asterisk (*). As an example, use domain*username instead of domain\\username.}

\item{webId}{The ID of the attribute category.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The security entry matching the specified condition.
}
\description{
Retrieve the security entry associated with the attribute category with the specified name.
}
