% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTemplateApi.r
\name{attributeTemplate$createAttributeTemplate}
\alias{attributeTemplate$createAttributeTemplate}
\title{Create an attribute template as a child of another attribute template.}
\arguments{
\item{webId}{The ID of the parent attribute template on which to create the attribute template.}

\item{PIAttributeTemplate}{The attribute template definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The attribute template was created. The response's Location header is a link to the created resource.
}
\description{
Create an attribute template as a child of another attribute template.
}
