% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTemplateApi.r
\name{attributeTemplate$get}
\alias{attributeTemplate$get}
\title{Retrieve an attribute template.}
\arguments{
\item{webId}{The ID of the attribute template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified attribute template.
}
\description{
Retrieve an attribute template.
}
