% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTemplateApi.r
\name{attributeTemplate$getAttributeTemplates}
\alias{attributeTemplate$getAttributeTemplates}
\title{Retrieve an attribute template's child attribute templates.}
\arguments{
\item{webId}{The ID of the attribute template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{column}{the value of the user label column}
}
\value{
A list of child attribute templates.
}
\description{
Retrieve an attribute template's child attribute templates.
}
