% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTemplateApi.r
\name{attributeTemplate$getCategories}
\alias{attributeTemplate$getCategories}
\title{Get an attribute template's categories.}
\arguments{
\item{webId}{The ID of the attribute template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of categories.
}
\description{
Get an attribute template's categories.
}
