% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTemplateApi.r
\name{attributeTemplate$update}
\alias{attributeTemplate$update}
\title{Update an existing attribute template by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the attribute template.}

\item{PIAttributeTemplate}{A partial attribute template containing the desired changes.}
}
\value{
The attribute template was updated.
}
\description{
Update an existing attribute template by replacing items in its definition.
}
