% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTraitApi.r
\name{attributeTrait$get}
\alias{attributeTrait$get}
\title{Retrieve an attribute trait.}
\arguments{
\item{name}{The name or abbreviation of the attribute trait.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested attribute trait.
}
\description{
Retrieve an attribute trait.
}
