% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTraitApi.r
\name{attributeTrait$getByCategory}
\alias{attributeTrait$getByCategory}
\title{Retrieve all attribute traits of the specified category/categories.}
\arguments{
\item{category}{The category of the attribute traits. Multiple categories may be specified with multiple instances of the parameter. If the parameter is not specified, or if its value is "all", then all attribute traits of all categories will be returned.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of attribute traits.
}
\description{
Retrieve all attribute traits of the specified category/categories.
}
