% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationApi.r
\name{calculation$getAtTimes}
\alias{calculation$getAtTimes}
\title{Returns the result of evaluating the expression at the specified timestamps.}
\arguments{
\item{expression}{A string containing the expression to be evaluated. The syntax for the expression generally follows the Performance Equation syntax as described in the PI Server documentation, with the exception that expressions which target AF objects use attribute names in place of tag names in the equation.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{time}{A list of timestamps at which to calculate the expression.}

\item{webId}{The ID of the target object of the expression. A target object can be a Data Server, a database, an element, an event frame or an attribute. References to attributes or points are based on the target. If this parameter is not provided, the target object is set to null.}
}
\value{
The calculated values.
}
\description{
Returns the result of evaluating the expression at the specified timestamps.
}
