% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculationApi.r
\name{calculation$getSummary}
\alias{calculation$getSummary}
\title{Returns the result of evaluating the expression over the time range from the start time to the end time. The time range is first divided into a number of summary intervals. Then the calculation is performed for the specified summaries over each interval.}
\arguments{
\item{calculationBasis}{Specifies the method of evaluating the data over the time range. The default is 'TimeWeighted'.}

\item{endTime}{An optional end time. The default is '*' for element attributes and points. For event frame attributes, the default is the event frame's end time, or '*' if that is not set. Note that if endTime is earlier than startTime, the resulting values will be in time-descending order.}

\item{expression}{A string containing the expression to be evaluated. The syntax for the expression generally follows the Performance Equation syntax as described in the PI Server documentation, with the exception that expressions which target AF objects use attribute names in place of tag names in the equation.}

\item{sampleInterval}{A time span specifies how often the filter expression is evaluated when computing the summary for an interval, if the sampleType is 'Interval'.}

\item{sampleType}{A flag which specifies one or more summaries to compute for each interval over the time range. The default is 'ExpressionRecordedValues'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{startTime}{An optional start time. The default is '*-1d' for element attributes and points. For event frame attributes, the default is the event frame's start time, or '*-1d' if that is not set.}

\item{summaryDuration}{The duration of each summary interval.}

\item{summaryType}{Specifies the kinds of summaries to produce over the range. The default is 'Total'. Multiple summary types may be specified by using multiple instances of summaryType.}

\item{timeType}{Specifies how to calculate the timestamp for each interval. The default is 'Auto'.}

\item{webId}{The ID of the target object of the expression. A target object can be a Data Server, a database, an element, an event frame or an attribute. References to attributes or points are based on the target. If this parameter is not provided, the target object is set to null.}
}
\value{
The calculated values.
}
\description{
Returns the result of evaluating the expression over the time range from the start time to the end time. The time range is first divided into a number of summary intervals. Then the calculation is performed for the specified summaries over each interval.
}
