% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/channelApi.r
\name{channel$instances}
\alias{channel$instances}
\title{Retrieves a list of currently running channel instances.}
\value{
A list of currently running channel instances.
}
\description{
Retrieves a list of currently running channel instances.
}
