% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurationApi.r
\name{configuration$delete}
\alias{configuration$delete}
\title{Delete a configuration item.}
\arguments{
\item{key}{The key of the configuration item to remove.}
}
\value{
The change has been accepted, and will take effect when the service can cleanly restart.
}
\description{
Delete a configuration item.
}
