% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurationApi.r
\name{configuration$get}
\alias{configuration$get}
\title{Get the value of a configuration item.}
\arguments{
\item{key}{The key of the configuration item.}
}
\value{
The value of the configuration item.
}
\description{
Get the value of a configuration item.
}
