% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configurationApi.r
\name{configuration$list}
\alias{configuration$list}
\title{Get the current system configuration.}
\arguments{
\item{column}{the value of the user label column}
}
\value{
The current system configuration.
}
\description{
Get the current system configuration.
}
