% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r, R/dataApi.r
\name{dataApi$getMultipleRecordedValues}
\alias{dataApi$getMultipleRecordedValues}
\alias{dataApi$getMultipleRecordedValues}
\title{Returns an array of data frames with recorded values of the specified streams.}
\arguments{
\item{paths}{The paths of streams.}

\item{boundaryType}{An optional value that determines how the times and values of the returned end points are determined. The default is 'Inside'.}

\item{endTime}{An optional end time. The default is '*'. Note that if endTime is earlier than startTime, the resulting values will be in time-descending order.}

\item{filterExpression}{An optional string containing a filter expression. Expression variables are relative to the data point. Use '.' to reference the containing attribute. The default is no filtering.}

\item{includeFilteredValues}{Specify 'true' to indicate that values which fail the filter criteria are present in the returned data at the times where they occurred with a value set to a 'Filtered' enumeration value with bad status. Repeated consecutive failures are omitted.}

\item{maxCount}{The maximum number of values to be returned. The default is 1000.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{startTime}{An optional start time. The default is '*-1d'.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}

\item{paths}{The paths of streams.}

\item{boundaryType}{An optional value that determines how the times and values of the returned end points are determined. The default is 'Inside'.}

\item{endTime}{An optional end time. The default is '*'. Note that if endTime is earlier than startTime, the resulting values will be in time-descending order.}

\item{filterExpression}{An optional string containing a filter expression. Expression variables are relative to the data point. Use '.' to reference the containing attribute. The default is no filtering.}

\item{includeFilteredValues}{Specify 'true' to indicate that values which fail the filter criteria are present in the returned data at the times where they occurred with a value set to a 'Filtered' enumeration value with bad status. Repeated consecutive failures are omitted.}

\item{maxCount}{The maximum number of values to be returned. The default is 1000.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. For better performance, The default is unsorted. 'Name' is the only supported field by which to sort.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'}

\item{startTime}{An optional start time. The default is '*-1d'.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Recorded values of the streams that meet the specified conditions.

Recorded values of the streams that meet the specified conditions.
}
\description{
Returns an array of data frames with recorded values of the specified streams.

Returns an array of data frames with recorded values of the specified streams.
}
