% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServerApi.r
\name{dataServer$createEnumerationSet}
\alias{dataServer$createEnumerationSet}
\title{Create an enumeration set on the Data Server.}
\arguments{
\item{webId}{The ID of the server on which to create the enumeration set.}

\item{PIEnumerationSet}{The new enumeration set definition.}
}
\value{
The enumeration was created. The response's Location header is a link to the created enumeration set.
}
\description{
Create an enumeration set on the Data Server.
}
