% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServerApi.r
\name{dataServer$createPoint}
\alias{dataServer$createPoint}
\title{Create a point in the specified Data Server.}
\arguments{
\item{webId}{The ID of the server.}

\item{PIPoint}{The new point definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The point was created. The response's Location header is a link to the point.
}
\description{
Create a point in the specified Data Server.
}
