% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServerApi.r
\name{dataServer$getByName}
\alias{dataServer$getByName}
\title{Retrieve a Data Server by name.}
\arguments{
\item{name}{The name of the server.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested server.
}
\description{
Retrieve a Data Server by name.
}
