% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServerApi.r
\name{dataServer$getByPath}
\alias{dataServer$getByPath}
\title{Retrieve a Data Server by path.}
\arguments{
\item{path}{The path to the server. Note that the path supplied to this method must be of the form '\\\\servername'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested server.
}
\description{
Retrieve a Data Server by path.
}
