% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServerApi.r
\name{dataServer$getLicense}
\alias{dataServer$getLicense}
\title{Retrieves the specified license for the given Data Server. The fields of the response object are string representations of the numerical values reported by the Data Server, with "Infinity" representing a license field with no limit.}
\arguments{
\item{webId}{The ID of the server.}

\item{module}{The case-sensitive name of the module.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The specified license.
}
\description{
Retrieves the specified license for the given Data Server. The fields of the response object are string representations of the numerical values reported by the Data Server, with "Infinity" representing a license field with no limit.
}
