% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataServerApi.r
\name{dataServer$getPoints}
\alias{dataServer$getPoints}
\title{Retrieve a list of points on a specified Data Server.}
\arguments{
\item{webId}{The ID of the server.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{nameFilter}{A query string for filtering by point name. The default is no filter.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is '0'.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of matching points.
}
\description{
Retrieve a list of points on a specified Data Server.
}
