% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$addReferencedElement}
\alias{element$addReferencedElement}
\title{Add a reference to an existing element to the child elements collection.}
\arguments{
\item{webId}{The ID of the element which the referenced element will be added to.}

\item{referencedElementWebId}{The ID of the referenced element. Multiple referenced elements may be specified with multiple instances of the parameter.}

\item{referenceType}{The name of the reference type between the parent and the referenced element. The default is "parent-child".}
}
\value{
The referenced element was successfully added.
}
\description{
Add a reference to an existing element to the child elements collection.
}
