% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$createAnalysis}
\alias{element$createAnalysis}
\title{Create an Analysis.}
\arguments{
\item{webId}{The ID of the element on which to create the Analysis.}

\item{analysis}{The new Analysis definition.}

\item{column}{the value of the user label column}
}
\value{
The Analysis was created. The response's Location header is a link to the Analysis.
}
\description{
Create an Analysis.
}
