% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$createConfig}
\alias{element$createConfig}
\title{Executes the create configuration function of the data references found within the attributes of the element, and optionally, its children.}
\arguments{
\item{webId}{The ID of the element.}

\item{includeChildElements}{If true, includes the child elements of the specified element.}
}
\value{
Process log of operations. Operations completed with no errors.
}
\description{
Executes the create configuration function of the data references found within the attributes of the element, and optionally, its children.
}
