% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$delete}
\alias{element$delete}
\title{Delete an element.}
\arguments{
\item{webId}{The ID of the element.}

\item{column}{the value of the user label column}
}
\value{
The element was deleted.
}
\description{
Delete an element.
}
