% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$executeSearchByAttribute}
\alias{element$executeSearchByAttribute}
\title{Execute a "Search Elements By Attribute Value" operation.}
\arguments{
\item{searchId}{The encoded search Id of the "Search Elements By Attribute Value" operation.}

\item{categoryName}{Specify that the owner of the returned attributes must have this category. For Asset Servers older than 2.7, a 400 status code (Bad Request) will be returned if this parameter is specified. The default is no filter.}

\item{descriptionFilter}{The element description filter string used for finding objects. Only the first 440 characters of the description will be searched. For Asset Servers older than 2.7, a 400 status code (Bad Request) will be returned if this parameter is specified. The default is no filter.}

\item{maxCount}{The maximum number of objects to be returned. The default is 1000.}

\item{nameFilter}{The name query string used for finding objects. The default is no filter.}

\item{searchFullHierarchy}{Specifies if the search should include objects nested further than the immediate children of the searchRoot. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
a list of matching Elements.
}
\description{
Execute a "Search Elements By Attribute Value" operation.
}
