% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$get}
\alias{element$get}
\title{Retrieve an element.}
\arguments{
\item{webId}{The ID of the element.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified element.
}
\description{
Retrieve an element.
}
