% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$getAnalyses}
\alias{element$getAnalyses}
\title{Retrieve analyses based on the specified conditions.}
\arguments{
\item{webId}{The ID of the element, which is the Target of the analyses.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of analyses matching the specified conditions.
}
\description{
Retrieve analyses based on the specified conditions.
}
