% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$getCategories}
\alias{element$getCategories}
\title{Get an element's categories.}
\arguments{
\item{webId}{The ID of the element.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of categories.
}
\description{
Get an element's categories.
}
