% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$getElements}
\alias{element$getElements}
\title{Retrieve elements based on the specified conditions. By default, this method selects immediate children of the specified element.}
\arguments{
\item{webId}{The ID of the element to use as the root of the search.}

\item{categoryName}{Specify that returned elements must have this category. The default is no category filter.}

\item{descriptionFilter}{Specify that returned elements must have this description. The default is no description filter.}

\item{elementType}{Specify that returned elements must have this type. The default type is 'Any'.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{nameFilter}{The name query string used for finding objects. The default is no filter.}

\item{searchFullHierarchy}{Specifies if the search should include objects nested further than the immediate children of the searchRoot. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{templateName}{Specify that returned elements must have this template or a template derived from this template. The default is no template filter.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of elements matching the specified conditions.
}
\description{
Retrieve elements based on the specified conditions. By default, this method selects immediate children of the specified element.
}
