% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$getEventFrames}
\alias{element$getEventFrames}
\title{Retrieve event frames that reference this element based on the specified conditions. By default, returns all event frames that reference this element that have been active in the past 8 hours.}
\arguments{
\item{webId}{The ID of the element whose related event frames are sought.}

\item{canBeAcknowledged}{Specify the returned event frames' canBeAcknowledged property. The default is no canBeAcknowledged filter.}

\item{categoryName}{Specify that returned event frames must have this category. The default is no category filter.}

\item{endTime}{The ending time for the search. The endTime must be greater than or equal to the startTime. The searchMode parameter will control whether the comparison will be performed against the event frame's startTime or endTime. The default is '*' if searchMode is not one of the 'Backward*' or 'Forward*' values.}

\item{isAcknowledged}{Specify the returned event frames' isAcknowledged property. The default no isAcknowledged filter.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{nameFilter}{The name query string used for finding event frames. The default is no filter.}

\item{searchMode}{Determines how the startTime and endTime parameters are treated when searching for event frame objects to be included in the returned collection. If this parameter is one of the 'Backward*' or 'Forward*' values, none of endTime, sortField, or sortOrder may be specified. The default is 'Overlapped'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{severity}{Specify that returned event frames must have this severity. Multiple severity values may be specified with multiple instances of the parameter. The default is no severity filter.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name' if searchMode is not one of the 'Backward*' or 'Forward*' values.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending' if searchMode is not one of the 'Backward*' or 'Forward*' values.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{startTime}{The starting time for the search. startTime must be less than or equal to the endTime. The searchMode parameter will control whether the comparison will be performed against the event frame's startTime or endTime. The default is '*-8h'.}

\item{templateName}{Specify that returned event frames must have this template or a template derived from this template. The default is no template filter. Specify this parameter by name.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of event frames matching the specified conditions.
}
\description{
Retrieve event frames that reference this element based on the specified conditions. By default, returns all event frames that reference this element that have been active in the past 8 hours.
}
