% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$getPaths}
\alias{element$getPaths}
\title{Get a list of the full or relative paths to this element.}
\arguments{
\item{webId}{The ID of the element.}

\item{relativePath}{The full path in ShortName format to the parent object that the returned paths should be relative. For example, "\\\\Server1\\Database2" would return all the paths to the element relative to the database. A path of "\\\\Server1\\Database2\\RootElement" would return all paths to the element relative to "RootElement". If null, then all the full paths to the element will be returned.}
}
\value{
A list of the full or relative paths to this element.
}
\description{
Get a list of the full or relative paths to this element.
}
