% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$removeReferencedElement}
\alias{element$removeReferencedElement}
\title{Remove a reference to an existing element from the child elements collection.}
\arguments{
\item{webId}{The ID of the element which the referenced element will be removed from.}

\item{referencedElementWebId}{The ID of the referenced element. Multiple referenced elements may be specified with multiple instances of the parameter.}
}
\value{
The referenced element was successfully removed.
}
\description{
Remove a reference to an existing element from the child elements collection.
}
