% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$update}
\alias{element$update}
\title{Update an element by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the element.}

\item{element}{A partial element containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The element was updated.
}
\description{
Update an element by replacing items in its definition.
}
