% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementCategoryApi.r
\name{elementCategory$delete}
\alias{elementCategory$delete}
\title{Delete an element category.}
\arguments{
\item{webId}{The ID of the element category to delete.}
}
\value{
The element category was deleted.
}
\description{
Delete an element category.
}
