% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementCategoryApi.r
\name{elementCategory$get}
\alias{elementCategory$get}
\title{Retrieve an element category.}
\arguments{
\item{webId}{The id of the element category.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified element category.
}
\description{
Retrieve an element category.
}
