% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementCategoryApi.r
\name{elementCategory$update}
\alias{elementCategory$update}
\title{Update an element category by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the element category to update.}

\item{PIElementCategory}{A partial element category containing the desired changes.}
}
\value{
The element category was updated.
}
\description{
Update an element category by replacing items in its definition.
}
