% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementTemplateApi.r
\name{elementTemplate$delete}
\alias{elementTemplate$delete}
\title{Delete an element template.}
\arguments{
\item{webId}{The ID of the element template to update.}
}
\value{
The element template was deleted.
}
\description{
Delete an element template.
}
