% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementTemplateApi.r
\name{elementTemplate$getAttributeTemplates}
\alias{elementTemplate$getAttributeTemplates}
\title{Get child attribute templates for an element template.}
\arguments{
\item{webId}{The ID of the element template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{showInherited}{Specifies if the result should include attribute templates inherited from base element templates. The default is 'false'.}
}
\value{
A list of attribute templates for the specified element template.
}
\description{
Get child attribute templates for an element template.
}
