% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementTemplateApi.r
\name{elementTemplate$getBaseElementTemplates}
\alias{elementTemplate$getBaseElementTemplates}
\title{Get base element templates for an element template.}
\arguments{
\item{webId}{The ID of the element template.}

\item{maxCount}{The maximum number of objects to be returned. The default is 1000.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of base element templates for the specified element template.
}
\description{
Get base element templates for an element template.
}
