% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementTemplateApi.r
\name{elementTemplate$getByPath}
\alias{elementTemplate$getByPath}
\title{Retrieve an element template by path.}
\arguments{
\item{path}{The path to the element template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified element template.
}
\description{
Retrieve an element template by path.
}
