% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementTemplateApi.r
\name{elementTemplate$update}
\alias{elementTemplate$update}
\title{Update an element template by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the element template to update.}

\item{PIElementTemplate}{A partial element template containing the desired changes.}
}
\value{
The element template was updated.
}
\description{
Update an element template by replacing items in its definition.
}
