% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationSetApi.r
\name{enumerationSet$delete}
\alias{enumerationSet$delete}
\title{Delete an enumeration set.}
\arguments{
\item{webId}{The ID of the enumeration set to delete.}

\item{column}{the value of the user label column}
}
\value{
The enumeration set was deleted.
}
\description{
Delete an enumeration set.
}
