% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationSetApi.r
\name{enumerationSet$get}
\alias{enumerationSet$get}
\title{Retrieve an enumeration set.}
\arguments{
\item{webId}{The ID of the enumeration set.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified enumeration sets.
}
\description{
Retrieve an enumeration set.
}
