% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationSetApi.r
\name{enumerationSet$getByPath}
\alias{enumerationSet$getByPath}
\title{Retrieve an enumeration set by path.}
\arguments{
\item{path}{The path to the target enumeration set.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified enumeration set.
}
\description{
Retrieve an enumeration set by path.
}
