% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationSetApi.r
\name{enumerationSet$getValues}
\alias{enumerationSet$getValues}
\title{Retrieve an enumeration set's values.}
\arguments{
\item{webId}{The ID of the enumeration set.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified enumeration set's values
}
\description{
Retrieve an enumeration set's values.
}
