% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationSetApi.r
\name{enumerationSet$update}
\alias{enumerationSet$update}
\title{Update an enumeration set by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the enumeration set to update.}

\item{enumerationSet}{A partial enumeration set containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The enumeration set was updated.
}
\description{
Update an enumeration set by replacing items in its definition.
}
