% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationValueApi.r
\name{enumerationValue$deleteEnumerationValue}
\alias{enumerationValue$deleteEnumerationValue}
\title{Delete an enumeration value from an enumeration set.}
\arguments{
\item{webId}{The ID of the enumeration value.}
}
\value{
The enumeration value was deleted.
}
\description{
Delete an enumeration value from an enumeration set.
}
