% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationValueApi.r
\name{enumerationValue$get}
\alias{enumerationValue$get}
\title{Retrieve an enumeration value mapping}
\arguments{
\item{webId}{The ID of the enumeration value.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified enumeration value mapping
}
\description{
Retrieve an enumeration value mapping
}
