% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationValueApi.r
\name{enumerationValue$getByPath}
\alias{enumerationValue$getByPath}
\title{Retrieve an enumeration value by path.}
\arguments{
\item{path}{The path to the target enumeration value.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified enumeration set.
}
\description{
Retrieve an enumeration value by path.
}
