% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$acknowledge}
\alias{eventFrame$acknowledge}
\title{Calls the EventFrame's Acknowledge method.}
\arguments{
\item{webId}{The ID of the event frame.}
}
\value{
The event frame has been acknowledged.
}
\description{
Calls the EventFrame's Acknowledge method.
}
