% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$captureValues}
\alias{eventFrame$captureValues}
\title{Calls the EventFrame's CaptureValues method.}
\arguments{
\item{webId}{The ID of the event frame.}
}
\value{
The event frame attribute's values were captured.
}
\description{
Calls the EventFrame's CaptureValues method.
}
