% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$createConfig}
\alias{eventFrame$createConfig}
\title{Executes the create configuration function of the data references found within the attributes of the event frame, and optionally, its children.}
\arguments{
\item{webId}{The ID of the event frame.}

\item{includeChildElements}{If true, includes the child event frames of the specified event frame.}
}
\value{
Process log of operations. Operations completed with no errors.
}
\description{
Executes the create configuration function of the data references found within the attributes of the event frame, and optionally, its children.
}
