% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$delete}
\alias{eventFrame$delete}
\title{Delete an event frame.}
\arguments{
\item{webId}{The ID of the event frame to delete.}
}
\value{
The event frame was deleted.
}
\description{
Delete an event frame.
}
