% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$deleteAnnotation}
\alias{eventFrame$deleteAnnotation}
\title{Delete an annotation on an event frame. If the annotation has attached media, the attached media will also be deleted.}
\arguments{
\item{id}{The Annotation identifier of the annotation to be deleted.}

\item{webId}{The ID of the owner event frame of the annotation to delete.}
}
\value{
The annotation was deleted.
}
\description{
Delete an annotation on an event frame. If the annotation has attached media, the attached media will also be deleted.
}
