% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$deleteAnnotationAttachmentMediaById}
\alias{eventFrame$deleteAnnotationAttachmentMediaById}
\title{Delete attached media from an annotation on an event frame.}
\arguments{
\item{id}{The Annotation identifier of the annotation to delete the attached media of.}

\item{webId}{The ID of the owner event frame of the annotation to delete the attached media of.}
}
\value{
The attached media was deleted.
}
\description{
Delete attached media from an annotation on an event frame.
}
