% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$executeSearchByAttribute}
\alias{eventFrame$executeSearchByAttribute}
\title{Execute a "Search EventFrames By Attribute Value" operation.}
\arguments{
\item{searchId}{The encoded search Id of the "Search EventFrames By Attribute Value" operation.}

\item{canBeAcknowledged}{Specify the returned event frames' canBeAcknowledged property. The default is no canBeAcknowledged filter.}

\item{endTime}{The ending time for the search. endTime must be greater than or equal to the startTime. The searchMode parameter will control whether the comparison will be performed against the event frame's startTime or endTime. The default is '*'.}

\item{isAcknowledged}{Specify the returned event frames' isAcknowledged property. The default no isAcknowledged filter.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{nameFilter}{The name query string used for finding event frames. The default is no filter.}

\item{referencedElementNameFilter}{The name query string which must match the name of a referenced element. The default is no filter.}

\item{searchFullHierarchy}{Specifies whether the search should include objects nested further than the immediate children of the search root. The default is 'false'.}

\item{searchMode}{Determines how the startTime and endTime parameters are treated when searching for event frame objects to be included in the returned collection. The default is 'Overlapped'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{severity}{Specify that returned event frames must have this severity. Multiple severity values may be specified with multiple instances of the parameter. The default is no severity filter.}

\item{sortField}{The field or property of the object used to sort the returned collection. The default is 'Name'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{startTime}{The starting time for the search. startTime must be less than or equal to the endTime. The searchMode parameter will control whether the comparison will be performed against the event frame's startTime or endTime. The default is '*-8h'.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
a list of matching EventFrames.
}
\description{
Execute a "Search EventFrames By Attribute Value" operation.
}
