% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$get}
\alias{eventFrame$get}
\title{Retrieve an event frame.}
\arguments{
\item{webId}{The ID of the event frame.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified event frame.
}
\description{
Retrieve an event frame.
}
