% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$getAnnotationById}
\alias{eventFrame$getAnnotationById}
\title{Get a specific annotation on an event frame.}
\arguments{
\item{id}{The Annotation identifier of the specific annotation.}

\item{webId}{The ID of the owner event frame.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested annotation.
}
\description{
Get a specific annotation on an event frame.
}
