% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$getAnnotations}
\alias{eventFrame$getAnnotations}
\title{Get an event frame's annotations.}
\arguments{
\item{webId}{The ID of the owner event frame.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of annotations.
}
\description{
Get an event frame's annotations.
}
