% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$getCategories}
\alias{eventFrame$getCategories}
\title{Get an event frame's categories.}
\arguments{
\item{webId}{The ID of the event frame.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of categories.
}
\description{
Get an event frame's categories.
}
