% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$getEventFramesQuery}
\alias{eventFrame$getEventFramesQuery}
\title{Retrieve event frames based on the specified conditions. Returns event frames using the specified search query string.}
\arguments{
\item{databaseWebId}{The ID of the asset database to use as the root of the query.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{query}{The query string is a list of filters used to perform an AFSearch for the eventframes in the asset database. An example would be: "query=Name:=MyEventFrame* Category:=MyCategory Template:=EFTemplate*".}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of event frames matching the specified conditions.
}
\description{
Retrieve event frames based on the specified conditions. Returns event frames using the specified search query string.
}
