% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$updateAnnotation}
\alias{eventFrame$updateAnnotation}
\title{Update an annotation on an event frame by replacing items in its definition.}
\arguments{
\item{id}{The Annotation identifier of the annotation to be updated.}

\item{webId}{The ID of the owner event frame of the annotation to update.}

\item{PIAnnotation}{A partial annotation containing the desired changes.}
}
\value{
The annotation was updated.
}
\description{
Update an annotation on an event frame by replacing items in its definition.
}
