% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/homeApi.r
\name{home$get}
\alias{home$get}
\title{Get top level links for this PI System Web API instance.}
\value{
Links to top level collections and system information.
}
\description{
Get top level links for this PI System Web API instance.
}
