% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notificationRuleTemplateApi.r
\name{notificationRuleTemplate$get}
\alias{notificationRuleTemplate$get}
\title{Get the specified notification rule template.}
\arguments{
\item{webId}{The ID of the notification rule template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The specified notification rule template.
}
\description{
Get the specified notification rule template.
}
