% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notificationRuleTemplateApi.r
\name{notificationRuleTemplate$getNotificationRuleTemplatesQuery}
\alias{notificationRuleTemplate$getNotificationRuleTemplatesQuery}
\title{Retrieve Notification rule templates based on the specified conditions. Returns Notification rule templates using the specified search query string.}
\arguments{
\item{databaseWebId}{The ID of the asset database to use as the root of the query.}

\item{maxCount}{The maximum number of objects to be returned per call (page size). The default is 1000.}

\item{query}{The query string is a list of filters used to perform an AFSearch for the Notification rule templates in the asset database. An example would be: "query=NotificationRuleTemplate:{ Name:='MyNotificationRuleTemplate' } Type:=Int32".}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{startIndex}{The starting index (zero based) of the items to be returned. The default is 0.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of notification rule templates matching the specified conditions.
}
\description{
Retrieve Notification rule templates based on the specified conditions. Returns Notification rule templates using the specified search query string.
}
