% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/notificationRuleTemplateApi.r
\name{notificationRuleTemplate$getNotificationRuleTemplateSubscribers}
\alias{notificationRuleTemplate$getNotificationRuleTemplateSubscribers}
\title{Retrieve notification rule template subscribers.}
\arguments{
\item{webId}{The ID of the resource to use as the root of the search.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of notification rule template subscribers.
}
\description{
Retrieve notification rule template subscribers.
}
