% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAmbiguous.r
\name{PIAmbiguous}
\alias{PIAmbiguous}
\title{Generate an instance of the PIAmbiguous PI Web API class}
\usage{
PIAmbiguous(reason = NULL)
}
\arguments{
\item{reason}{(string)}
}
\value{
PIAmbiguous
}
\description{
Generate an instance of the PIAmbiguous PI Web API class
}
\examples{
ambiguous <- PIAmbiguous(reason = "The requested resource has multiple representations.")

}
