% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysis.r
\name{PIAnalysis}
\alias{PIAnalysis}
\title{Generate an instance of the PIAnalysis PI Web API class}
\usage{
PIAnalysis(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, analysisRulePlugInName = NULL, autoCreated = NULL,
  categoryNames = NULL, groupId = NULL, hasNotification = NULL,
  hasTarget = NULL, hasTemplate = NULL, isConfigured = NULL,
  isTimeRuleDefinedByTemplate = NULL, maximumQueueSize = NULL,
  outputTime = NULL, priority = NULL, publishResults = NULL,
  status = NULL, targetWebId = NULL, templateName = NULL,
  timeRulePlugInName = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{analysisRulePlugInName}{(string)}

\item{autoCreated}{(boolean)}

\item{categoryNames}{(array)}

\item{groupId}{(integer)}

\item{hasNotification}{(boolean)}

\item{hasTarget}{(boolean)}

\item{hasTemplate}{(boolean)}

\item{isConfigured}{(boolean)}

\item{isTimeRuleDefinedByTemplate}{(boolean)}

\item{maximumQueueSize}{(integer)}

\item{outputTime}{(string)}

\item{priority}{(string)}

\item{publishResults}{(boolean)}

\item{status}{(string)}

\item{targetWebId}{(string)}

\item{templateName}{(string)}

\item{timeRulePlugInName}{(string)}

\item{links}{(object)}
}
\value{
PIAnalysis
}
\description{
Generate an instance of the PIAnalysis PI Web API class
}
\examples{
analysis <- PIAnalysis(webId = "x0DqD5loBNH0erqeqJodtALAGgF3cBcv5hGA2AAVXYRDBAUkVTVFVOSVRcNkQwQjc2MDItOTYxMi00OTRBLUE4MEQtODQyRDMyOTZENTYxXDkyMjFGMDhGLUZCMTUtNDZFMy1BQzYzLURCNDkzRDhGNkJFQlxBTkFMWVNFU1swRUFFNDU3My0wMDY0LTQxMUUtODQwRC0wOEMwOTlEM0VGNTNd", id = "7077011a-2f17-11e6-80d8-00155d844304", name = "MyAnalysis", description = "", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement\\\\Analyses[MyAnalysis]", analysisRulePlugInName = "PerformanceEquation", autoCreated = FALSE, groupId = 0, hasNotification = FALSE, hasTarget = FALSE, hasTemplate = FALSE, isConfigured = FALSE, isTimeRuleDefinedByTemplate = FALSE, maximumQueueSize = 0, outputTime = "", priority = "High", publishResults = FALSE, status = "Disabled", targetWebId = "E0DqD5loBNH0erqeqJodtALAFwF3cBcv5hGA2AAVXYRDBAUkVTVFVOSVRcNkQwQjc2MDItOTYxMi00OTRBLUE4MEQtODQyRDMyOTZENTYxXDkyMjFGMDhGLUZCMTUtNDZFMy1BQzYzLURCNDkzRDhGNkJFQg", templateName = "", timeRulePlugInName = "Periodic")

}
