% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piAnalysisCategory.r
\name{piAnalysisCategory}
\alias{piAnalysisCategory}
\title{Generate an instance of the piAnalysisCategory PI Web API class}
\usage{
piAnalysisCategory(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{(object)}
}
\value{
analysisCategory
}
\description{
Generate an instance of the piAnalysisCategory PI Web API class
}
\examples{
analysisCategory(webId = "n0DqD5loBNH0erqeqJodtALAoko2-UoOVEibhTWQCk1MDwUkVTVFVOSVRcUk9OR1xDQVRFR09SSUVTQU5BTFlTSVNbQ0FURUdPUlkxXQ", id = "f9364aa2-0e4a-4854-9b85-35900a4d4c0f", name = "CategoryName", description = "Relative energy use per ton of process feed.", path = "\\\\\\\\MyServer\\\\Database\\\\CategoriesAnalysis[CategoryName]")

}
