% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisCategoryLinks.r
\name{PIAnalysisCategoryLinks}
\alias{PIAnalysisCategoryLinks}
\title{Generate an instance of the PIAnalysisCategoryLinks PI Web API class}
\usage{
PIAnalysisCategoryLinks(self = NULL, database = NULL, security = NULL,
  securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIAnalysisCategoryLinks
}
\description{
Generate an instance of the PIAnalysisCategoryLinks PI Web API class
}
\examples{
analysisCategoryLinks <- PIAnalysisCategoryLinks(self = "", database = "", security = "", securityEntries = "")

}
