% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisLinks.r
\name{PIAnalysisLinks}
\alias{PIAnalysisLinks}
\title{Generate an instance of the PIAnalysisLinks PI Web API class}
\usage{
PIAnalysisLinks(self = NULL, target = NULL, database = NULL,
  categories = NULL, template = NULL, analysisRule = NULL,
  analysisRulePlugIn = NULL, timeRule = NULL, timeRulePlugIn = NULL,
  security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{target}{(string)}

\item{database}{(string)}

\item{categories}{(string)}

\item{template}{(string)}

\item{analysisRule}{(string)}

\item{analysisRulePlugIn}{(string)}

\item{timeRule}{(string)}

\item{timeRulePlugIn}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIAnalysisLinks
}
\description{
Generate an instance of the PIAnalysisLinks PI Web API class
}
\examples{
analysisLinks <- PIAnalysisLinks(self = "", target = "", database = "", categories = "", template = "", analysisRule = "", analysisRulePlugIn = "", timeRule = "", timeRulePlugIn = "", security = "", securityEntries = "")

}
