% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisRuleLinks.r
\name{PIAnalysisRuleLinks}
\alias{PIAnalysisRuleLinks}
\title{Generate an instance of the PIAnalysisRuleLinks PI Web API class}
\usage{
PIAnalysisRuleLinks(self = NULL, analysisRules = NULL, analysis = NULL,
  analysisTemplate = NULL, parent = NULL, plugIn = NULL)
}
\arguments{
\item{self}{(string)}

\item{analysisRules}{(string)}

\item{analysis}{(string)}

\item{analysisTemplate}{(string)}

\item{parent}{(string)}

\item{plugIn}{(string)}
}
\value{
PIAnalysisRuleLinks
}
\description{
Generate an instance of the PIAnalysisRuleLinks PI Web API class
}
\examples{
analysisRuleLinks <- PIAnalysisRuleLinks(self = "", analysisRules = "", analysis = "", analysisTemplate = "", parent = "", plugIn = "")

}
