% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisRulePlugIn.r
\name{PIAnalysisRulePlugIn}
\alias{PIAnalysisRulePlugIn}
\title{Generate an instance of the PIAnalysisRulePlugIn PI Web API class}
\usage{
PIAnalysisRulePlugIn(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, assemblyFileName = NULL,
  assemblyID = NULL, assemblyLoadProperties = NULL, assemblyTime = NULL,
  compatibilityVersion = NULL, isBrowsable = NULL,
  isNonEditableConfig = NULL, loadedAssemblyTime = NULL,
  loadedVersion = NULL, version = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{assemblyFileName}{(string)}

\item{assemblyID}{(string)}

\item{assemblyLoadProperties}{(array)}

\item{assemblyTime}{(string)}

\item{compatibilityVersion}{(integer)}

\item{isBrowsable}{(boolean)}

\item{isNonEditableConfig}{(boolean)}

\item{loadedAssemblyTime}{(string)}

\item{loadedVersion}{(string)}

\item{version}{(string)}

\item{links}{(object)}
}
\value{
PIAnalysisRulePlugIn
}
\description{
Generate an instance of the PIAnalysisRulePlugIn PI Web API class
}
\examples{
analysisRulePlugIn <- PIAnalysisRulePlugIn(webId = "L0DqD5loBNH0erqeqJodtALAAc7dxPJowEixZDO6X94w4AUkVTVFVOSVRcUExVR0lOU0FOQUxZU0lTUlVMRVtJTUJBTEFOQ0Vd", id = "c4ddce01-68f2-48c0-b164-33ba5fde30e0", name = "Imbalance", description = "Calculates simple material imbalances on nodes of a model.", path = "\\\\\\\\MYSERVER\\\\PlugInsAnalysisRule[Imbalance]", assemblyFileName = "OSIsoft.AF.Analysis.AnalysisRule.Imbalance.dll", assemblyID = "1acd00c2-79d0-4818-a03c-0616c4513750", assemblyTime = "5/25/2016 2:01:07 PM", compatibilityVersion = 2, isBrowsable = FALSE, isNonEditableConfig = FALSE, loadedAssemblyTime = "1/1/1970 12:00:00 AM", loadedVersion = "", version = "2.8.5.7552")

}
