% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisRulePlugInLinks.r
\name{PIAnalysisRulePlugInLinks}
\alias{PIAnalysisRulePlugInLinks}
\title{Generate an instance of the PIAnalysisRulePlugInLinks PI Web API class}
\usage{
PIAnalysisRulePlugInLinks(self = NULL, assetServer = NULL)
}
\arguments{
\item{self}{(string)}

\item{assetServer}{(string)}
}
\value{
PIAnalysisRulePlugInLinks
}
\description{
Generate an instance of the PIAnalysisRulePlugInLinks PI Web API class
}
\examples{
analysisRulePlugInLinks <- PIAnalysisRulePlugInLinks(self = "", assetServer = "")

}
