% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisTemplate.r
\name{PIAnalysisTemplate}
\alias{PIAnalysisTemplate}
\title{Generate an instance of the PIAnalysisTemplate PI Web API class}
\usage{
PIAnalysisTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, analysisRulePlugInName = NULL,
  categoryNames = NULL, createEnabled = NULL, groupId = NULL,
  hasNotificationTemplate = NULL, hasTarget = NULL, outputTime = NULL,
  targetName = NULL, timeRulePlugInName = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{analysisRulePlugInName}{(string)}

\item{categoryNames}{(array)}

\item{createEnabled}{(boolean)}

\item{groupId}{(integer)}

\item{hasNotificationTemplate}{(boolean)}

\item{hasTarget}{(boolean)}

\item{outputTime}{(string)}

\item{targetName}{(string)}

\item{timeRulePlugInName}{(string)}

\item{links}{(object)}
}
\value{
PIAnalysisTemplate
}
\description{
Generate an instance of the PIAnalysisTemplate PI Web API class
}
\examples{
analysisTemplate <- PIAnalysisTemplate(webId = "Y0DqD5loBNH0erqeqJodtALA0IiywCJ7mkyCQYNQSlvYWwUkVTVFVOSVRcOTdCRjMyMzMtNDgyMi00OTYxLTg4M0YtQzU2REFERERDODUzXEFOQUxZU0lTVEVNUExBVEVTW0EzMjcwNzI2LTkzMDEtNDY5NS1BQTE1LTBCOEM0NUNCNDJDQ10", id = "c0b288d0-7b22-4c9a-8241-83504a5bd85b", name = "MyAnalysisTemplate", description = "", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\AnalysisTemplates[MyAnalysisTemplate]", analysisRulePlugInName = "PerformanceEquation", createEnabled = FALSE, groupId = 0, hasNotificationTemplate = FALSE, hasTarget = FALSE, outputTime = "", targetName = "MyElementTemplate", timeRulePlugInName = "Periodic")

}
