% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisTemplateLinks.r
\name{PIAnalysisTemplateLinks}
\alias{PIAnalysisTemplateLinks}
\title{Generate an instance of the PIAnalysisTemplateLinks PI Web API class}
\usage{
PIAnalysisTemplateLinks(self = NULL, database = NULL, categories = NULL,
  analysisRule = NULL, analysisRulePlugIn = NULL, timeRule = NULL,
  timeRulePlugIn = NULL, target = NULL, security = NULL,
  securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{categories}{(string)}

\item{analysisRule}{(string)}

\item{analysisRulePlugIn}{(string)}

\item{timeRule}{(string)}

\item{timeRulePlugIn}{(string)}

\item{target}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIAnalysisTemplateLinks
}
\description{
Generate an instance of the PIAnalysisTemplateLinks PI Web API class
}
\examples{
analysisTemplateLinks <- PIAnalysisTemplateLinks(self = "", database = "", categories = "", analysisRule = "", analysisRulePlugIn = "", timeRule = "", timeRulePlugIn = "", target = "", security = "", securityEntries = "")

}
