% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piAnnotation.r
\name{piAnnotation}
\alias{piAnnotation}
\title{Generate an instance of the piAnnotation PI Web API class}
\usage{
piAnnotation(id = NULL, name = NULL, description = NULL, value = NULL,
  creator = NULL, creationDate = NULL, modifier = NULL,
  modifyDate = NULL, links = NULL)
}
\arguments{
\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{value}{(object)}

\item{creator}{(string)}

\item{creationDate}{(string)}

\item{modifier}{(string)}

\item{modifyDate}{(string)}

\item{links}{(object)}
}
\value{
annotation
}
\description{
Generate an instance of the piAnnotation PI Web API class
}
\examples{
annotation(id = "512B6616-CE39-4F70-9048-8C6A025FB592", name = "Temperature Annotation", description = "Signifies a spike in temperature.", creator = "MyDomain\\\\UserA", creationDate = "21/06/2016 14:45:50", modifier = "MyDomain\\\\UserA", modifyDate = "21/06/2016 14:45:50")

}
