% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnnotationLinks.r
\name{PIAnnotationLinks}
\alias{PIAnnotationLinks}
\title{Generate an instance of the PIAnnotationLinks PI Web API class}
\usage{
PIAnnotationLinks(self = NULL, owner = NULL)
}
\arguments{
\item{self}{(string)}

\item{owner}{(string)}
}
\value{
PIAnnotationLinks
}
\description{
Generate an instance of the PIAnnotationLinks PI Web API class
}
\examples{
annotationLinks <- PIAnnotationLinks(self = "", owner = "")

}
