% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAssetDatabase.r
\name{PIAssetDatabase}
\alias{PIAssetDatabase}
\title{Generate an instance of the PIAssetDatabase PI Web API class}
\usage{
PIAssetDatabase(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, extendedProperties = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{extendedProperties}{(object)}

\item{links}{(object)}
}
\value{
PIAssetDatabase
}
\description{
Generate an instance of the PIAssetDatabase PI Web API class
}
\examples{
assetDatabase <- PIAssetDatabase(webId = "D0NxzXSxtlKkGzAaZhKOB-KABJ2buwfWrkye3YhdL2FOUAUEhMQUZTMDRcQgYUUEVSRk9STUFOQ0UgVEVTVElORw", id = "23913034-072C-41AC-BBBD-95FB6C4122D5", name = "NuGreen", description = "PI BI Project Asset Model", path = "\\\\\\\\MyServer\\\\NuGreen")

}
