% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAssetDatabaseLinks.r
\name{PIAssetDatabaseLinks}
\alias{PIAssetDatabaseLinks}
\title{Generate an instance of the PIAssetDatabaseLinks PI Web API class}
\usage{
PIAssetDatabaseLinks(self = NULL, elements = NULL,
  elementTemplates = NULL, eventFrames = NULL, assetServer = NULL,
  elementCategories = NULL, attributeCategories = NULL,
  tableCategories = NULL, analysisCategories = NULL,
  analysisTemplates = NULL, enumerationSets = NULL, tables = NULL,
  security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{elements}{(string)}

\item{elementTemplates}{(string)}

\item{eventFrames}{(string)}

\item{assetServer}{(string)}

\item{elementCategories}{(string)}

\item{attributeCategories}{(string)}

\item{tableCategories}{(string)}

\item{analysisCategories}{(string)}

\item{analysisTemplates}{(string)}

\item{enumerationSets}{(string)}

\item{tables}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIAssetDatabaseLinks
}
\description{
Generate an instance of the PIAssetDatabaseLinks PI Web API class
}
\examples{
assetDatabaseLinks <- PIAssetDatabaseLinks(self = "", elements = "", elementTemplates = "", eventFrames = "", assetServer = "", elementCategories = "", attributeCategories = "", tableCategories = "", analysisCategories = "", analysisTemplates = "", enumerationSets = "", tables = "", security = "", securityEntries = "")

}
