% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAssetServerLinks.r
\name{PIAssetServerLinks}
\alias{PIAssetServerLinks}
\title{Generate an instance of the PIAssetServerLinks PI Web API class}
\usage{
PIAssetServerLinks(self = NULL, databases = NULL,
  notificationContactTemplates = NULL, securityIdentities = NULL,
  securityMappings = NULL, unitClasses = NULL, analysisRulePlugIns = NULL,
  timeRulePlugIns = NULL, security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{databases}{(string)}

\item{notificationContactTemplates}{(string)}

\item{securityIdentities}{(string)}

\item{securityMappings}{(string)}

\item{unitClasses}{(string)}

\item{analysisRulePlugIns}{(string)}

\item{timeRulePlugIns}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIAssetServerLinks
}
\description{
Generate an instance of the PIAssetServerLinks PI Web API class
}
\examples{
assetServerLinks <- PIAssetServerLinks(self = "", databases = "", notificationContactTemplates = "", securityIdentities = "", securityMappings = "", unitClasses = "", analysisRulePlugIns = "", timeRulePlugIns = "", security = "", securityEntries = "")

}
