% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeCategoryLinks.r
\name{PIAttributeCategoryLinks}
\alias{PIAttributeCategoryLinks}
\title{Generate an instance of the PIAttributeCategoryLinks PI Web API class}
\usage{
PIAttributeCategoryLinks(self = NULL, database = NULL, security = NULL,
  securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIAttributeCategoryLinks
}
\description{
Generate an instance of the PIAttributeCategoryLinks PI Web API class
}
\examples{
attributeCategoryLinks <- PIAttributeCategoryLinks(self = "", database = "", security = "", securityEntries = "")

}
