% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeLinks.r
\name{PIAttributeLinks}
\alias{PIAttributeLinks}
\title{Generate an instance of the PIAttributeLinks PI Web API class}
\usage{
PIAttributeLinks(self = NULL, attributes = NULL, element = NULL,
  eventFrame = NULL, parent = NULL, template = NULL,
  interpolatedData = NULL, recordedData = NULL, plotData = NULL,
  summaryData = NULL, value = NULL, endValue = NULL, point = NULL,
  categories = NULL, enumerationSet = NULL, enumerationValues = NULL,
  trait = NULL)
}
\arguments{
\item{self}{(string)}

\item{attributes}{(string)}

\item{element}{(string)}

\item{eventFrame}{(string)}

\item{parent}{(string)}

\item{template}{(string)}

\item{interpolatedData}{(string)}

\item{recordedData}{(string)}

\item{plotData}{(string)}

\item{summaryData}{(string)}

\item{value}{(string)}

\item{endValue}{(string)}

\item{point}{(string)}

\item{categories}{(string)}

\item{enumerationSet}{(string)}

\item{enumerationValues}{(string)}

\item{trait}{(string)}
}
\value{
PIAttributeLinks
}
\description{
Generate an instance of the PIAttributeLinks PI Web API class
}
\examples{
attributeLinks <- PIAttributeLinks(self = "", attributes = "", element = "", eventFrame = "", parent = "", template = "", interpolatedData = "", recordedData = "", plotData = "", summaryData = "", value = "", endValue = "", point = "", categories = "", enumerationSet = "", enumerationValues = "", trait = "")

}
