% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeTemplate.r
\name{PIAttributeTemplate}
\alias{PIAttributeTemplate}
\title{Generate an instance of the PIAttributeTemplate PI Web API class}
\usage{
PIAttributeTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, type = NULL, typeQualifier = NULL,
  defaultUnitsName = NULL, defaultValue = NULL,
  dataReferencePlugIn = NULL, configString = NULL,
  isConfigurationItem = NULL, isExcluded = NULL, isHidden = NULL,
  isManualDataEntry = NULL, hasChildren = NULL, categoryNames = NULL,
  traitName = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{type}{(string)}

\item{typeQualifier}{(string)}

\item{defaultUnitsName}{(string)}

\item{defaultValue}{(object)}

\item{dataReferencePlugIn}{(string)}

\item{configString}{(string)}

\item{isConfigurationItem}{(boolean)}

\item{isExcluded}{(boolean)}

\item{isHidden}{(boolean)}

\item{isManualDataEntry}{(boolean)}

\item{hasChildren}{(boolean)}

\item{categoryNames}{(array)}

\item{traitName}{(string)}

\item{links}{(object)}
}
\value{
PIAttributeTemplate
}
\description{
Generate an instance of the PIAttributeTemplate PI Web API class
}
\examples{
attributeTemplate <- PIAttributeTemplate(webId = "t0EDqD5loBNH0erqeqJodtALACd8tc2JnkUWeQ2KsKq30iA1W9lw3cYRkCQ5TRZsrkLQAUkVTVFVORVScU1VHQVJNQU5cRUxFTUVOVFRFTVBMQVRFUTQt1EFOVF18UExBTlRUUlVOQ0FURUROQU1F", id = "F82605BC-3A9C-4091-A57E-C0D9341BAD1A", name = "Water(2008)", description = "2008 Water Use", path = "\\\\\\\\MyServer\\\\NuGreen\\\\ElementTemplates[Boiler]|Water(2008)", type = "Int32", typeQualifier = "", defaultUnitsName = "liters", dataReferencePlugIn = "Table Lookup", configString = "SELECT [Water Use] FROM [Energy Use 2008] WHERE [Asset ID] = '\%Element\%'", isConfigurationItem = FALSE, isExcluded = FALSE, isHidden = FALSE, isManualDataEntry = FALSE, hasChildren = FALSE, traitName = "LimitLoLo")

}
