% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeTemplateLinks.r
\name{PIAttributeTemplateLinks}
\alias{PIAttributeTemplateLinks}
\title{Generate an instance of the PIAttributeTemplateLinks PI Web API class}
\usage{
PIAttributeTemplateLinks(self = NULL, attributeTemplates = NULL,
  elementTemplate = NULL, parent = NULL, categories = NULL,
  trait = NULL)
}
\arguments{
\item{self}{(string)}

\item{attributeTemplates}{(string)}

\item{elementTemplate}{(string)}

\item{parent}{(string)}

\item{categories}{(string)}

\item{trait}{(string)}
}
\value{
PIAttributeTemplateLinks
}
\description{
Generate an instance of the PIAttributeTemplateLinks PI Web API class
}
\examples{
attributeTemplateLinks <- PIAttributeTemplateLinks(self = "", attributeTemplates = "", elementTemplate = "", parent = "", categories = "", trait = "")

}
