% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piAttributeTrait.r
\name{piAttributeTrait}
\alias{piAttributeTrait}
\title{Generate an instance of the piAttributeTrait PI Web API class}
\usage{
piAttributeTrait(name = NULL, abbreviation = NULL,
  allowChildAttributes = NULL, allowDuplicates = NULL,
  isAllowedOnRootAttribute = NULL, isTypeInherited = NULL,
  isUOMInherited = NULL, requireNumeric = NULL, requireString = NULL,
  links = NULL)
}
\arguments{
\item{name}{(string)}

\item{abbreviation}{(string)}

\item{allowChildAttributes}{(boolean)}

\item{allowDuplicates}{(boolean)}

\item{isAllowedOnRootAttribute}{(boolean)}

\item{isTypeInherited}{(boolean)}

\item{isUOMInherited}{(boolean)}

\item{requireNumeric}{(boolean)}

\item{requireString}{(boolean)}

\item{links}{(object)}
}
\value{
attributeTrait
}
\description{
Generate an instance of the piAttributeTrait PI Web API class
}
\examples{
attributeTrait(name = "LimitLoLo", abbreviation = "LoLo", allowChildAttributes = FALSE, allowDuplicates = FALSE, isAllowedOnRootAttribute = FALSE, isTypeInherited = FALSE, isUOMInherited = FALSE, requireNumeric = FALSE, requireString = FALSE)

}
