% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeTraitLinks.r
\name{PIAttributeTraitLinks}
\alias{PIAttributeTraitLinks}
\title{Generate an instance of the PIAttributeTraitLinks PI Web API class}
\usage{
PIAttributeTraitLinks(self = NULL)
}
\arguments{
\item{self}{(string)}
}
\value{
PIAttributeTraitLinks
}
\description{
Generate an instance of the PIAttributeTraitLinks PI Web API class
}
\examples{
attributeTraitLinks <- PIAttributeTraitLinks(self = "")

}
