% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piCacheInstance.r
\name{piCacheInstance}
\alias{piCacheInstance}
\title{Generate an instance of the piCacheInstance PI Web API class}
\usage{
piCacheInstance(id = NULL, lastRefreshTime = NULL,
  willRefreshAfter = NULL, scheduledExpirationTime = NULL, user = NULL)
}
\arguments{
\item{id}{(string)}

\item{lastRefreshTime}{(string)}

\item{willRefreshAfter}{(string)}

\item{scheduledExpirationTime}{(string)}

\item{user}{(string)}
}
\value{
piCacheInstance
}
\description{
Generate an instance of the piCacheInstance PI Web API class
}
\examples{
cacheInstance <- piCacheInstance(id = "8a5033c9-b77f-431c-9820-c9f538d80dc6", lastRefreshTime = "20/07/2014 14:08:00", willRefreshAfter = "20/07/2014 18:08:00", scheduledExpirationTime = "20/07/2014 23:08:00", user = "MyDomain\\\\Username")

}
