% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIChannelInstance.r
\name{PIChannelInstance}
\alias{PIChannelInstance}
\title{Generate an instance of the PIChannelInstance PI Web API class}
\usage{
PIChannelInstance(id = NULL, startTime = NULL, lastMessageSentTime = NULL,
  sentMessageCount = NULL, webException = NULL)
}
\arguments{
\item{id}{(string)}

\item{startTime}{(string)}

\item{lastMessageSentTime}{(string)}

\item{sentMessageCount}{(integer)}

\item{webException}{()}
}
\value{
PIChannelInstance
}
\description{
Generate an instance of the PIChannelInstance PI Web API class
}
\examples{
channelInstance <- PIChannelInstance(id = "", startTime = "", lastMessageSentTime = "", sentMessageCount = )

}
