% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDataServerLicense.r
\name{PIDataServerLicense}
\alias{PIDataServerLicense}
\title{Generate an instance of the PIDataServerLicense PI Web API class}
\usage{
PIDataServerLicense(amountLeft = NULL, amountUsed = NULL, name = NULL,
  totalAmount = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{amountLeft}{(string)}

\item{amountUsed}{(string)}

\item{name}{(string)}

\item{totalAmount}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIDataServerLicense
}
\description{
Generate an instance of the PIDataServerLicense PI Web API class
}
\examples{
dataServerLicense <- PIDataServerLicense(amountLeft = "85675", amountUsed = "1325", name = "Component.Module", totalAmount = "10000")

}
