% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDataServerLicenseLinks.r
\name{PIDataServerLicenseLinks}
\alias{PIDataServerLicenseLinks}
\title{Generate an instance of the PIDataServerLicenseLinks PI Web API class}
\usage{
PIDataServerLicenseLinks(self = NULL, parent = NULL)
}
\arguments{
\item{self}{(string)}

\item{parent}{(string)}
}
\value{
PIDataServerLicenseLinks
}
\description{
Generate an instance of the PIDataServerLicenseLinks PI Web API class
}
\examples{
dataServerLicenseLinks <- PIDataServerLicenseLinks(self = "", parent = "")

}
