% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIDataServerLinks.r
\name{PIDataServerLinks}
\alias{PIDataServerLinks}
\title{Generate an instance of the PIDataServerLinks PI Web API class}
\usage{
PIDataServerLinks(self = NULL, points = NULL, enumerationSets = NULL)
}
\arguments{
\item{self}{(string)}

\item{points}{(string)}

\item{enumerationSets}{(string)}
}
\value{
PIDataServerLinks
}
\description{
Generate an instance of the PIDataServerLinks PI Web API class
}
\examples{
dataServerLinks <- PIDataServerLinks(self = "", points = "", enumerationSets = "")

}
