% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElement.r
\name{PIElement}
\alias{PIElement}
\title{Generate an instance of the PIElement PI Web API class}
\usage{
PIElement(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, templateName = NULL, hasChildren = NULL,
  categoryNames = NULL, extendedProperties = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{templateName}{(string)}

\item{hasChildren}{(boolean)}

\item{categoryNames}{(array)}

\item{extendedProperties}{(object)}

\item{links}{(object)}
}
\value{
PIElement
}
\description{
Generate an instance of the PIElement PI Web API class
}
\examples{
element <- PIElement(webId = "E0DqD5loBNH0erqeqJodtALAvGEqktDl4xG-twAFmjx6UAUkVTVFVOSVRUc1VHQVJNQU5V1TcHUkVFTg", id = "FCBC2D39-4085-491B-A0EE-E25767C85FA5", name = "B-210", description = "Boiler B-210", path = "\\\\\\\\MyServer\\\\NuGreen\\\\NuGreen\\\\Houston\\\\Cracking Process\\\\Equipment\\\\B-210", templateName = "Boiler", hasChildren = FALSE)

}
