% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElementCategory.r
\name{PIElementCategory}
\alias{PIElementCategory}
\title{Generate an instance of the PIElementCategory PI Web API class}
\usage{
PIElementCategory(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{(object)}
}
\value{
PIElementCategory
}
\description{
Generate an instance of the PIElementCategory PI Web API class
}
\examples{
elementCategory <- PIElementCategory(webId = "e0DqD5loBNH0erqeqJodtALAQ_lRME1-QUKrnEUKhMgEUAUkVTVFVOSVRcUk9OR1xDQVRFR09SSUVTRUxFTUVOVFtDQVRFR09SWTFd", id = "3051f943-7e4d-4241-ab9c-450a84c80450", name = "CategoryName", description = "Relative energy use per ton of process feed.", path = "\\\\\\\\MyServer\\\\Database\\\\CategoriesElement[CategoryName]")

}
