% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElementCategoryLinks.r
\name{PIElementCategoryLinks}
\alias{PIElementCategoryLinks}
\title{Generate an instance of the PIElementCategoryLinks PI Web API class}
\usage{
PIElementCategoryLinks(self = NULL, database = NULL, security = NULL,
  securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIElementCategoryLinks
}
\description{
Generate an instance of the PIElementCategoryLinks PI Web API class
}
\examples{
elementCategoryLinks <- PIElementCategoryLinks(self = "", database = "", security = "", securityEntries = "")

}
