% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElementLinks.r
\name{PIElementLinks}
\alias{PIElementLinks}
\title{Generate an instance of the PIElementLinks PI Web API class}
\usage{
PIElementLinks(self = NULL, analyses = NULL, attributes = NULL,
  elements = NULL, database = NULL, parent = NULL, template = NULL,
  categories = NULL, defaultAttribute = NULL, eventFrames = NULL,
  interpolatedData = NULL, recordedData = NULL, plotData = NULL,
  summaryData = NULL, value = NULL, endValue = NULL, security = NULL,
  securityEntries = NULL, notificationRules = NULL)
}
\arguments{
\item{self}{(string)}

\item{analyses}{(string)}

\item{attributes}{(string)}

\item{elements}{(string)}

\item{database}{(string)}

\item{parent}{(string)}

\item{template}{(string)}

\item{categories}{(string)}

\item{defaultAttribute}{(string)}

\item{eventFrames}{(string)}

\item{interpolatedData}{(string)}

\item{recordedData}{(string)}

\item{plotData}{(string)}

\item{summaryData}{(string)}

\item{value}{(string)}

\item{endValue}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}

\item{notificationRules}{(string)}
}
\value{
PIElementLinks
}
\description{
Generate an instance of the PIElementLinks PI Web API class
}
\examples{
elementLinks <- PIElementLinks(self = "", analyses = "", attributes = "", elements = "", database = "", parent = "", template = "", categories = "", defaultAttribute = "", eventFrames = "", interpolatedData = "", recordedData = "", plotData = "", summaryData = "", value = "", endValue = "", security = "", securityEntries = "", notificationRules = "")

}
