% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElementTemplateLinks.r
\name{PIElementTemplateLinks}
\alias{PIElementTemplateLinks}
\title{Generate an instance of the PIElementTemplateLinks PI Web API class}
\usage{
PIElementTemplateLinks(self = NULL, analysisTemplates = NULL,
  attributeTemplates = NULL, database = NULL, categories = NULL,
  baseTemplate = NULL, baseTemplates = NULL, derivedTemplates = NULL,
  defaultAttribute = NULL, notificationRuleTemplates = NULL,
  security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{analysisTemplates}{(string)}

\item{attributeTemplates}{(string)}

\item{database}{(string)}

\item{categories}{(string)}

\item{baseTemplate}{(string)}

\item{baseTemplates}{(string)}

\item{derivedTemplates}{(string)}

\item{defaultAttribute}{(string)}

\item{notificationRuleTemplates}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIElementTemplateLinks
}
\description{
Generate an instance of the PIElementTemplateLinks PI Web API class
}
\examples{
elementTemplateLinks <- PIElementTemplateLinks(self = "", analysisTemplates = "", attributeTemplates = "", database = "", categories = "", baseTemplate = "", baseTemplates = "", derivedTemplates = "", defaultAttribute = "", notificationRuleTemplates = "", security = "", securityEntries = "")

}
