% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEnumerationSet.r
\name{PIEnumerationSet}
\alias{PIEnumerationSet}
\title{Generate an instance of the PIEnumerationSet PI Web API class}
\usage{
PIEnumerationSet(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL,
  serializeDescription = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{(object)}

\item{serializeDescription}{(boolean)}
}
\value{
PIEnumerationSet
}
\description{
Generate an instance of the PIEnumerationSet PI Web API class
}
\examples{
enumerationSet <- PIEnumerationSet(webId = "U0qDSD5loBNH0erqeqJodtALAChs3W0xcXUG6x0ibnAnQdgUkVTVFVOScVRU1VHQVJNQU5cRX5VTUVSQVRJT05TRVRTW0VOVU1FUkFUSU9OIFNFVCAzXQ", id = "C186EF94-2028-4954-A136-FAEA79CEE9AA", name = "Model Number", description = "Model numbers by brand of vehicle", path = "\\\\\\\\MyServer\\\\NuGreen\\\\EnumerationSets[Model Number]", serializeDescription = FALSE)

}
