% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEnumerationSetLinks.r
\name{PIEnumerationSetLinks}
\alias{PIEnumerationSetLinks}
\title{Generate an instance of the PIEnumerationSetLinks PI Web API class}
\usage{
PIEnumerationSetLinks(self = NULL, database = NULL, dataServer = NULL,
  values = NULL, security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{dataServer}{(string)}

\item{values}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIEnumerationSetLinks
}
\description{
Generate an instance of the PIEnumerationSetLinks PI Web API class
}
\examples{
enumerationSetLinks <- PIEnumerationSetLinks(self = "", database = "", dataServer = "", values = "", security = "", securityEntries = "")

}
