% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEnumerationValue.r
\name{PIEnumerationValue}
\alias{PIEnumerationValue}
\title{Generate an instance of the PIEnumerationValue PI Web API class}
\usage{
PIEnumerationValue(webId = NULL, id = NULL, name = NULL,
  description = NULL, value = NULL, path = NULL, links = NULL,
  serializeWebId = NULL, serializeId = NULL, serializeDescription = NULL,
  serializePath = NULL, serializeLinks = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{value}{(integer)}

\item{path}{(string)}

\item{links}{(object)}

\item{serializeWebId}{(boolean)}

\item{serializeId}{(boolean)}

\item{serializeDescription}{(boolean)}

\item{serializePath}{(boolean)}

\item{serializeLinks}{(boolean)}
}
\value{
PIEnumerationValue
}
\description{
Generate an instance of the PIEnumerationValue PI Web API class
}
\examples{
enumerationValue <- PIEnumerationValue(webId = "V0SDqD5loBNH0erqeqJodtALAChs3W0xcXUG6x0ibnAnQdgUN6jDehdLkWYmU1CkJMzgNUkTVVFVOSVR1CuVHQVJNQU5cRU5VTUVSQVRJT05TRVRTW0VOVU1FUkFUSU9OIFNFVCAzXVxB", id = "7DC2449E-B31D-41AF-AA08-75A17AA32D6A", name = "BMW", description = "Model Number of BMW", value = 2005, path = "\\\\\\\\MyServer\\\\NuGreen\\\\EnumerationSets[Model Number]\\\\BMW", serializeWebId = FALSE, serializeId = FALSE, serializeDescription = FALSE, serializePath = FALSE, serializeLinks = FALSE)

}
