% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEnumerationValueLinks.r
\name{PIEnumerationValueLinks}
\alias{PIEnumerationValueLinks}
\title{Generate an instance of the PIEnumerationValueLinks PI Web API class}
\usage{
PIEnumerationValueLinks(self = NULL, enumerationSet = NULL, parent = NULL)
}
\arguments{
\item{self}{(string)}

\item{enumerationSet}{(string)}

\item{parent}{(string)}
}
\value{
PIEnumerationValueLinks
}
\description{
Generate an instance of the PIEnumerationValueLinks PI Web API class
}
\examples{
enumerationValueLinks <- PIEnumerationValueLinks(self = "", enumerationSet = "", parent = "")

}
