% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piErrors.r
\name{piErrors}
\alias{piErrors}
\title{Generate an instance of the piErrors PI Web API class}
\usage{
piErrors(errors = NULL)
}
\arguments{
\item{errors}{(array)}
}
\value{
piErrors
}
\description{
Generate an instance of the piErrors PI Web API class
}
