% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEventFrame.r
\name{PIEventFrame}
\alias{PIEventFrame}
\title{Generate an instance of the PIEventFrame PI Web API class}
\usage{
PIEventFrame(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, templateName = NULL, hasChildren = NULL,
  categoryNames = NULL, extendedProperties = NULL, startTime = NULL,
  endTime = NULL, severity = NULL, acknowledgedBy = NULL,
  acknowledgedDate = NULL, canBeAcknowledged = NULL,
  isAcknowledged = NULL, isAnnotated = NULL, isLocked = NULL,
  areValuesCaptured = NULL, refElementWebIds = NULL, security = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{templateName}{(string)}

\item{hasChildren}{(boolean)}

\item{categoryNames}{(array)}

\item{extendedProperties}{(object)}

\item{startTime}{(string)}

\item{endTime}{(string)}

\item{severity}{(string)}

\item{acknowledgedBy}{(string)}

\item{acknowledgedDate}{(string)}

\item{canBeAcknowledged}{(boolean)}

\item{isAcknowledged}{(boolean)}

\item{isAnnotated}{(boolean)}

\item{isLocked}{(boolean)}

\item{areValuesCaptured}{(boolean)}

\item{refElementWebIds}{(array)}

\item{security}{()}

\item{links}{(object)}
}
\value{
PIEventFrame
}
\description{
Generate an instance of the PIEventFrame PI Web API class
}
\examples{
eventFrame <- PIEventFrame(webId = "F0DqD5loBNH0ereJqqodtALADqD5loBNH0cAAAAAAAOJeQVKUTVFVOSVRcU1VHQVJNQU5ZVRcFTlRGUkFNRVNbRUYyMDE0MDcyMS0wMDFd", id = "6131900F-2FDD-4DD7-B082-8BB326E134F8", name = "EF20140725-001", description = "Event Frame of Past Week", path = "\\\\\\\\MyServer\\\\NuGreen\\\\EventFrames[EF20140725-001]", templateName = "Template", hasChildren = FALSE, startTime = "7/18/2014 2:45:29 PM", endTime = "7/25/2014 2:45:29 PM", severity = "None", acknowledgedBy = "MyDomain\\\\UserA", acknowledgedDate = "7/30/2014 11:04:23 AM", canBeAcknowledged = FALSE, isAcknowledged = FALSE, isAnnotated = FALSE, isLocked = FALSE, areValuesCaptured = FALSE)

}
