% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEventFrameLinks.r
\name{PIEventFrameLinks}
\alias{PIEventFrameLinks}
\title{Generate an instance of the PIEventFrameLinks PI Web API class}
\usage{
PIEventFrameLinks(self = NULL, attributes = NULL, eventFrames = NULL,
  database = NULL, referencedElements = NULL,
  primaryReferencedElement = NULL, parent = NULL, template = NULL,
  defaultAttribute = NULL, categories = NULL, annotations = NULL,
  interpolatedData = NULL, recordedData = NULL, plotData = NULL,
  summaryData = NULL, value = NULL, endValue = NULL, security = NULL,
  securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{attributes}{(string)}

\item{eventFrames}{(string)}

\item{database}{(string)}

\item{referencedElements}{(string)}

\item{primaryReferencedElement}{(string)}

\item{parent}{(string)}

\item{template}{(string)}

\item{defaultAttribute}{(string)}

\item{categories}{(string)}

\item{annotations}{(string)}

\item{interpolatedData}{(string)}

\item{recordedData}{(string)}

\item{plotData}{(string)}

\item{summaryData}{(string)}

\item{value}{(string)}

\item{endValue}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PIEventFrameLinks
}
\description{
Generate an instance of the PIEventFrameLinks PI Web API class
}
\examples{
eventFrameLinks <- PIEventFrameLinks(self = "", attributes = "", eventFrames = "", database = "", referencedElements = "", primaryReferencedElement = "", parent = "", template = "", defaultAttribute = "", categories = "", annotations = "", interpolatedData = "", recordedData = "", plotData = "", summaryData = "", value = "", endValue = "", security = "", securityEntries = "")

}
