% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIExtendedTimedValue.r
\name{PIExtendedTimedValue}
\alias{PIExtendedTimedValue}
\title{Generate an instance of the PIExtendedTimedValue PI Web API class}
\usage{
PIExtendedTimedValue(annotations = NULL, timestamp = NULL,
  unitsAbbreviation = NULL, good = NULL, questionable = NULL,
  substituted = NULL, annotated = NULL, value = NULL, errors = NULL,
  webException = NULL)
}
\arguments{
\item{annotations}{(array)}

\item{timestamp}{(string)}

\item{unitsAbbreviation}{(string)}

\item{good}{(boolean)}

\item{questionable}{(boolean)}

\item{substituted}{(boolean)}

\item{annotated}{(boolean)}

\item{value}{(object)}

\item{errors}{(array)}

\item{webException}{()}
}
\value{
PIExtendedTimedValue
}
\description{
Generate an instance of the PIExtendedTimedValue PI Web API class
}
\examples{
extendedTimedValue <- PIExtendedTimedValue(timestamp = "7/22/2014 2:00:00 PM", unitsAbbreviation = "m", good = FALSE, questionable = FALSE, substituted = FALSE, annotated = FALSE)

}
