% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIExtendedTimedValues.r
\name{PIExtendedTimedValues}
\alias{PIExtendedTimedValues}
\title{Generate an instance of the PIExtendedTimedValues PI Web API class}
\usage{
PIExtendedTimedValues(items = NULL, unitsAbbreviation = NULL,
  webException = NULL)
}
\arguments{
\item{items}{(array)}

\item{unitsAbbreviation}{(string)}

\item{webException}{()}
}
\value{
PIExtendedTimedValues
}
\description{
Generate an instance of the PIExtendedTimedValues PI Web API class
}
\examples{
extendedTimedValues <- PIExtendedTimedValues(unitsAbbreviation = "")

}
