% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemAttribute.r
\name{piItemAttribute}
\alias{piItemAttribute}
\title{Generate an instance of the piItemAttribute PI Web API class}
\usage{
piItemAttribute(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
itemAttribute
}
\description{
Generate an instance of the piItemAttribute PI Web API class
}
\examples{
itemAttribute(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
