% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemElement.r
\name{piItemElement}
\alias{piItemElement}
\title{Generate an instance of the piItemElement PI Web API class}
\usage{
piItemElement(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
piItemElement
}
\description{
Generate an instance of the piItemElement PI Web API class
}
\examples{
itemElement <- piItemElement(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
