% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemEventFrame.r
\name{piItemEventFrame}
\alias{piItemEventFrame}
\title{Generate an instance of the piItemEventFrame PI Web API class}
\usage{
piItemEventFrame(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
piItemEventFrame
}
\description{
Generate an instance of the piItemEventFrame PI Web API class
}
\examples{
itemEventFrame <- piItemEventFrame(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
