% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemPoint.r
\name{piItemPoint}
\alias{piItemPoint}
\title{Generate an instance of the piItemPoint PI Web API class}
\usage{
piItemPoint(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
itemPoint
}
\description{
Generate an instance of the piItemPoint PI Web API class
}
\examples{
itemPoint(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
