% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAnalysis.r
\name{piItemsAnalysis}
\alias{piItemsAnalysis}
\title{Generate an instance of the piItemsAnalysis PI Web API class}
\usage{
piItemsAnalysis(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsAnalysis
}
\description{
Generate an instance of the piItemsAnalysis PI Web API class
}
