% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAnalysisCategory.r
\name{piItemsAnalysisCategory}
\alias{piItemsAnalysisCategory}
\title{Generate an instance of the piItemsAnalysisCategory PI Web API class}
\usage{
piItemsAnalysisCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAnalysisCategory
}
\description{
Generate an instance of the piItemsAnalysisCategory PI Web API class
}
