% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAnalysisRule.r
\name{piItemsAnalysisRule}
\alias{piItemsAnalysisRule}
\title{Generate an instance of the piItemsAnalysisRule PI Web API class}
\usage{
piItemsAnalysisRule(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAnalysisRule
}
\description{
Generate an instance of the piItemsAnalysisRule PI Web API class
}
