% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAnalysisRulePlugIn.r
\name{piItemsAnalysisRulePlugIn}
\alias{piItemsAnalysisRulePlugIn}
\title{Generate an instance of the piItemsAnalysisRulePlugIn PI Web API class}
\usage{
piItemsAnalysisRulePlugIn(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsAnalysisRulePlugIn
}
\description{
Generate an instance of the piItemsAnalysisRulePlugIn PI Web API class
}
