% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAnalysisTemplate.r
\name{piItemsAnalysisTemplate}
\alias{piItemsAnalysisTemplate}
\title{Generate an instance of the piItemsAnalysisTemplate PI Web API class}
\usage{
piItemsAnalysisTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsAnalysisTemplate
}
\description{
Generate an instance of the piItemsAnalysisTemplate PI Web API class
}
