% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAnnotation.r
\name{piItemsAnnotation}
\alias{piItemsAnnotation}
\title{Generate an instance of the piItemsAnnotation PI Web API class}
\usage{
piItemsAnnotation(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAnnotation
}
\description{
Generate an instance of the piItemsAnnotation PI Web API class
}
