% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAssetDatabase.r
\name{piItemsAssetDatabase}
\alias{piItemsAssetDatabase}
\title{Generate an instance of the piItemsAssetDatabase PI Web API class}
\usage{
piItemsAssetDatabase(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAssetDatabase
}
\description{
Generate an instance of the piItemsAssetDatabase PI Web API class
}
