% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAttribute.r
\name{piItemsAttribute}
\alias{piItemsAttribute}
\title{Generate an instance of the piItemsAttribute PI Web API class}
\usage{
piItemsAttribute(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAttribute
}
\description{
Generate an instance of the piItemsAttribute PI Web API class
}
