% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAttributeCategory.r
\name{piItemsAttributeCategory}
\alias{piItemsAttributeCategory}
\title{Generate an instance of the piItemsAttributeCategory PI Web API class}
\usage{
piItemsAttributeCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsAttributeCategory
}
\description{
Generate an instance of the piItemsAttributeCategory PI Web API class
}
