% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAttributeTemplate.r
\name{piItemsAttributeTemplate}
\alias{piItemsAttributeTemplate}
\title{Generate an instance of the piItemsAttributeTemplate PI Web API class}
\usage{
piItemsAttributeTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsAttributeTemplate
}
\description{
Generate an instance of the piItemsAttributeTemplate PI Web API class
}
