% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAttributeTrait.r
\name{piItemsAttributeTrait}
\alias{piItemsAttributeTrait}
\title{Generate an instance of the piItemsAttributeTrait PI Web API class}
\usage{
piItemsAttributeTrait(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAttributeTrait
}
\description{
Generate an instance of the piItemsAttributeTrait PI Web API class
}
