% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsCacheInstance.r
\name{piItemsCacheInstance}
\alias{piItemsCacheInstance}
\title{Generate an instance of the piItemsCacheInstance PI Web API class}
\usage{
piItemsCacheInstance(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsCacheInstance
}
\description{
Generate an instance of the piItemsCacheInstance PI Web API class
}
