% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsChannelInstance.r
\name{PIItemsChannelInstance}
\alias{PIItemsChannelInstance}
\title{Generate an instance of the PIItemsChannelInstance PI Web API class}
\usage{
PIItemsChannelInstance(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsChannelInstance
}
\description{
Generate an instance of the PIItemsChannelInstance PI Web API class
}
