% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsDataServer.r
\name{piItemsDataServer}
\alias{piItemsDataServer}
\title{Generate an instance of the piItemsDataServer PI Web API class}
\usage{
piItemsDataServer(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsDataServer
}
\description{
Generate an instance of the piItemsDataServer PI Web API class
}
