% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsElement.r
\name{piItemsElement}
\alias{piItemsElement}
\title{Generate an instance of the piItemsElement PI Web API class}
\usage{
piItemsElement(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsElement
}
\description{
Generate an instance of the piItemsElement PI Web API class
}
