% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsElementCategory.r
\name{piItemsElementCategory}
\alias{piItemsElementCategory}
\title{Generate an instance of the piItemsElementCategory PI Web API class}
\usage{
piItemsElementCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsElementCategory
}
\description{
Generate an instance of the piItemsElementCategory PI Web API class
}
