% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsElementTemplate.r
\name{piItemsElementTemplate}
\alias{piItemsElementTemplate}
\title{Generate an instance of the piItemsElementTemplate PI Web API class}
\usage{
piItemsElementTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsElementTemplate
}
\description{
Generate an instance of the piItemsElementTemplate PI Web API class
}
