% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsEnumerationSet.r
\name{PIItemsEnumerationSet}
\alias{PIItemsEnumerationSet}
\title{Generate an instance of the PIItemsEnumerationSet PI Web API class}
\usage{
PIItemsEnumerationSet(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsEnumerationSet
}
\description{
Generate an instance of the PIItemsEnumerationSet PI Web API class
}
