% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsEnumerationValue.r
\name{piItemsEnumerationValue}
\alias{piItemsEnumerationValue}
\title{Generate an instance of the piItemsEnumerationValue PI Web API class}
\usage{
piItemsEnumerationValue(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsEnumerationValue
}
\description{
Generate an instance of the piItemsEnumerationValue PI Web API class
}
