% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsEventFrame.r
\name{piItemsEventFrame}
\alias{piItemsEventFrame}
\title{Generate an instance of the piItemsEventFrame PI Web API class}
\usage{
piItemsEventFrame(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsEventFrame
}
\description{
Generate an instance of the piItemsEventFrame PI Web API class
}
