% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsItemAttribute.r
\name{piItemsItemAttribute}
\alias{piItemsItemAttribute}
\title{Generate an instance of the piItemsItemAttribute PI Web API class}
\usage{
piItemsItemAttribute(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsItemAttribute
}
\description{
Generate an instance of the piItemsItemAttribute PI Web API class
}
