% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsItemElement.r
\name{piItemsItemElement}
\alias{piItemsItemElement}
\title{Generate an instance of the piItemsItemElement PI Web API class}
\usage{
piItemsItemElement(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsItemElement
}
\description{
Generate an instance of the piItemsItemElement PI Web API class
}
