% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsItemEventFrame.r
\name{piItemsItemEventFrame}
\alias{piItemsItemEventFrame}
\title{Generate an instance of the piItemsItemEventFrame PI Web API class}
\usage{
piItemsItemEventFrame(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsItemEventFrame
}
\description{
Generate an instance of the piItemsItemEventFrame PI Web API class
}
