% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsItemPoint.r
\name{piItemsItemPoint}
\alias{piItemsItemPoint}
\title{Generate an instance of the piItemsItemPoint PI Web API class}
\usage{
piItemsItemPoint(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsItemPoint
}
\description{
Generate an instance of the piItemsItemPoint PI Web API class
}
