% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsItemsSubstatus.r
\name{piItemsItemsSubstatus}
\alias{piItemsItemsSubstatus}
\title{Generate an instance of the piItemsItemsSubstatus PI Web API class}
\usage{
piItemsItemsSubstatus(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsItemsSubstatus
}
\description{
Generate an instance of the piItemsItemsSubstatus PI Web API class
}
