% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsNotificationContactTemplate.r
\name{PIItemsNotificationContactTemplate}
\alias{PIItemsNotificationContactTemplate}
\title{Generate an instance of the PIItemsNotificationContactTemplate PI Web API class}
\usage{
PIItemsNotificationContactTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsNotificationContactTemplate
}
\description{
Generate an instance of the PIItemsNotificationContactTemplate PI Web API class
}
