% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsNotificationRule.r
\name{PIItemsNotificationRule}
\alias{PIItemsNotificationRule}
\title{Generate an instance of the PIItemsNotificationRule PI Web API class}
\usage{
PIItemsNotificationRule(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsNotificationRule
}
\description{
Generate an instance of the PIItemsNotificationRule PI Web API class
}
