% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsNotificationRuleSubscriber.r
\name{PIItemsNotificationRuleSubscriber}
\alias{PIItemsNotificationRuleSubscriber}
\title{Generate an instance of the PIItemsNotificationRuleSubscriber PI Web API class}
\usage{
PIItemsNotificationRuleSubscriber(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsNotificationRuleSubscriber
}
\description{
Generate an instance of the PIItemsNotificationRuleSubscriber PI Web API class
}
