% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsNotificationRuleTemplate.r
\name{PIItemsNotificationRuleTemplate}
\alias{PIItemsNotificationRuleTemplate}
\title{Generate an instance of the PIItemsNotificationRuleTemplate PI Web API class}
\usage{
PIItemsNotificationRuleTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsNotificationRuleTemplate
}
\description{
Generate an instance of the PIItemsNotificationRuleTemplate PI Web API class
}
