% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsPoint.r
\name{piItemsPoint}
\alias{piItemsPoint}
\title{Generate an instance of the piItemsPoint PI Web API class}
\usage{
piItemsPoint(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsPoint
}
\description{
Generate an instance of the piItemsPoint PI Web API class
}
