% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsPointAttribute.r
\name{piItemsPointAttribute}
\alias{piItemsPointAttribute}
\title{Generate an instance of the piItemsPointAttribute PI Web API class}
\usage{
piItemsPointAttribute(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsPointAttribute
}
\description{
Generate an instance of the piItemsPointAttribute PI Web API class
}
