% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsSecurityEntry.r
\name{piItemsSecurityEntry}
\alias{piItemsSecurityEntry}
\title{Generate an instance of the piItemsSecurityEntry PI Web API class}
\usage{
piItemsSecurityEntry(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsSecurityEntry
}
\description{
Generate an instance of the piItemsSecurityEntry PI Web API class
}
