% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsSecurityIdentity.r
\name{piItemsSecurityIdentity}
\alias{piItemsSecurityIdentity}
\title{Generate an instance of the piItemsSecurityIdentity PI Web API class}
\usage{
piItemsSecurityIdentity(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsSecurityIdentity
}
\description{
Generate an instance of the piItemsSecurityIdentity PI Web API class
}
