% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsSecurityMapping.r
\name{piItemsSecurityMapping}
\alias{piItemsSecurityMapping}
\title{Generate an instance of the piItemsSecurityMapping PI Web API class}
\usage{
piItemsSecurityMapping(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsSecurityMapping
}
\description{
Generate an instance of the piItemsSecurityMapping PI Web API class
}
