% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsSecurityRights.r
\name{piItemsSecurityRights}
\alias{piItemsSecurityRights}
\title{Generate an instance of the piItemsSecurityRights PI Web API class}
\usage{
piItemsSecurityRights(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsSecurityRights
}
\description{
Generate an instance of the piItemsSecurityRights PI Web API class
}
