% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsStreamSummaries.r
\name{piItemsStreamSummaries}
\alias{piItemsStreamSummaries}
\title{Generate an instance of the piItemsStreamSummaries PI Web API class}
\usage{
piItemsStreamSummaries(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsStreamSummaries
}
\description{
Generate an instance of the piItemsStreamSummaries PI Web API class
}
