% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsStreamUpdatesRegister.r
\name{PIItemsStreamUpdatesRegister}
\alias{PIItemsStreamUpdatesRegister}
\title{Generate an instance of the PIItemsStreamUpdatesRegister PI Web API class}
\usage{
PIItemsStreamUpdatesRegister(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsStreamUpdatesRegister
}
\description{
Generate an instance of the PIItemsStreamUpdatesRegister PI Web API class
}
