% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsStreamUpdatesRetrieve.r
\name{PIItemsStreamUpdatesRetrieve}
\alias{PIItemsStreamUpdatesRetrieve}
\title{Generate an instance of the PIItemsStreamUpdatesRetrieve PI Web API class}
\usage{
PIItemsStreamUpdatesRetrieve(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsStreamUpdatesRetrieve
}
\description{
Generate an instance of the PIItemsStreamUpdatesRetrieve PI Web API class
}
