% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsStreamValue.r
\name{piItemsStreamValue}
\alias{piItemsStreamValue}
\title{Generate an instance of the piItemsStreamValue PI Web API class}
\usage{
piItemsStreamValue(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsStreamValue
}
\description{
Generate an instance of the piItemsStreamValue PI Web API class
}
