% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsStreamValues.r
\name{piItemsStreamValues}
\alias{piItemsStreamValues}
\title{Generate an instance of the piItemsStreamValues PI Web API class}
\usage{
piItemsStreamValues(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsStreamValues
}
\description{
Generate an instance of the piItemsStreamValues PI Web API class
}
