% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsstring.r
\name{PIItemsstring}
\alias{PIItemsstring}
\title{Generate an instance of the PIItemsstring PI Web API class}
\usage{
PIItemsstring(items = NULL)
}
\arguments{
\item{items}{(array)}
}
\value{
PIItemsstring
}
\description{
Generate an instance of the PIItemsstring PI Web API class
}
