% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsSubstatus.r
\name{piItemsSubstatus}
\alias{piItemsSubstatus}
\title{Generate an instance of the piItemsSubstatus PI Web API class}
\usage{
piItemsSubstatus(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsSubstatus
}
\description{
Generate an instance of the piItemsSubstatus PI Web API class
}
