% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsSummaryValue.r
\name{piItemsSummaryValue}
\alias{piItemsSummaryValue}
\title{Generate an instance of the piItemsSummaryValue PI Web API class}
\usage{
piItemsSummaryValue(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsSummaryValue
}
\description{
Generate an instance of the piItemsSummaryValue PI Web API class
}
