% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsTable.r
\name{piItemsTable}
\alias{piItemsTable}
\title{Generate an instance of the piItemsTable PI Web API class}
\usage{
piItemsTable(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsTable
}
\description{
Generate an instance of the piItemsTable PI Web API class
}
