% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsTableCategory.r
\name{piItemsTableCategory}
\alias{piItemsTableCategory}
\title{Generate an instance of the piItemsTableCategory PI Web API class}
\usage{
piItemsTableCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsTableCategory
}
\description{
Generate an instance of the piItemsTableCategory PI Web API class
}
