% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsTimeRulePlugIn.r
\name{piItemsTimeRulePlugIn}
\alias{piItemsTimeRulePlugIn}
\title{Generate an instance of the piItemsTimeRulePlugIn PI Web API class}
\usage{
piItemsTimeRulePlugIn(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
itemsTimeRulePlugIn
}
\description{
Generate an instance of the piItemsTimeRulePlugIn PI Web API class
}
