% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsUnitClass.r
\name{piItemsUnitClass}
\alias{piItemsUnitClass}
\title{Generate an instance of the piItemsUnitClass PI Web API class}
\usage{
piItemsUnitClass(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsUnitClass
}
\description{
Generate an instance of the piItemsUnitClass PI Web API class
}
