% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piLanding.r
\name{piLanding}
\alias{piLanding}
\title{Generate an instance of the piLanding PI Web API class}
\usage{
piLanding(links = NULL)
}
\arguments{
\item{links}{(object)}
}
\value{
landing
}
\description{
Generate an instance of the piLanding PI Web API class
}
