% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILandingLinks.r
\name{PILandingLinks}
\alias{PILandingLinks}
\title{Generate an instance of the PILandingLinks PI Web API class}
\usage{
PILandingLinks(self = NULL, assetServers = NULL, dataServers = NULL,
  search = NULL, system = NULL)
}
\arguments{
\item{self}{(string)}

\item{assetServers}{(string)}

\item{dataServers}{(string)}

\item{search}{(string)}

\item{system}{(string)}
}
\value{
PILandingLinks
}
\description{
Generate an instance of the PILandingLinks PI Web API class
}
\examples{
landingLinks <- PILandingLinks(self = "", assetServers = "", dataServers = "", search = "", system = "")

}
