% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIMediaMetadata.r
\name{PIMediaMetadata}
\alias{PIMediaMetadata}
\title{Generate an instance of the PIMediaMetadata PI Web API class}
\usage{
PIMediaMetadata(author = NULL, changeDate = NULL, description = NULL,
  name = NULL, size = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{author}{(string)}

\item{changeDate}{(string)}

\item{description}{(string)}

\item{name}{(string)}

\item{size}{(number)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIMediaMetadata
}
\description{
Generate an instance of the PIMediaMetadata PI Web API class
}
\examples{
mediaMetadata <- PIMediaMetadata(author = "DOMAIN\\\\username", changeDate = "7/22/2014 2:00:00 PM", description = "The description attached to this media.", name = "ExampleFile.txt")

}
