% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIMediaMetadataLinks.r
\name{PIMediaMetadataLinks}
\alias{PIMediaMetadataLinks}
\title{Generate an instance of the PIMediaMetadataLinks PI Web API class}
\usage{
PIMediaMetadataLinks(self = NULL, owner = NULL, mediaData = NULL)
}
\arguments{
\item{self}{(string)}

\item{owner}{(string)}

\item{mediaData}{(string)}
}
\value{
PIMediaMetadataLinks
}
\description{
Generate an instance of the PIMediaMetadataLinks PI Web API class
}
\examples{
mediaMetadataLinks <- PIMediaMetadataLinks(self = "", owner = "", mediaData = "")

}
