% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINotificationContactTemplate.r
\name{PINotificationContactTemplate}
\alias{PINotificationContactTemplate}
\title{Generate an instance of the PINotificationContactTemplate PI Web API class}
\usage{
PINotificationContactTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, available = NULL,
  configString = NULL, contactType = NULL, plugInName = NULL,
  hasChildren = NULL, maximumRetries = NULL,
  minimumAcknowledgements = NULL, notifyWhenInstanceEnded = NULL,
  escalationTimeout = NULL, retryInterval = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{available}{(boolean)}

\item{configString}{(string)}

\item{contactType}{(string)}

\item{plugInName}{(string)}

\item{hasChildren}{(boolean)}

\item{maximumRetries}{(integer)}

\item{minimumAcknowledgements}{(integer)}

\item{notifyWhenInstanceEnded}{(boolean)}

\item{escalationTimeout}{(string)}

\item{retryInterval}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PINotificationContactTemplate
}
\description{
Generate an instance of the PINotificationContactTemplate PI Web API class
}
\examples{
notificationContactTemplate <- PINotificationContactTemplate(webId = "I1NCEDqD5loBNH0erqeqJodtALAYIKyyz2F5BGAxQAVXYRDBAGyPedZG1sUmxOOclp3Flwg", id = "cbb28260-853d-11e4-80c5-00155d844304", name = "Plant Manager", description = "", path = "\\\\\\\\MyAssetServer\\\\NotificationContactTemplates[Plant Manager]", available = FALSE, configString = "", contactType = "Individual", plugInName = "", hasChildren = FALSE, maximumRetries = 5, minimumAcknowledgements = 2, notifyWhenInstanceEnded = FALSE, escalationTimeout = "PT3S", retryInterval = "PT5S")

}
