% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINotificationContactTemplateLinks.r
\name{PINotificationContactTemplateLinks}
\alias{PINotificationContactTemplateLinks}
\title{Generate an instance of the PINotificationContactTemplateLinks PI Web API class}
\usage{
PINotificationContactTemplateLinks(self = NULL, assetServer = NULL)
}
\arguments{
\item{self}{(string)}

\item{assetServer}{(string)}
}
\value{
PINotificationContactTemplateLinks
}
\description{
Generate an instance of the PINotificationContactTemplateLinks PI Web API class
}
\examples{
notificationContactTemplateLinks <- PINotificationContactTemplateLinks(self = "", assetServer = "")

}
