% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINotificationRule.r
\name{PINotificationRule}
\alias{PINotificationRule}
\title{Generate an instance of the PINotificationRule PI Web API class}
\usage{
PINotificationRule(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, autoCreated = NULL,
  categoryNames = NULL, criteria = NULL, multiTriggerEventOption = NULL,
  nonrepetitionInterval = NULL, resendInterval = NULL, status = NULL,
  templateName = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{autoCreated}{(boolean)}

\item{categoryNames}{(array)}

\item{criteria}{(string)}

\item{multiTriggerEventOption}{(string)}

\item{nonrepetitionInterval}{(string)}

\item{resendInterval}{(string)}

\item{status}{(string)}

\item{templateName}{(string)}

\item{webException}{()}
}
\value{
PINotificationRule
}
\description{
Generate an instance of the PINotificationRule PI Web API class
}
\examples{
notificationRule <- PINotificationRule(webId = "I1NRDqD5loBNH0erqeqJodtALA5bYfWno26BGgMQAVXYR0Ag", id = "e9a984d0-f47c-11e7-8454-00155d029708", name = "MachineName", description = "Manufacturing Equipment MachineName", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\CityName\\\\EngineeringProcess\\\\Equipment\\\\MachineName\\\\NotificationRules[RuleName]", autoCreated = FALSE, criteria = "Name: EventFrameCriteriaName", multiTriggerEventOption = "HighestSeverity", nonrepetitionInterval = "PT3S", resendInterval = "PT5S", status = "Disabled", templateName = "MachineName Notification Rule")

}
