% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINotificationRuleSubscriber.r
\name{PINotificationRuleSubscriber}
\alias{PINotificationRuleSubscriber}
\title{Generate an instance of the PINotificationRuleSubscriber PI Web API class}
\usage{
PINotificationRuleSubscriber(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, configString = NULL,
  contactTemplateName = NULL, contactType = NULL,
  deliveryFormatName = NULL, plugInName = NULL, escalationTimeout = NULL,
  maximumRetries = NULL, notifyOption = NULL, retryInterval = NULL,
  webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{configString}{(string)}

\item{contactTemplateName}{(string)}

\item{contactType}{(string)}

\item{deliveryFormatName}{(string)}

\item{plugInName}{(string)}

\item{escalationTimeout}{(string)}

\item{maximumRetries}{(integer)}

\item{notifyOption}{(string)}

\item{retryInterval}{(string)}

\item{webException}{()}
}
\value{
PINotificationRuleSubscriber
}
\description{
Generate an instance of the PINotificationRuleSubscriber PI Web API class
}
\examples{
notificationRuleSubscriber <- PINotificationRuleSubscriber(webId = "I1NSLDqD5loBNH0erqeqJodtALA5bYfWno26BGgMQAVXYR0AgPUJJXNlEW1w78rnCXDmcDA", id = "e9a984d0-f47c-11e7-8454-00155d029708", name = "NotificationSubscriptionRuleName", description = "Manufacturing Machine MachineName subscription", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\NotificationRules[NotificationSubscriptionRuleName]", configString = "ToEmail=mike@testemail.com", contactTemplateName = "Plant Manager", contactType = "Individual", deliveryFormatName = "DeliveryFormat", plugInName = "Email", escalationTimeout = "PT3S", maximumRetries = 3, notifyOption = "EventStart", retryInterval = "PT3S")

}
