% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PINotificationRuleTemplate.r
\name{PINotificationRuleTemplate}
\alias{PINotificationRuleTemplate}
\title{Generate an instance of the PINotificationRuleTemplate PI Web API class}
\usage{
PINotificationRuleTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, categoryNames = NULL,
  criteria = NULL, multiTriggerEventOption = NULL,
  nonrepetitionInterval = NULL, resendInterval = NULL, status = NULL,
  templateName = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{categoryNames}{(array)}

\item{criteria}{(string)}

\item{multiTriggerEventOption}{(string)}

\item{nonrepetitionInterval}{(string)}

\item{resendInterval}{(string)}

\item{status}{(string)}

\item{templateName}{(string)}

\item{webException}{()}
}
\value{
PINotificationRuleTemplate
}
\description{
Generate an instance of the PINotificationRuleTemplate PI Web API class
}
\examples{
notificationRuleTemplate <- PINotificationRuleTemplate(webId = "I1NTDqD5loBNH0erqeqJodtALAtdcX5JH_5xGEKAAVXTSaAg", id = "e9a984d0-f47c-11e7-8454-00155d029708", name = "MachineName", description = "Description entered by the user", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\CityName\\\\ElementTemplates[Template3]\\\\NotificationRuleTemplates[Notification Rule Template]", criteria = "Name: EventFrameCriteriaName", multiTriggerEventOption = "HighestSeverity", nonrepetitionInterval = "PT3S", resendInterval = "PT5S", status = "Functioning", templateName = "MachineName Notification Rule Template")

}
