% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPoint.r
\name{PIPoint}
\alias{PIPoint}
\title{Generate an instance of the PIPoint PI Web API class}
\usage{
PIPoint(webId = NULL, id = NULL, name = NULL, path = NULL,
  descriptor = NULL, pointClass = NULL, pointType = NULL,
  digitalSetName = NULL, engineeringUnits = NULL, step = NULL,
  future = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(integer)}

\item{name}{(string)}

\item{path}{(string)}

\item{descriptor}{(string)}

\item{pointClass}{(string)}

\item{pointType}{(string)}

\item{digitalSetName}{(string)}

\item{engineeringUnits}{(string)}

\item{step}{(boolean)}

\item{future}{(boolean)}

\item{links}{(object)}
}
\value{
PIPoint
}
\description{
Generate an instance of the PIPoint PI Web API class
}
\examples{
point <- PIPoint(webId = "P0ZkdQMwx9oUCCzSWmnpgoEwTB8AAAUEhMVlpOXDAwR0FMDjBDRjAwMV9YUTAx", id = 8012, name = "PointName", path = "\\\\\\\\MyPIServer\\\\PointName", descriptor = "12 Hour Sine Wave", pointClass = "classic", pointType = "Float32", digitalSetName = "", engineeringUnits = "", step = FALSE, future = FALSE)

}
