% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piPointAttribute.r
\name{piPointAttribute}
\alias{piPointAttribute}
\title{Generate an instance of the piPointAttribute PI Web API class}
\usage{
piPointAttribute(name = NULL, value = NULL, links = NULL)
}
\arguments{
\item{name}{(string)}

\item{value}{(object)}

\item{links}{(object)}
}
\value{
piPointAttribute
}
\description{
Generate an instance of the piPointAttribute PI Web API class
}
\examples{
pointAttribute <- piPointAttribute(name = "typicalvalue")

}
