% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPointAttributeLinks.r
\name{PIPointAttributeLinks}
\alias{PIPointAttributeLinks}
\title{Generate an instance of the PIPointAttributeLinks PI Web API class}
\usage{
PIPointAttributeLinks(self = NULL, point = NULL)
}
\arguments{
\item{self}{(string)}

\item{point}{(string)}
}
\value{
PIPointAttributeLinks
}
\description{
Generate an instance of the PIPointAttributeLinks PI Web API class
}
\examples{
pointAttributeLinks <- PIPointAttributeLinks(self = "", point = "")

}
