% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPointLinks.r
\name{PIPointLinks}
\alias{PIPointLinks}
\title{Generate an instance of the PIPointLinks PI Web API class}
\usage{
PIPointLinks(self = NULL, dataServer = NULL, attributes = NULL,
  interpolatedData = NULL, recordedData = NULL, plotData = NULL,
  summaryData = NULL, value = NULL, endValue = NULL)
}
\arguments{
\item{self}{(string)}

\item{dataServer}{(string)}

\item{attributes}{(string)}

\item{interpolatedData}{(string)}

\item{recordedData}{(string)}

\item{plotData}{(string)}

\item{summaryData}{(string)}

\item{value}{(string)}

\item{endValue}{(string)}
}
\value{
PIPointLinks
}
\description{
Generate an instance of the PIPointLinks PI Web API class
}
\examples{
pointLinks <- PIPointLinks(self = "", dataServer = "", attributes = "", interpolatedData = "", recordedData = "", plotData = "", summaryData = "", value = "", endValue = "")

}
