% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPropertyError.r
\name{PIPropertyError}
\alias{PIPropertyError}
\title{Generate an instance of the PIPropertyError PI Web API class}
\usage{
PIPropertyError(fieldName = NULL, message = NULL)
}
\arguments{
\item{fieldName}{(string)}

\item{message}{(array)}
}
\value{
PIPropertyError
}
\description{
Generate an instance of the PIPropertyError PI Web API class
}
\examples{
propertyError <- PIPropertyError(fieldName = "Value")

}
