% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRequest.r
\name{piRequest}
\alias{piRequest}
\title{Generate an instance of the piRequest PI Web API class}
\usage{
piRequest(method = NULL, resource = NULL, requestTemplate = NULL,
  parameters = NULL, headers = NULL, content = NULL, parentIds = NULL)
}
\arguments{
\item{method}{(string)}

\item{resource}{(string)}

\item{requestTemplate}{()}

\item{parameters}{(array)}

\item{headers}{(object)}

\item{content}{(string)}

\item{parentIds}{(array)}
}
\value{
piRequest
}
\description{
Generate an instance of the piRequest PI Web API class
}
\examples{
request <- piRequest(method = "GET", resource = "", content = "")

}
