% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piRequestTemplate.r
\name{piRequestTemplate}
\alias{piRequestTemplate}
\title{Generate an instance of the piRequestTemplate PI Web API class}
\usage{
piRequestTemplate(resource = NULL)
}
\arguments{
\item{resource}{(string)}
}
\value{
piRequestTemplate
}
\description{
Generate an instance of the piRequestTemplate PI Web API class
}
\examples{
requestTemplate <- piRequestTemplate(resource = "")

}
