% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piResponse.r
\name{piResponse}
\alias{piResponse}
\title{Generate an instance of the piResponse PI Web API class}
\usage{
piResponse(status = NULL, headers = NULL, content = NULL)
}
\arguments{
\item{status}{(integer)}

\item{headers}{(object)}

\item{content}{(object)}
}
\value{
response
}
\description{
Generate an instance of the piResponse PI Web API class
}
\examples{
response(status = 200)

}
