% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISearchByAttribute.r
\name{PISearchByAttribute}
\alias{PISearchByAttribute}
\title{Generate an instance of the PISearchByAttribute PI Web API class}
\usage{
PISearchByAttribute(searchRoot = NULL, elementTemplate = NULL,
  items = NULL, webException = NULL)
}
\arguments{
\item{searchRoot}{(string)}

\item{elementTemplate}{(string)}

\item{items}{(array)}

\item{webException}{()}
}
\value{
PISearchByAttribute
}
\description{
Generate an instance of the PISearchByAttribute PI Web API class
}
\examples{
searchByAttribute <- PISearchByAttribute(searchRoot = "F1FmDqD5loBNH0erqeqJodtALADqD5loBNH0cAAAAAACR1_wUkVTVFVOSVRcQUVGVEVTVFxFVkVOVEZSQU1FU1t-QC9FRl4jXQ", elementTemplate = "F1ETDqD5loBNH0erqeqJodtALAkzfYXwSS9kmTgO1h1pM7ngUkVTVFVOSVRcQUVGVEVTVFxFTEVNRU5UVEVNUExBVEVTW1hYWFhd")

}
