% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSecurity.r
\name{piSecurity}
\alias{piSecurity}
\title{Generate an instance of the piSecurity PI Web API class}
\usage{
piSecurity(canAnnotate = NULL, canDelete = NULL, canExecute = NULL,
  canRead = NULL, canReadData = NULL, canSubscribe = NULL,
  canSubscribeOthers = NULL, canWrite = NULL, canWriteData = NULL,
  hasAdmin = NULL, rights = NULL)
}
\arguments{
\item{canAnnotate}{(boolean)}

\item{canDelete}{(boolean)}

\item{canExecute}{(boolean)}

\item{canRead}{(boolean)}

\item{canReadData}{(boolean)}

\item{canSubscribe}{(boolean)}

\item{canSubscribeOthers}{(boolean)}

\item{canWrite}{(boolean)}

\item{canWriteData}{(boolean)}

\item{hasAdmin}{(boolean)}

\item{rights}{(array)}
}
\value{
security
}
\description{
Generate an instance of the piSecurity PI Web API class
}
\examples{
security(canAnnotate = FALSE, canDelete = FALSE, canExecute = FALSE, canRead = FALSE, canReadData = FALSE, canSubscribe = FALSE, canSubscribeOthers = FALSE, canWrite = FALSE, canWriteData = FALSE, hasAdmin = FALSE)

}
