% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSecurityEntry.r
\name{piSecurityEntry}
\alias{piSecurityEntry}
\title{Generate an instance of the piSecurityEntry PI Web API class}
\usage{
piSecurityEntry(name = NULL, securityIdentityName = NULL,
  allowRights = NULL, denyRights = NULL, links = NULL)
}
\arguments{
\item{name}{(string)}

\item{securityIdentityName}{(string)}

\item{allowRights}{(array)}

\item{denyRights}{(array)}

\item{links}{(object)}
}
\value{
securityEntry
}
\description{
Generate an instance of the piSecurityEntry PI Web API class
}
\examples{
securityEntry(name = "domain\\\\user1", securityIdentityName = "domain\\\\user1")

}
