% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityEntryLinks.r
\name{PISecurityEntryLinks}
\alias{PISecurityEntryLinks}
\title{Generate an instance of the PISecurityEntryLinks PI Web API class}
\usage{
PISecurityEntryLinks(self = NULL, securableObject = NULL,
  securityIdentity = NULL)
}
\arguments{
\item{self}{(string)}

\item{securableObject}{(string)}

\item{securityIdentity}{(string)}
}
\value{
PISecurityEntryLinks
}
\description{
Generate an instance of the PISecurityEntryLinks PI Web API class
}
\examples{
securityEntryLinks <- PISecurityEntryLinks(self = "", securableObject = "", securityIdentity = "")

}
