% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityIdentity.r
\name{PISecurityIdentity}
\alias{PISecurityIdentity}
\title{Generate an instance of the PISecurityIdentity PI Web API class}
\usage{
PISecurityIdentity(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, isEnabled = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{isEnabled}{(boolean)}

\item{links}{(object)}
}
\value{
PISecurityIdentity
}
\description{
Generate an instance of the PISecurityIdentity PI Web API class
}
\examples{
securityIdentity <- PISecurityIdentity(webId = "I07v5To0Ux7Eus2sjMsdqq_AUEhMQUZTMDU", id = "a353feee-3145-4bec-acda-c8ccb1daaafc", name = "MySecurityIdentity", description = "", path = "\\\\\\\\RESTUNIT\\\\SecurityIdentities[MySecurityIdentity]", isEnabled = FALSE)

}
