% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityIdentityLinks.r
\name{PISecurityIdentityLinks}
\alias{PISecurityIdentityLinks}
\title{Generate an instance of the PISecurityIdentityLinks PI Web API class}
\usage{
PISecurityIdentityLinks(self = NULL, assetServer = NULL,
  securityMappings = NULL, security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{assetServer}{(string)}

\item{securityMappings}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PISecurityIdentityLinks
}
\description{
Generate an instance of the PISecurityIdentityLinks PI Web API class
}
\examples{
securityIdentityLinks <- PISecurityIdentityLinks(self = "", assetServer = "", securityMappings = "", security = "", securityEntries = "")

}
