% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityMappingLinks.r
\name{PISecurityMappingLinks}
\alias{PISecurityMappingLinks}
\title{Generate an instance of the PISecurityMappingLinks PI Web API class}
\usage{
PISecurityMappingLinks(self = NULL, assetServer = NULL,
  securityIdentity = NULL, security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{assetServer}{(string)}

\item{securityIdentity}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PISecurityMappingLinks
}
\description{
Generate an instance of the PISecurityMappingLinks PI Web API class
}
\examples{
securityMappingLinks <- PISecurityMappingLinks(self = "", assetServer = "", securityIdentity = "", security = "", securityEntries = "")

}
