% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityRights.r
\name{PISecurityRights}
\alias{PISecurityRights}
\title{Generate an instance of the PISecurityRights PI Web API class}
\usage{
PISecurityRights(ownerWebId = NULL, securityItem = NULL,
  userIdentity = NULL, links = NULL)
}
\arguments{
\item{ownerWebId}{(string)}

\item{securityItem}{(string)}

\item{userIdentity}{(string)}

\item{links}{(object)}
}
\value{
PISecurityRights
}
\description{
Generate an instance of the PISecurityRights PI Web API class
}
\examples{
securityRights <- PISecurityRights(ownerWebId = "E0DqD5loBNH0erqeqJodtALAvGEqktDl4xG-twAFmjx6UAUkVTVFVOSVRUc1VHQVJNQU5V1TcHUkVFTg", securityItem = "Default", userIdentity = "domain\\\\user")

}
