% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityRightsLinks.r
\name{PISecurityRightsLinks}
\alias{PISecurityRightsLinks}
\title{Generate an instance of the PISecurityRightsLinks PI Web API class}
\usage{
PISecurityRightsLinks(self = NULL, owner = NULL)
}
\arguments{
\item{self}{(string)}

\item{owner}{(string)}
}
\value{
PISecurityRightsLinks
}
\description{
Generate an instance of the PISecurityRightsLinks PI Web API class
}
\examples{
securityRightsLinks <- PISecurityRightsLinks(self = "", owner = "")

}
