% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamSummaries.r
\name{PIStreamSummaries}
\alias{PIStreamSummaries}
\title{Generate an instance of the PIStreamSummaries PI Web API class}
\usage{
PIStreamSummaries(webId = NULL, name = NULL, path = NULL, items = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{items}{(array)}

\item{links}{(object)}
}
\value{
PIStreamSummaries
}
\description{
Generate an instance of the PIStreamSummaries PI Web API class
}
\examples{
streamSummaries <- PIStreamSummaries(webId = "A0EDqD5loBNH0erqeqJodtALAUvtkqIDl4xG-twAFmjx6AA79jxF7HiqFwpsgMBxwgWmgUkVTVFVOSVRcU1VNJQVHQU5cTlVHUkVZTlxIT1VTVE9f0FBMQU5UQ0lUWQ", name = "Water", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement|Water")

}
