% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamSummariesLinks.r
\name{PIStreamSummariesLinks}
\alias{PIStreamSummariesLinks}
\title{Generate an instance of the PIStreamSummariesLinks PI Web API class}
\usage{
PIStreamSummariesLinks(source = NULL)
}
\arguments{
\item{source}{(string)}
}
\value{
PIStreamSummariesLinks
}
\description{
Generate an instance of the PIStreamSummariesLinks PI Web API class
}
\examples{
streamSummariesLinks <- PIStreamSummariesLinks(source = "")

}
