% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamUpdatesRegister.r
\name{PIStreamUpdatesRegister}
\alias{PIStreamUpdatesRegister}
\title{Generate an instance of the PIStreamUpdatesRegister PI Web API class}
\usage{
PIStreamUpdatesRegister(status = NULL, source = NULL, sourceName = NULL,
  sourcePath = NULL, latestMarker = NULL, exception = NULL)
}
\arguments{
\item{status}{(string)}

\item{source}{(string)}

\item{sourceName}{(string)}

\item{sourcePath}{(string)}

\item{latestMarker}{(string)}

\item{exception}{()}
}
\value{
PIStreamUpdatesRegister
}
\description{
Generate an instance of the PIStreamUpdatesRegister PI Web API class
}
\examples{
streamUpdatesRegister <- PIStreamUpdatesRegister(status = "Succeeded", source = "I1AbEDqD5loBNH0erqeqJodtALA9iLxz4m_5RGAxgAVXYUACw_yrS377vX0uZVHTgkanusw", sourceName = "Water", sourcePath = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\MyElement|Water", latestMarker = "0cbbec466f3a42288c8c47112f54b852_12")

}
