% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamUpdatesRetrieve.r
\name{PIStreamUpdatesRetrieve}
\alias{PIStreamUpdatesRetrieve}
\title{Generate an instance of the PIStreamUpdatesRetrieve PI Web API class}
\usage{
PIStreamUpdatesRetrieve(source = NULL, sourceName = NULL,
  sourcePath = NULL, requestedMarker = NULL, latestMarker = NULL,
  status = NULL, events = NULL, exception = NULL)
}
\arguments{
\item{source}{(string)}

\item{sourceName}{(string)}

\item{sourcePath}{(string)}

\item{requestedMarker}{(string)}

\item{latestMarker}{(string)}

\item{status}{(string)}

\item{events}{(array)}

\item{exception}{()}
}
\value{
PIStreamUpdatesRetrieve
}
\description{
Generate an instance of the PIStreamUpdatesRetrieve PI Web API class
}
\examples{
streamUpdatesRetrieve <- PIStreamUpdatesRetrieve(source = "I1AbEDqD5loBNH0erqeqJodtALAYIKyyz2F5BGAxQAVXYRDBAGyPedZG1sUmxOOclp3Flwg", sourceName = "Water", sourcePath = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\MyElement|Water", requestedMarker = "0cbbec466f3a42288c8c47112f54b852_0", latestMarker = "0cbbec466f3a42288c8c47112f54b852_12", status = "Created")

}
