% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamValue.r
\name{PIStreamValue}
\alias{PIStreamValue}
\title{Generate an instance of the PIStreamValue PI Web API class}
\usage{
PIStreamValue(webId = NULL, name = NULL, path = NULL, value = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{value}{()}

\item{links}{(object)}
}
\value{
PIStreamValue
}
\description{
Generate an instance of the PIStreamValue PI Web API class
}
\examples{
streamValue <- PIStreamValue(webId = "A0EDqD5loBNH0erqeqJodtALAUvtkqIDl4xG-twAFmjx6AA79jxF7HiqFwpsgMBxwgWmgUkVTVFVOSVRcU1VNJQVHQU5cTlVHUkVZTlxIT1VTVE9f0FBMQU5UQ0lUWQ", name = "Water", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement|Water")

}
