% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamValueLinks.r
\name{PIStreamValueLinks}
\alias{PIStreamValueLinks}
\title{Generate an instance of the PIStreamValueLinks PI Web API class}
\usage{
PIStreamValueLinks(source = NULL)
}
\arguments{
\item{source}{(string)}
}
\value{
PIStreamValueLinks
}
\description{
Generate an instance of the PIStreamValueLinks PI Web API class
}
\examples{
streamValueLinks <- PIStreamValueLinks(source = "")

}
