% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamValues.r
\name{PIStreamValues}
\alias{PIStreamValues}
\title{Generate an instance of the PIStreamValues PI Web API class}
\usage{
PIStreamValues(webId = NULL, name = NULL, path = NULL, items = NULL,
  unitsAbbreviation = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{items}{(array)}

\item{unitsAbbreviation}{(string)}

\item{links}{(object)}
}
\value{
PIStreamValues
}
\description{
Generate an instance of the PIStreamValues PI Web API class
}
\examples{
streamValues <- PIStreamValues(webId = "A0EDqD5loBNH0erqeqJodtALAUvtkqIDl4xG-twAFmjx6AA79jxF7HiqFwpsgMBxwgWmgUkVTVFVOSVRcU1VNJQVHQU5cTlVHUkVZTlxIT1VTVE9f0FBMQU5UQ0lUWQ", name = "Water", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement|Water", unitsAbbreviation = "m")

}
