% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamValuesLinks.r
\name{PIStreamValuesLinks}
\alias{PIStreamValuesLinks}
\title{Generate an instance of the PIStreamValuesLinks PI Web API class}
\usage{
PIStreamValuesLinks(source = NULL)
}
\arguments{
\item{source}{(string)}
}
\value{
PIStreamValuesLinks
}
\description{
Generate an instance of the PIStreamValuesLinks PI Web API class
}
\examples{
streamValuesLinks <- PIStreamValuesLinks(source = "")

}
