% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSummaryValue.r
\name{piSummaryValue}
\alias{piSummaryValue}
\title{Generate an instance of the piSummaryValue PI Web API class}
\usage{
piSummaryValue(type = NULL, value = NULL)
}
\arguments{
\item{type}{(string)}

\item{value}{()}
}
\value{
piSummaryValue
}
\description{
Generate an instance of the piSummaryValue PI Web API class
}
\examples{
summaryValue <- piSummaryValue(type = "Total")

}
