% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSystemLanding.r
\name{piSystemLanding}
\alias{piSystemLanding}
\title{Generate an instance of the piSystemLanding PI Web API class}
\usage{
piSystemLanding(productTitle = NULL, productVersion = NULL, links = NULL)
}
\arguments{
\item{productTitle}{(string)}

\item{productVersion}{(string)}

\item{links}{(object)}
}
\value{
systemLanding
}
\description{
Generate an instance of the piSystemLanding PI Web API class
}
\examples{
systemLanding(productTitle = "PI Web API 2016 R2", productVersion = "1.8.0.201")

}
