% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISystemLandingLinks.r
\name{PISystemLandingLinks}
\alias{PISystemLandingLinks}
\title{Generate an instance of the PISystemLandingLinks PI Web API class}
\usage{
PISystemLandingLinks(self = NULL, cacheInstances = NULL,
  configuration = NULL, userInfo = NULL, versions = NULL, status = NULL)
}
\arguments{
\item{self}{(string)}

\item{cacheInstances}{(string)}

\item{configuration}{(string)}

\item{userInfo}{(string)}

\item{versions}{(string)}

\item{status}{(string)}
}
\value{
PISystemLandingLinks
}
\description{
Generate an instance of the PISystemLandingLinks PI Web API class
}
\examples{
systemLandingLinks <- PISystemLandingLinks(self = "", cacheInstances = "", configuration = "", userInfo = "", versions = "", status = "")

}
