% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITable.r
\name{PITable}
\alias{PITable}
\title{Generate an instance of the PITable PI Web API class}
\usage{
PITable(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, categoryNames = NULL, timeZone = NULL,
  convertToLocalTime = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{categoryNames}{(array)}

\item{timeZone}{(string)}

\item{convertToLocalTime}{(boolean)}

\item{links}{(object)}
}
\value{
PITable
}
\description{
Generate an instance of the PITable PI Web API class
}
\examples{
table <- PITable(webId = "B0DqD5loBNH0erqeqJodtALAgAMa5wsjA0acPM-MJytrPAUkVTVFVOSVRcTUFSS1xUQUJMRVNbVEVTVFRBQkxFXQ", id = "e71a0380-230b-4603-9c3c-cf8c272b6b3c", name = "CarInfo", description = "Table of car info.", path = "\\\\\\\\MyServer\\\\MyDatabase\\\\Tables[CarInfo]", timeZone = "Eastern Standard Time", convertToLocalTime = FALSE)

}
