% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITableCategory.r
\name{PITableCategory}
\alias{PITableCategory}
\title{Generate an instance of the PITableCategory PI Web API class}
\usage{
PITableCategory(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{(object)}
}
\value{
PITableCategory
}
\description{
Generate an instance of the PITableCategory PI Web API class
}
\examples{
tableCategory <- PITableCategory(webId = "a0DqD5loBNH0erqeqJodtALAwgzHiSFSd06HP4lKPqYefQUkVTVFVOSVRcUk9OR1xDQVRFR09SSUVTVEFCTEVbQ0FURUdPUlkxXQ", id = "89c70cc2-5221-4e77-873f-894a3ea61e7d", name = "CategoryName", description = "Relative energy use per ton of process feed.", path = "\\\\\\\\MyServer\\\\Database\\\\CategoriesTable[CategoryName]")

}
