% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITableCategoryLinks.r
\name{PITableCategoryLinks}
\alias{PITableCategoryLinks}
\title{Generate an instance of the PITableCategoryLinks PI Web API class}
\usage{
PITableCategoryLinks(self = NULL, database = NULL, security = NULL,
  securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PITableCategoryLinks
}
\description{
Generate an instance of the PITableCategoryLinks PI Web API class
}
\examples{
tableCategoryLinks <- PITableCategoryLinks(self = "", database = "", security = "", securityEntries = "")

}
