% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piTableData.r
\name{piTableData}
\alias{piTableData}
\title{Generate an instance of the piTableData PI Web API class}
\usage{
piTableData(columns = NULL, rows = NULL)
}
\arguments{
\item{columns}{(object)}

\item{rows}{(array)}
}
\value{
piTableData
}
\description{
Generate an instance of the piTableData PI Web API class
}
