% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITableLinks.r
\name{PITableLinks}
\alias{PITableLinks}
\title{Generate an instance of the PITableLinks PI Web API class}
\usage{
PITableLinks(self = NULL, database = NULL, data = NULL,
  categories = NULL, security = NULL, securityEntries = NULL)
}
\arguments{
\item{self}{(string)}

\item{database}{(string)}

\item{data}{(string)}

\item{categories}{(string)}

\item{security}{(string)}

\item{securityEntries}{(string)}
}
\value{
PITableLinks
}
\description{
Generate an instance of the PITableLinks PI Web API class
}
\examples{
tableLinks <- PITableLinks(self = "", database = "", data = "", categories = "", security = "", securityEntries = "")

}
