% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITimedValues.r
\name{PITimedValues}
\alias{PITimedValues}
\title{Generate an instance of the PITimedValues PI Web API class}
\usage{
PITimedValues(items = NULL, unitsAbbreviation = NULL)
}
\arguments{
\item{items}{(array)}

\item{unitsAbbreviation}{(string)}
}
\value{
PITimedValues
}
\description{
Generate an instance of the PITimedValues PI Web API class
}
\examples{
timedValues <- PITimedValues(unitsAbbreviation = "")

}
