% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITimeRule.r
\name{PITimeRule}
\alias{PITimeRule}
\title{Generate an instance of the PITimeRule PI Web API class}
\usage{
PITimeRule(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, configString = NULL, configStringStored = NULL,
  displayString = NULL, editorType = NULL, isConfigured = NULL,
  isInitializing = NULL, mergeDuplicatedItems = NULL, plugInName = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{configString}{(string)}

\item{configStringStored}{(string)}

\item{displayString}{(string)}

\item{editorType}{(string)}

\item{isConfigured}{(boolean)}

\item{isInitializing}{(boolean)}

\item{mergeDuplicatedItems}{(boolean)}

\item{plugInName}{(string)}

\item{links}{(object)}
}
\value{
PITimeRule
}
\description{
Generate an instance of the PITimeRule PI Web API class
}
\examples{
timeRule <- PITimeRule(webId = "w0xDqD5loBNH0erqeqJodtALAPLXQTRgv5hGA2AAVXYRDBAPLXQTRgv5hGA2AAVXYRDBAUkVTVFVOSVRcMEMzQjczQkYtMzc2NS00QkYyLTg3N0EtRjlENTE4MDFCNkQ0XEM5NzZDOTFDLUEyOEEtNENDMS05MTQ4LUFBRTNFRTM0RERFM1xBTkFMWVNFU1syRDNCODI1NS02OTZGLTQ1MUQtQjUwRC05REJBOEQ3RDFGODBdXFRJTUVSVUxF", id = "4dd0b53c-2f18-11e6-80d8-00155d844304", name = "Periodic", description = "Creates regular periodic time periods.", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement\\\\Analyses[MyAnalysis]\\\\TimeRule", configString = "Frequency=300", configStringStored = "", displayString = "Frequency=300", editorType = "OSIsoft.AF.Time.TimeRule.PeriodicConfig", isConfigured = FALSE, isInitializing = FALSE, mergeDuplicatedItems = FALSE, plugInName = "Periodic")

}
