% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITimeRuleLinks.r
\name{PITimeRuleLinks}
\alias{PITimeRuleLinks}
\title{Generate an instance of the PITimeRuleLinks PI Web API class}
\usage{
PITimeRuleLinks(self = NULL, analysis = NULL, analysisTemplate = NULL,
  plugIn = NULL)
}
\arguments{
\item{self}{(string)}

\item{analysis}{(string)}

\item{analysisTemplate}{(string)}

\item{plugIn}{(string)}
}
\value{
PITimeRuleLinks
}
\description{
Generate an instance of the PITimeRuleLinks PI Web API class
}
\examples{
timeRuleLinks <- PITimeRuleLinks(self = "", analysis = "", analysisTemplate = "", plugIn = "")

}
