% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITimeRulePlugIn.r
\name{PITimeRulePlugIn}
\alias{PITimeRulePlugIn}
\title{Generate an instance of the PITimeRulePlugIn PI Web API class}
\usage{
PITimeRulePlugIn(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, assemblyFileName = NULL,
  assemblyID = NULL, assemblyLoadProperties = NULL, assemblyTime = NULL,
  compatibilityVersion = NULL, isBrowsable = NULL,
  isNonEditableConfig = NULL, loadedAssemblyTime = NULL,
  loadedVersion = NULL, version = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{assemblyFileName}{(string)}

\item{assemblyID}{(string)}

\item{assemblyLoadProperties}{(array)}

\item{assemblyTime}{(string)}

\item{compatibilityVersion}{(integer)}

\item{isBrowsable}{(boolean)}

\item{isNonEditableConfig}{(boolean)}

\item{loadedAssemblyTime}{(string)}

\item{loadedVersion}{(string)}

\item{version}{(string)}

\item{links}{(object)}
}
\value{
PITimeRulePlugIn
}
\description{
Generate an instance of the PITimeRulePlugIn PI Web API class
}
\examples{
timeRulePlugIn <- PITimeRulePlugIn(webId = "W0NxzXSxtlKkGzAhZfHOB-KAjzvCrVcGmUePCcTgk3_HzAUEhMQUZTMDRcUExVR0lOU1RJTUVSVUxFW1BFUklPRElDXQ", id = "adc23b8f-0657-4799-8f09-c4e0937fc7cc", name = "Periodic", description = "Creates regular periodic time periods.", path = "\\\\\\\\MYSERVER\\\\PlugInsTimeRule[Periodic]", assemblyFileName = "OSIsoft.AF.Time.TimeRule.Periodic.dll", assemblyID = "7a635b27-3c3c-4a5f-bba9-01f06e9e525b", assemblyTime = "5/25/2016 2:01:07 PM", compatibilityVersion = 2, isBrowsable = FALSE, isNonEditableConfig = FALSE, loadedAssemblyTime = "1/1/1970 12:00:00 AM", loadedVersion = "", version = "2.8.5.7552")

}
