% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITimeRulePlugInLinks.r
\name{PITimeRulePlugInLinks}
\alias{PITimeRulePlugInLinks}
\title{Generate an instance of the PITimeRulePlugInLinks PI Web API class}
\usage{
PITimeRulePlugInLinks(self = NULL, assetServer = NULL)
}
\arguments{
\item{self}{(string)}

\item{assetServer}{(string)}
}
\value{
PITimeRulePlugInLinks
}
\description{
Generate an instance of the PITimeRulePlugInLinks PI Web API class
}
\examples{
timeRulePlugInLinks <- PITimeRulePlugInLinks(self = "", assetServer = "")

}
