% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUnit.r
\name{PIUnit}
\alias{PIUnit}
\title{Generate an instance of the PIUnit PI Web API class}
\usage{
PIUnit(webId = NULL, id = NULL, name = NULL, abbreviation = NULL,
  description = NULL, path = NULL, factor = NULL, offset = NULL,
  referenceFactor = NULL, referenceOffset = NULL,
  referenceUnitAbbreviation = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{abbreviation}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{factor}{(number)}

\item{offset}{(number)}

\item{referenceFactor}{(number)}

\item{referenceOffset}{(number)}

\item{referenceUnitAbbreviation}{(string)}

\item{links}{(object)}
}
\value{
PIUnit
}
\description{
Generate an instance of the PIUnit PI Web API class
}
\examples{
unit <- PIUnit(webId = "u0DqD5loBNH0eqreqJodtALAR5X5Ku22eUqbeawQJoWIdAUkVVTFVOSVRcVU9NFREUQUJBU0VcZEVSVFo", id = "609049B3-108A-4A9A-94B3-359AA092C5A0", name = "Hertz", abbreviation = "Hz", description = "Hertz Unit", path = "\\\\\\\\MyServer\\\\UOMDatabase\\\\Hertz", referenceUnitAbbreviation = "")

}
