% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUnitClass.r
\name{PIUnitClass}
\alias{PIUnitClass}
\title{Generate an instance of the PIUnitClass PI Web API class}
\usage{
PIUnitClass(webId = NULL, id = NULL, name = NULL, description = NULL,
  canonicalUnitName = NULL, canonicalUnitAbbreviation = NULL, path = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{canonicalUnitName}{(string)}

\item{canonicalUnitAbbreviation}{(string)}

\item{path}{(string)}

\item{links}{(object)}
}
\value{
PIUnitClass
}
\description{
Generate an instance of the PIUnitClass PI Web API class
}
\examples{
unitClass <- PIUnitClass(webId = "K0DqD5loBNH0erqeqJodtALAEYKC2hx4DU2jygG2_AT33gVUkTVFVOSVRcVU9NREFUQUJUB0VcVU9NQ0xBU1NFU1tRVFUOVElUWV2", id = "09FBCC8A-796D-4568-9287-C8B7BC5EA129", name = "Power", description = "Power Unit Class", canonicalUnitName = "watt", canonicalUnitAbbreviation = "W", path = "\\\\\\\\MyServer\\\\UOMDatabase\\\\UOMClasses[Power]")

}
