% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUnitClassLinks.r
\name{PIUnitClassLinks}
\alias{PIUnitClassLinks}
\title{Generate an instance of the PIUnitClassLinks PI Web API class}
\usage{
PIUnitClassLinks(self = NULL, canonicalUnit = NULL, units = NULL,
  assetServer = NULL)
}
\arguments{
\item{self}{(string)}

\item{canonicalUnit}{(string)}

\item{units}{(string)}

\item{assetServer}{(string)}
}
\value{
PIUnitClassLinks
}
\description{
Generate an instance of the PIUnitClassLinks PI Web API class
}
\examples{
unitClassLinks <- PIUnitClassLinks(self = "", canonicalUnit = "", units = "", assetServer = "")

}
