% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUnitLinks.r
\name{PIUnitLinks}
\alias{PIUnitLinks}
\title{Generate an instance of the PIUnitLinks PI Web API class}
\usage{
PIUnitLinks(self = NULL, class = NULL, referenceUnit = NULL)
}
\arguments{
\item{self}{(string)}

\item{class}{(string)}

\item{referenceUnit}{(string)}
}
\value{
PIUnitLinks
}
\description{
Generate an instance of the PIUnitLinks PI Web API class
}
\examples{
unitLinks <- PIUnitLinks(self = "", class = "", referenceUnit = "")

}
