% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piUserInfo.r
\name{piUserInfo}
\alias{piUserInfo}
\title{Generate an instance of the piUserInfo PI Web API class}
\usage{
piUserInfo(identityType = NULL, name = NULL, isAuthenticated = NULL,
  sID = NULL, impersonationLevel = NULL)
}
\arguments{
\item{identityType}{(string)}

\item{name}{(string)}

\item{isAuthenticated}{(boolean)}

\item{sID}{(string)}

\item{impersonationLevel}{(string)}
}
\value{
userInfo
}
\description{
Generate an instance of the piUserInfo PI Web API class
}
\examples{
userInfo(identityType = "WindowsIdentity", name = "MyDomain\\\\Username", isAuthenticated = FALSE, sID = "S-1-5-21-2083424470-1893017325-1469799231-71366", impersonationLevel = "None")

}
