% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piValue.r
\name{piValue}
\alias{piValue}
\title{Generate an instance of the piValue PI Web API class}
\usage{
piValue(value = NULL, exception = NULL)
}
\arguments{
\item{value}{(object)}

\item{exception}{()}
}
\value{
piValue
}
\description{
Generate an instance of the piValue PI Web API class
}
