% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIValueQuery.r
\name{PIValueQuery}
\alias{PIValueQuery}
\title{Generate an instance of the PIValueQuery PI Web API class}
\usage{
PIValueQuery(attributeName = NULL, attributeUOM = NULL,
  attributeValue = NULL, searchOperator = NULL, webException = NULL)
}
\arguments{
\item{attributeName}{(string)}

\item{attributeUOM}{(string)}

\item{attributeValue}{(object)}

\item{searchOperator}{(string)}

\item{webException}{()}
}
\value{
PIValueQuery
}
\description{
Generate an instance of the PIValueQuery PI Web API class
}
\examples{
valueQuery <- PIValueQuery(attributeName = "Pressure", attributeUOM = "pound-force per square inch", searchOperator = "LessThan")

}
