% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piVersion.r
\name{piVersion}
\alias{piVersion}
\title{Generate an instance of the piVersion PI Web API class}
\usage{
piVersion(fullVersion = NULL, majorMinorRevision = NULL, build = NULL)
}
\arguments{
\item{fullVersion}{(string)}

\item{majorMinorRevision}{(string)}

\item{build}{(string)}
}
\value{
version
}
\description{
Generate an instance of the piVersion PI Web API class
}
\examples{
version(fullVersion = "1.8.0.201", majorMinorRevision = "1.8.0", build = "201")

}
