% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piwebapi.r
\name{piwebapi}
\alias{piwebapi}
\title{PI Web  API Top level object
Generate an instance of the top level object of this library}
\description{
PI Web  API Top level object
Generate an instance of the top level object of this library
}
\section{Fields}{

\describe{
\item{\code{serviceBase}}{PI Web API endpoint}

\item{\code{username}}{PI Web API User name}

\item{\code{password}}{PI Web API Password}

\item{\code{useKerberos}}{Kerberos Authentication method selected}

\item{\code{debug}}{Show HTTP requests debug}
}}

