% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIWebException.r
\name{PIWebException}
\alias{PIWebException}
\title{Generate an instance of the PIWebException PI Web API class}
\usage{
PIWebException(statusCode = NULL, errors = NULL)
}
\arguments{
\item{statusCode}{(integer)}

\item{errors}{(array)}
}
\value{
PIWebException
}
\description{
Generate an instance of the PIWebException PI Web API class
}
\examples{
webException <- PIWebException(statusCode = 500)

}
