% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointApi.r
\name{point$getAttributeByName}
\alias{point$getAttributeByName}
\title{Get a point attribute by name.}
\arguments{
\item{name}{The name of the attribute.}

\item{webId}{The ID of the point.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A point attribute specified by its name.
}
\description{
Get a point attribute by name.
}
