% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointApi.r
\name{point$getByPath}
\alias{point$getByPath}
\title{Get a point by path.}
\arguments{
\item{path}{The path to the point.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{column}{the value of the user label column}
}
\value{
The specified point.
}
\description{
Get a point by path.
}
