% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityIdentityApi.r
\name{securityIdentity$delete}
\alias{securityIdentity$delete}
\title{Delete a security identity.}
\arguments{
\item{webId}{The ID of the security identity.}

\item{column}{the value of the user label column}
}
\value{
The security identity was deleted.
}
\description{
Delete a security identity.
}
