% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityIdentityApi.r
\name{securityIdentity$get}
\alias{securityIdentity$get}
\title{Retrieve a security identity.}
\arguments{
\item{webId}{The ID of the security identity.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified security identity.
}
\description{
Retrieve a security identity.
}
