% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityIdentityApi.r
\name{securityIdentity$getByPath}
\alias{securityIdentity$getByPath}
\title{Retrieve a security identity by path.}
\arguments{
\item{path}{The path to the security identity.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified security identity.
}
\description{
Retrieve a security identity by path.
}
