% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityIdentityApi.r
\name{securityIdentity$getSecurityEntries}
\alias{securityIdentity$getSecurityEntries}
\title{Retrieve the security entries associated with the security identity based on the specified criteria. By default, all security entries for this security identity are returned.}
\arguments{
\item{webId}{The ID of the security identity.}

\item{nameFilter}{The name query string used for filtering security entries. The default is no filter.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of security entries matching the specified condition.
}
\description{
Retrieve the security entries associated with the security identity based on the specified criteria. By default, all security entries for this security identity are returned.
}
