% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityIdentityApi.r
\name{securityIdentity$update}
\alias{securityIdentity$update}
\title{Update a security identity by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the security identity.}

\item{securityIdentity}{A partial security identity containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The security identity was updated.
}
\description{
Update a security identity by replacing items in its definition.
}
