% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityMappingApi.r
\name{securityMapping$delete}
\alias{securityMapping$delete}
\title{Delete a security mapping.}
\arguments{
\item{webId}{The ID of the security mapping.}

\item{column}{the value of the user label column}
}
\value{
The security mapping was deleted.
}
\description{
Delete a security mapping.
}
