% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityMappingApi.r
\name{securityMapping$getByPath}
\alias{securityMapping$getByPath}
\title{Retrieve a security mapping by path.}
\arguments{
\item{path}{The path to the security mapping.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified security mapping.
}
\description{
Retrieve a security mapping by path.
}
