% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/securityMappingApi.r
\name{securityMapping$update}
\alias{securityMapping$update}
\title{Update a security mapping by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the security mapping.}

\item{securityMapping}{A partial security mapping containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The security mapping was updated.
}
\description{
Update a security mapping by replacing items in its definition.
}
