% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$getChannel}
\alias{stream$getChannel}
\title{Opens a channel that will send messages about any value changes for the specified stream.}
\arguments{
\item{webId}{The ID of the stream.}

\item{includeInitialValues}{Specified if the channel should send a message with the current value of the stream after the connection is opened. The default is 'false'.}
}
\value{
Switches to the Web Socket protocol.
}
\description{
Opens a channel that will send messages about any value changes for the specified stream.
}
