% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$getEnd}
\alias{stream$getEnd}
\title{Returns the end-of-stream value of the stream.}
\arguments{
\item{webId}{The ID of the stream.}

\item{desiredUnits}{The name or abbreviation of the desired units of measure for the returned value, as found in the UOM database associated with the attribute. If not specified for an attribute, the attribute's default unit of measure is used. If the underlying stream is a point, this value may not be specified, as points are not associated with a unit of measure.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The value at the specified time value.
}
\description{
Returns the end-of-stream value of the stream.
}
