% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$getInterpolatedAtTimes}
\alias{stream$getInterpolatedAtTimes}
\title{Retrieves interpolated values over the specified time range at the specified sampling interval.}
\arguments{
\item{webId}{The ID of the stream.}

\item{desiredUnits}{The name or abbreviation of the desired units of measure for the returned value, as found in the UOM database associated with the attribute. If not specified for an attribute, the attribute's default unit of measure is used. If the underlying stream is a point, this value may not be specified, as points are not associated with a unit of measure.}

\item{filterExpression}{An optional string containing a filter expression. Expression variables are relative to the data point. Use '.' to reference the containing attribute. If the attribute does not support filtering, the filter will be ignored. The default is no filtering.}

\item{includeFilteredValues}{Specify 'true' to indicate that values which fail the filter criteria are present in the returned data at the times where they occurred with a value set to a 'Filtered' enumeration value with bad status. Repeated consecutive failures are omitted.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{time}{The timestamp at which to retrieve an interpolated value. Multiple timestamps may be specified with multiple instances of the parameter.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}
}
\value{
The values that meet the specified conditions.
}
\description{
Retrieves interpolated values over the specified time range at the specified sampling interval.
}
