% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$getRecordedAtTimes}
\alias{stream$getRecordedAtTimes}
\title{Retrieves recorded values at the specified times.}
\arguments{
\item{webId}{The ID of the stream.}

\item{associations}{Associated values to return in the response, separated by semicolons (;). This call supports Annotations to return events with annotation values. If this parameter is not specified, annotation values are not returned.}

\item{desiredUnits}{The name or abbreviation of the desired units of measure for the returned value, as found in the UOM database associated with the attribute. If not specified for an attribute, the attribute's default unit of measure is used. If the underlying stream is a point, this value may not be specified, as points are not associated with a unit of measure.}

\item{retrievalMode}{An optional value that determines the value to return when a value doesn't exist at the exact time specified. The default is 'Auto'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{time}{The timestamp at which to retrieve a recorded value. Multiple timestamps may be specified with multiple instances of the parameter.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}
}
\value{
The values that meet the specified conditions.
}
\description{
Retrieves recorded values at the specified times.
}
