% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$getSummary}
\alias{stream$getSummary}
\title{Returns a summary over the specified time range for the stream.}
\arguments{
\item{webId}{The ID of the stream.}

\item{calculationBasis}{Specifies the method of evaluating the data over the time range. The default is 'TimeWeighted'.}

\item{endTime}{An optional end time. The default is '*' for element attributes and points. For event frame attributes, the default is the event frame's end time, or '*' if that is not set. Note that if endTime is earlier than startTime, the resulting values will be in time-descending order.}

\item{filterExpression}{A string containing a filter expression. Expression variables are relative to the attribute. Use '.' to reference the containing attribute.}

\item{sampleInterval}{When the sampleType is Interval, sampleInterval specifies how often the filter expression is evaluated when computing the summary for an interval.}

\item{sampleType}{Defines the evaluation of an expression over a time range. The default is 'ExpressionRecordedValues'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{startTime}{An optional start time. The default is '*-1d' for element attributes and points. For event frame attributes, the default is the event frame's start time, or '*-1d' if that is not set.}

\item{summaryDuration}{The duration of each summary interval. If specified in hours, minutes, seconds, or milliseconds, the summary durations will be evenly spaced UTC time intervals. Longer interval types are interpreted using wall clock rules and are time zone dependent.}

\item{summaryType}{Specifies the kinds of summaries to produce over the range. The default is 'Total'. Multiple summary types may be specified by using multiple instances of summaryType.}

\item{timeType}{Specifies how to calculate the timestamp for each interval. The default is 'Auto'.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}
}
\value{
The summaries that meet the specified conditions.
}
\description{
Returns a summary over the specified time range for the stream.
}
