% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$getValue}
\alias{stream$getValue}
\title{Returns the value of the stream at the specified time. By default, this is usually the current value.}
\arguments{
\item{webId}{The ID of the stream.}

\item{desiredUnits}{The name or abbreviation of the desired units of measure for the returned value, as found in the UOM database associated with the attribute. If not specified for an attribute, the attribute's default unit of measure is used. If the underlying stream is a point, this value may not be specified, as points are not associated with a unit of measure.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{time}{An optional time. The default time context is determined from the owning object - for example, the time range of the event frame or transfer which holds this attribute. Otherwise, the implementation of the Data Reference determines the meaning of no context. For Points or simply configured PI Point Data References, this means the snapshot value of the PI Point on the Data Server.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}
}
\value{
The value at the specified time value.
}
\description{
Returns the value of the stream at the specified time. By default, this is usually the current value.
}
