% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$retrieveStreamUpdate}
\alias{stream$retrieveStreamUpdate}
\title{Receive stream updates}
\arguments{
\item{marker}{Identifier of stream source and current position}

\item{desiredUnits}{The name or abbreviation of the desired units of measure for the returned value, as found in the UOM database associated with the attribute. If not specified for an attribute, the attribute's default unit of measure is used. If the underlying stream is a point, this value may not be specified, as points are not associated with a unit of measure.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The stream updates have been successfully retrieved
}
\description{
Receive stream updates
}
