% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$updateValue}
\alias{stream$updateValue}
\title{Updates a value for the specified stream.}
\arguments{
\item{webId}{The ID of the stream.}

\item{PITimedValue}{The value to add or update.}

\item{bufferOption}{The desired AFBufferOption. The default is 'BufferIfPossible'.}

\item{updateOption}{The desired AFUpdateOption. The default is 'Replace'. This parameter is ignored if the attribute is a configuration item.}
}
\value{
The value was accepted for a potential buffered update. The response's Location header is a link that can be used to examine the result.
}
\description{
Updates a value for the specified stream.
}
