% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamApi.r
\name{stream$updateValues}
\alias{stream$updateValues}
\title{Updates multiple values for the specified stream.}
\arguments{
\item{webId}{The ID of the stream.}

\item{values}{The values to add or update.}

\item{bufferOption}{The desired AFBufferOption. The default is 'BufferIfPossible'.}

\item{updateOption}{The desired AFUpdateOption. The default is 'Replace'.}
}
\value{
All updates executed successfully.
}
\description{
Updates multiple values for the specified stream.
}
