% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$getChannelAdHoc}
\alias{streamSet$getChannelAdHoc}
\title{Opens a channel that will send messages about any value changes for the specified streams.}
\arguments{
\item{webId}{The ID of a stream.  Multiple streams may be specified with multiple instances of the parameter.}

\item{includeInitialValues}{Specified if the channel should send a message with the current values of all the streams after the connection is opened. The default is 'false'.}
}
\value{
Switches to the Web Socket protocol.
}
\description{
Opens a channel that will send messages about any value changes for the specified streams.
}
