% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$getEndAdHoc}
\alias{streamSet$getEndAdHoc}
\title{Returns End Of Stream values for attributes of the specified streams}
\arguments{
\item{webId}{The ID of a stream.  Multiple streams may be specified with multiple instances of the parameter.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
End of stream values of the streams that meet the specified conditions.
}
\description{
Returns End Of Stream values for attributes of the specified streams
}
