% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$getInterpolated}
\alias{streamSet$getInterpolated}
\title{Returns interpolated values of attributes for an element, event frame or attribute over the specified time range at the specified sampling interval.}
\arguments{
\item{webId}{The ID of an element, event frame or attribute, which is the base element or parent of all the stream attributes.}

\item{categoryName}{Specify that included attributes must have this category. The default is no category filter.}

\item{endTime}{An optional end time. The default is '*' for element attributes and points. For event frame attributes, the default is the event frame's end time, or '*' if that is not set. Note that if endTime is earlier than startTime, the resulting values will be in time-descending order.}

\item{filterExpression}{An optional string containing a filter expression. Expression variables are relative to the data point. Use '.' to reference the containing attribute. If the attribute does not support filtering, the filter will be ignored. The default is no filtering.}

\item{includeFilteredValues}{Specify 'true' to indicate that values which fail the filter criteria are present in the returned data at the times where they occurred with a value set to a 'Filtered' enumeration value with bad status. Repeated consecutive failures are omitted.}

\item{interval}{The sampling interval, in AFTimeSpan format.}

\item{nameFilter}{The name query string used for filtering attributes. The default is no filter.}

\item{searchFullHierarchy}{Specifies if the search should include attributes nested further than the immediate attributes of the searchRoot. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{showExcluded}{Specified if the search should include attributes with the Excluded property set. The default is 'false'.}

\item{showHidden}{Specified if the search should include attributes with the Hidden property set. The default is 'false'.}

\item{sortField}{The field or property of the object used to sort the returned collection. For better performance, by default no sorting is applied. 'Name' is the only supported field by which to sort.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{startTime}{An optional start time. The default is '*-1d' for element attributes and points. For event frame attributes, the default is the event frame's start time, or '*-1d' if that is not set.}

\item{syncTime}{An optional start time anchor, in AFTime format. When specified, interpolated data retrieval will use the sync time as the origin for calculating the interval times.}

\item{syncTimeBoundaryType}{An optional string specifying the boundary type to use when applying a syncTime. The allowed values are 'Inside' and 'Outside'. The default is 'Inside'.}

\item{templateName}{Specify that included attributes must be members of this template. The default is no template filter.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Interpolated values of the streams that meet the specified conditions.
}
\description{
Returns interpolated values of attributes for an element, event frame or attribute over the specified time range at the specified sampling interval.
}
