% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$getInterpolatedAtTimes}
\alias{streamSet$getInterpolatedAtTimes}
\title{Returns interpolated values of attributes for an element, event frame or attribute at the specified times.}
\arguments{
\item{webId}{The ID of an element, event frame or attribute, which is the base element or parent of all the stream attributes.}

\item{time}{The timestamp at which to retrieve a interpolated value. Multiple timestamps may be specified with multiple instances of the parameter.}

\item{categoryName}{Specify that included attributes must have this category. The default is no category filter.}

\item{filterExpression}{An optional string containing a filter expression. Expression variables are relative to the data point. Use '.' to reference the containing attribute. If the attribute does not support filtering, the filter will be ignored. The default is no filtering.}

\item{includeFilteredValues}{Specify 'true' to indicate that values which fail the filter criteria are present in the returned data at the times where they occurred with a value set to a 'Filtered' enumeration value with bad status. Repeated consecutive failures are omitted.}

\item{nameFilter}{The name query string used for filtering attributes. The default is no filter.}

\item{searchFullHierarchy}{Specifies if the search should include attributes nested further than the immediate attributes of the searchRoot. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{showExcluded}{Specified if the search should include attributes with the Excluded property set. The default is 'false'.}

\item{showHidden}{Specified if the search should include attributes with the Hidden property set. The default is 'false'.}

\item{sortOrder}{The order that the returned collection is sorted. The default is 'Ascending'.}

\item{templateName}{Specify that included attributes must be members of this template. The default is no template filter.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Interpolated values of the streams that meet the specified conditions.
}
\description{
Returns interpolated values of attributes for an element, event frame or attribute at the specified times.
}
