% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$getSummaries}
\alias{streamSet$getSummaries}
\title{Returns summary values of the attributes for an element, event frame or attribute.}
\arguments{
\item{webId}{The ID of an element, event frame or attribute, which is the base element or parent of all the stream attributes.}

\item{calculationBasis}{Specifies the method of evaluating the data over the time range. The default is 'TimeWeighted'.}

\item{categoryName}{Specify that included attributes must have this category. The default is no category filter.}

\item{endTime}{An optional end time. The default is '*' for element attributes and points. For event frame attributes, the default is the event frame's end time, or '*' if that is not set. Note that if endTime is earlier than startTime, the resulting values will be in time-descending order.}

\item{filterExpression}{A string containing a filter expression. Expression variables are relative to the attribute. Use '.' to reference the containing attribute. The default is no filtering.}

\item{nameFilter}{The name query string used for filtering attributes. The default is no filter.}

\item{sampleInterval}{A time span specifies how often the filter expression is evaluated when computing the summary for an interval, if the sampleType is 'Interval'.}

\item{sampleType}{A flag which specifies one or more summaries to compute for each interval over the time range. The default is 'ExpressionRecordedValues'.}

\item{searchFullHierarchy}{Specifies if the search should include attributes nested further than the immediate attributes of the searchRoot. The default is 'false'.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{showExcluded}{Specified if the search should include attributes with the Excluded property set. The default is 'false'.}

\item{showHidden}{Specified if the search should include attributes with the Hidden property set. The default is 'false'.}

\item{startTime}{An optional start time. The default is '*-1d' for element attributes and points. For event frame attributes, the default is the event frame's start time, or '*-1d' if that is not set.}

\item{summaryDuration}{The duration of each summary interval.}

\item{summaryType}{Specifies the kinds of summaries to produce over the range. The default is 'Total'. Multiple summary types may be specified by using multiple instances of summaryType.}

\item{templateName}{Specify that included attributes must be members of this template. The default is no template filter.}

\item{timeType}{Specifies how to calculate the timestamp for each interval. The default is 'Auto'.}

\item{timeZone}{The time zone in which the time string will be interpreted. This parameter will be ignored if a time zone is specified in the time string. If no time zone is specified in either places, the PI Web API server time zone will be used.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Summary values of the streams that meet the specified conditions.
}
\description{
Returns summary values of the attributes for an element, event frame or attribute.
}
