% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$registerStreamSetUpdates}
\alias{streamSet$registerStreamSetUpdates}
\title{Register for stream updates}
\arguments{
\item{webId}{The ID of a stream. Multiple streams may be specified with multiple instances of the parameter.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Successful Registration
}
\description{
Register for stream updates
}
