% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$updateValue}
\alias{streamSet$updateValue}
\title{Updates a single value for the specified streams.}
\arguments{
\item{webId}{The ID of the parent element, event frame, or attribute. Attributes specified in the body must be descendants of the specified object.}

\item{values}{The values to add or update.}

\item{bufferOption}{The desired AFBufferOption. The default is 'BufferIfPossible'.}

\item{updateOption}{The desired AFUpdateOption. The default is 'Replace'.}
}
\value{
All updates executed successfully.
}
\description{
Updates a single value for the specified streams.
}
