% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemApi.r
\name{system$cacheInstances}
\alias{system$cacheInstances}
\title{Get AF cache instances currently in use by the system. These are caches from which user requests are serviced. The number of instances depends on the number of users connected to the service, the service's authentication method, and the cache instance configuration.}
\value{
Cache instance information.
}
\description{
Get AF cache instances currently in use by the system. These are caches from which user requests are serviced. The number of instances depends on the number of users connected to the service, the service's authentication method, and the cache instance configuration.
}
