% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemApi.r
\name{system$status}
\alias{system$status}
\title{Get the system uptime, the system state and the number of cache instances for this PI System Web API instance.}
\value{
System Status.
}
\description{
Get the system uptime, the system state and the number of cache instances for this PI System Web API instance.
}
