% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemApi.r
\name{system$userInfo}
\alias{system$userInfo}
\title{Get information about the Windows identity used to fulfill the request. This depends on the service's authentication method and the credentials passed by the client. The impersonation level of the Windows identity is included.}
\value{
Information about the current user.
}
\description{
Get information about the Windows identity used to fulfill the request. This depends on the service's authentication method and the credentials passed by the client. The impersonation level of the Windows identity is included.
}
