% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemApi.r
\name{system$versions}
\alias{system$versions}
\title{Get the current versions of the PI Web API instance and all external plugins.}
\arguments{
\item{column}{the value of the user label column}
}
\value{
A dictionary of assembly name-version information pairs.
}
\description{
Get the current versions of the PI Web API instance and all external plugins.
}
