% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$delete}
\alias{table$delete}
\title{Delete a table.}
\arguments{
\item{webId}{The ID of the table to delete.}

\item{column}{the value of the user label column}
}
\value{
The table was deleted.
}
\description{
Delete a table.
}
