% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$deleteSecurityEntry}
\alias{table$deleteSecurityEntry}
\title{Delete a security entry owned by the table.}
\arguments{
\item{name}{The name of the security entry. For every backslash character (\\) in the security entry name, replace with asterisk (*). As an example, use domain*username instead of domain\\username.}

\item{webId}{The ID of the table where the security entry will be deleted.}

\item{applyToChildren}{If false, the new access permissions are only applied to the associated object. If true, the access permissions of children with any parent-child reference types will change when the permissions on the primary parent change.}
}
\value{
The security entry was deleted.
}
\description{
Delete a security entry owned by the table.
}
