% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$getCategories}
\alias{table$getCategories}
\title{Get a table's categories.}
\arguments{
\item{webId}{The ID of the table.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
A list of categories.
}
\description{
Get a table's categories.
}
