% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$getData}
\alias{table$getData}
\title{Get the table's data.}
\arguments{
\item{webId}{The ID of the table.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The data for the specified table.
}
\description{
Get the table's data.
}
