% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$updateData}
\alias{table$updateData}
\title{Update the table's data.}
\arguments{
\item{webId}{The ID of the table on which to update the data.}

\item{data}{The new table data definition.}

\item{column}{the value of the user label column}
}
\value{
The table data was updated.
}
\description{
Update the table's data.
}
