% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCategoryApi.r
\name{tableCategory$delete}
\alias{tableCategory$delete}
\title{Delete a table category.}
\arguments{
\item{webId}{The ID of the table category to delete.}
}
\value{
The table category was deleted.
}
\description{
Delete a table category.
}
