% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCategoryApi.r
\name{tableCategory$get}
\alias{tableCategory$get}
\title{Retrieve a table category.}
\arguments{
\item{webId}{The id of the table category.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified table category.
}
\description{
Retrieve a table category.
}
