% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCategoryApi.r
\name{tableCategory$getByPath}
\alias{tableCategory$getByPath}
\title{Retrieve a table category by path.}
\arguments{
\item{path}{The path to the target table category.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified table category.
}
\description{
Retrieve a table category by path.
}
