% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableCategoryApi.r
\name{tableCategory$update}
\alias{tableCategory$update}
\title{Update a table category by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the table category to update.}

\item{PITableCategory}{A partial table category containing the desired changes.}
}
\value{
The table category was updated.
}
\description{
Update a table category by replacing items in its definition.
}
