% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRuleApi.r
\name{timeRule$delete}
\alias{timeRule$delete}
\title{Delete a Time Rule.}
\arguments{
\item{webId}{The ID of the Time Rule.}

\item{column}{the value of the user label column}
}
\value{
The Time Rule was deleted.
}
\description{
Delete a Time Rule.
}
