% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRuleApi.r
\name{timeRule$getByPath}
\alias{timeRule$getByPath}
\title{Retrieve a Time Rule by path.}
\arguments{
\item{path}{The path to the Time Rule.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The requested Time Rule.
}
\description{
Retrieve a Time Rule by path.
}
