% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRuleApi.r
\name{timeRule$update}
\alias{timeRule$update}
\title{Update a Time Rule by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the Time Rule.}

\item{PITimeRule}{A partial Time Rule containing the desired changes.}
}
\value{
The Time Rule was updated.
}
\description{
Update a Time Rule by replacing items in its definition.
}
