% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeRulePlugInApi.r
\name{timeRulePlugIn$get}
\alias{timeRulePlugIn$get}
\title{Retrieve a Time Rule Plug-in.}
\arguments{
\item{webId}{The ID of the Time Rule Plug-in.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified Time Rule Plug-in.
}
\description{
Retrieve a Time Rule Plug-in.
}
