% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitApi.r
\name{unit$delete}
\alias{unit$delete}
\title{Delete a unit.}
\arguments{
\item{webId}{The ID of the unit.}

\item{column}{the value of the user label column}
}
\value{
The unit was deleted.
}
\description{
Delete a unit.
}
