% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitApi.r
\name{unit$get}
\alias{unit$get}
\title{Retrieve a unit.}
\arguments{
\item{webId}{The ID of the unit.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified unit.
}
\description{
Retrieve a unit.
}
