% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitApi.r
\name{unit$update}
\alias{unit$update}
\title{Update a unit.}
\arguments{
\item{webId}{The ID of the unit.}

\item{unitDTO}{A partial unit containing the desired changes.}

\item{column}{the value of the user label column}
}
\value{
The unit was updated.
}
\description{
Update a unit.
}
