% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$createUnit}
\alias{unitClass$createUnit}
\title{Create a unit in the specified Unit Class.}
\arguments{
\item{webId}{The ID of the server.}

\item{PIUnit}{The new unit definition.}
}
\value{
The unit was created. The response's Location header is a link to the unit.
}
\description{
Create a unit in the specified Unit Class.
}
