% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$delete}
\alias{unitClass$delete}
\title{Delete a unit class.}
\arguments{
\item{webId}{The ID of the unit class.}
}
\value{
The unit class was deleted.
}
\description{
Delete a unit class.
}
