% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$get}
\alias{unitClass$get}
\title{Retrieve a unit class.}
\arguments{
\item{webId}{The ID of the unit class.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The specified unit class.
}
\description{
Retrieve a unit class.
}
