% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$getByPath}
\alias{unitClass$getByPath}
\title{Retrieve a unit class by path.}
\arguments{
\item{path}{The path to the unit class.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The specified unit class.
}
\description{
Retrieve a unit class by path.
}
