% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$getUnits}
\alias{unitClass$getUnits}
\title{Get a list of all units belonging to the unit class.}
\arguments{
\item{webId}{The ID of unit class.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}
}
\value{
The list of units in the unit class.
}
\description{
Get a list of all units belonging to the unit class.
}
