% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisApi.r
\name{analysis$getCategories}
\alias{analysis$getCategories}
\title{Get an Analysis' categories.}
\arguments{
\item{webId}{The ID of the Analysis.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of categories.
}
\description{
Get an Analysis' categories.
}
