% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysisRulePlugInApi.r
\name{analysisRulePlugIn$get}
\alias{analysisRulePlugIn$get}
\title{Retrieve an Analysis Rule Plug-in.}
\arguments{
\item{webId}{The ID of the Analysis Rule Plug-in.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The specified Analysis Rule Plug-in.
}
\description{
Retrieve an Analysis Rule Plug-in.
}
