% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$createElement}
\alias{assetDatabase$createElement}
\title{Create a child element.}
\arguments{
\item{webId}{The ID of the asset database on which to create the element.}

\item{element}{The new element definition.}

\item{column}{the value of the user label column}
}
\value{
The element was created. The response's Location header is a link to the element.
}
\description{
Create a child element.
}
