% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$createElementCategory}
\alias{assetDatabase$createElementCategory}
\title{Create an element category at the Asset Database root.}
\arguments{
\item{webId}{The ID of the database in which to create the element category.}

\item{PIElementCategory}{The new element category definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The element category was created. The response's Location header is a link to the element category.
}
\description{
Create an element category at the Asset Database root.
}
