% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$createElementTemplate}
\alias{assetDatabase$createElementTemplate}
\title{Create a template at the Asset Database root. Specify InstanceType of "Element" or "EventFrame" to create element or event frame template respectively. Only these two types of templates can be created.}
\arguments{
\item{webId}{The ID of the database in which to create the element template.}

\item{PIElementTemplate}{The new element template definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The element template was created. The response's Location header is a link to the created template.
}
\description{
Create a template at the Asset Database root. Specify InstanceType of "Element" or "EventFrame" to create element or event frame template respectively. Only these two types of templates can be created.
}
