% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetDatabaseApi.r
\name{assetDatabase$getElementCategories}
\alias{assetDatabase$getElementCategories}
\title{Retrieve element categories for a given Asset Database.}
\arguments{
\item{webId}{The ID of the database.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{column}{the value of the user label column}
}
\value{
The element categories that exist in the datbase.
}
\description{
Retrieve element categories for a given Asset Database.
}
