% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$createAssetDatabase}
\alias{assetServer$createAssetDatabase}
\title{Create an asset database.}
\arguments{
\item{webId}{The ID of the asset server on which to create the database.}

\item{PIAssetDatabase}{The new database definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The database was created. The response's Location header is a link to the database.
}
\description{
Create an asset database.
}
