% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$createSecurityIdentity}
\alias{assetServer$createSecurityIdentity}
\title{Create a security identity.}
\arguments{
\item{webId}{The ID of the asset server on which to create the security identity.}

\item{securityIdentity}{The new security identity definition.}

\item{column}{the value of the user label column}
}
\value{
The security identity was created. The response's Location header is a link to the security identity.
}
\description{
Create a security identity.
}
