% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$createSecurityIdentity}
\alias{assetServer$createSecurityIdentity}
\title{Create a security identity.}
\arguments{
\item{webId}{The ID of the asset server on which to create the security identity.}

\item{PISecurityIdentity}{The new security identity definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The security identity was created. The response's Location header is a link to the security identity.
}
\description{
Create a security identity.
}
