% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$createUnitClass}
\alias{assetServer$createUnitClass}
\title{Create a unit class in the specified Asset Server.}
\arguments{
\item{webId}{The ID of the server.}

\item{unitClass}{The new unit class definition.}

\item{column}{the value of the user label column}
}
\value{
The unit class was created. The response's Location header is a link to the unit class.
}
\description{
Create a unit class in the specified Asset Server.
}
