% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$createUnitClass}
\alias{assetServer$createUnitClass}
\title{Create a unit class in the specified Asset Server.}
\arguments{
\item{webId}{The ID of the server.}

\item{PIUnitClass}{The new unit class definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The unit class was created. The response's Location header is a link to the unit class.
}
\description{
Create a unit class in the specified Asset Server.
}
