% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getSecurityEntries}
\alias{assetServer$getSecurityEntries}
\title{Retrieve the security entries of the specified security item associated with the asset server based on the specified criteria. By default, all security entries for this asset server are returned.}
\arguments{
\item{webId}{The ID of the asset server.}

\item{nameFilter}{The name query string used for filtering security entries. The default is no filter.}

\item{securityItem}{The security item of the desired security entries information to be returned. If the parameter is not specified, security entries of the 'Default' security item will be returned.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of security entries matching the specified condition.
}
\description{
Retrieve the security entries of the specified security item associated with the asset server based on the specified criteria. By default, all security entries for this asset server are returned.
}
