% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assetServerApi.r
\name{assetServer$getTimeRulePlugIns}
\alias{assetServer$getTimeRulePlugIns}
\title{Retrieve a list of all Time Rule Plug-in's.}
\arguments{
\item{webId}{The ID of the asset server, where the Time Rule Plug-in's are installed.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of Time Rule Plug-in's.
}
\description{
Retrieve a list of all Time Rule Plug-in's.
}
