% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeApi.r
\name{attribute$createConfig}
\alias{attribute$createConfig}
\title{Create or update an attribute's DataReference configuration (Create/Update PI point for PI Point DataReference).}
\arguments{
\item{webId}{The ID of the attribute.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The attribute's DataReference configuration was created or updated.
}
\description{
Create or update an attribute's DataReference configuration (Create/Update PI point for PI Point DataReference).
}
