% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeApi.r
\name{attribute$setValue}
\alias{attribute$setValue}
\title{Set the value of a configuration item attribute. For attributes with a data reference or non-configuration item attributes, consult the documentation for streams.}
\arguments{
\item{webId}{The ID of the attribute.}

\item{PITimedValue}{The value to write.}
}
\value{
The value was set successfully.
}
\description{
Set the value of a configuration item attribute. For attributes with a data reference or non-configuration item attributes, consult the documentation for streams.
}
