% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeCategoryApi.r
\name{attributeCategory$update}
\alias{attributeCategory$update}
\title{Update an attribute category by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the attribute category to update.}

\item{PIAttributeCategory}{A partial attribute category containing the desired changes.}
}
\value{
The attribute category was updated.
}
\description{
Update an attribute category by replacing items in its definition.
}
