% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributeTemplateApi.r
\name{attributeTemplate$delete}
\alias{attributeTemplate$delete}
\title{Delete an attribute template.}
\arguments{
\item{webId}{The ID of the attribute template.}
}
\value{
The attribute template was deleted.
}
\description{
Delete an attribute template.
}
