% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$createAnalysis}
\alias{element$createAnalysis}
\title{Create an Analysis.}
\arguments{
\item{webId}{The ID of the element on which to create the Analysis.}

\item{PIAnalysis}{The new Analysis definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The Analysis was created. The response's Location header is a link to the Analysis.
}
\description{
Create an Analysis.
}
