% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementApi.r
\name{element$createElement}
\alias{element$createElement}
\title{Create a child element.}
\arguments{
\item{webId}{The ID of the parent element on which to create the element.}

\item{PIElement}{The new element definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The element was created. The response's Location header is a link to the element.
}
\description{
Create a child element.
}
