% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elementTemplateApi.r
\name{elementTemplate$get}
\alias{elementTemplate$get}
\title{Retrieve an element template.}
\arguments{
\item{webId}{The ID of the element template.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{column}{the value of the user label column}
}
\value{
The specified element template.
}
\description{
Retrieve an element template.
}
