% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationSetApi.r
\name{enumerationSet$createValue}
\alias{enumerationSet$createValue}
\title{Create an enumeration value for a enumeration set.}
\arguments{
\item{webId}{The ID of the enumeration set on which to create the enumeration value.}

\item{PIEnumerationValue}{The new enumeration value definition.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The enumeration value was created. The response's Location header is a link to the enumeration value.
}
\description{
Create an enumeration value for a enumeration set.
}
