% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enumerationValueApi.r
\name{enumerationValue$updateEnumerationValue}
\alias{enumerationValue$updateEnumerationValue}
\title{Update an enumeration value by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the enumeration value to update.}

\item{PIEnumerationValue}{A partial enumeration value containing the desired changes.}
}
\value{
The enumeration set was updated.
}
\description{
Update an enumeration value by replacing items in its definition.
}
