% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$createAnnotation}
\alias{eventFrame$createAnnotation}
\title{Create an annotation on an event frame.}
\arguments{
\item{webId}{The ID of the owner event frame on which to create the annotation.}

\item{annotation}{The new annotation definition.}

\item{column}{the value of the user label column}
}
\value{
The annotation was created. The response's Location header is a link to the annotation.
}
\description{
Create an annotation on an event frame.
}
