% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$createAttribute}
\alias{eventFrame$createAttribute}
\title{Create a new attribute of the specified event frame.}
\arguments{
\item{webId}{The ID of the event frame on which to create the attribute.}

\item{PIAttribute}{The definition of the new attribute.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The attribute was created. The response's Location header is a link to the created attribute.
}
\description{
Create a new attribute of the specified event frame.
}
