% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$getReferencedElements}
\alias{eventFrame$getReferencedElements}
\title{Retrieve the event frame's referenced elements.}
\arguments{
\item{webId}{The ID of the event frame whose referenced elements should be retrieved.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
A list of referenced elements.
}
\description{
Retrieve the event frame's referenced elements.
}
