% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventFrameApi.r
\name{eventFrame$update}
\alias{eventFrame$update}
\title{Update an event frame by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the event frame to update.}

\item{PIEventFrame}{A partial event frame containing the desired changes.}
}
\value{
The event frame was updated.
}
\description{
Update an event frame by replacing items in its definition.
}
