% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisCategory.r
\name{PIAnalysisCategory}
\alias{PIAnalysisCategory}
\title{Generate an instance of the PIAnalysisCategory PI Web API class}
\usage{
PIAnalysisCategory(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAnalysisCategory
}
\description{
Generate an instance of the PIAnalysisCategory PI Web API class
}
\examples{
analysisCategory <- PIAnalysisCategory(webId = "I1XCDqD5loBNH0erqeqJodtALAoko2-UoOVEibhTWQCk1MDw", id = "f9364aa2-0e4a-4854-9b85-35900a4d4c0f", name = "CategoryName", description = "Relative energy use per ton of process feed.", path = "\\\\\\\\MyAssetServer\\\\Database\\\\CategoriesAnalysis[CategoryName]")

}
