% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisRule.r
\name{PIAnalysisRule}
\alias{PIAnalysisRule}
\title{Generate an instance of the PIAnalysisRule PI Web API class}
\usage{
PIAnalysisRule(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, configString = NULL, displayString = NULL,
  editorType = NULL, hasChildren = NULL, isConfigured = NULL,
  isInitializing = NULL, plugInName = NULL, supportedBehaviors = NULL,
  variableMapping = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{configString}{(string)}

\item{displayString}{(string)}

\item{editorType}{(string)}

\item{hasChildren}{(boolean)}

\item{isConfigured}{(boolean)}

\item{isInitializing}{(boolean)}

\item{plugInName}{(string)}

\item{supportedBehaviors}{(array)}

\item{variableMapping}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAnalysisRule
}
\description{
Generate an instance of the PIAnalysisRule PI Web API class
}
\examples{
analysisRule <- PIAnalysisRule(webId = "I1XRXDqD5loBNH0erqeqJodtALAfyWdysKU5xGEQwAVXYTCAAfyWdysKU5xGEQwAVXYTCAA", id = "ca9d257f-94c2-11e7-8443-00155d84c200", name = "PerformanceEquation", description = "Runs a performance equation.", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\MyElement\\\\Analyses[MyAnalysis]\\\\AnalysisRule", configString = "a := TagVal('sinusoid'); b := Pow(a, 2); c := a + b;", displayString = "a := TagVal('sinusoid'); b := Pow(a, 2); c := a + b;", editorType = "OSIsoft.AF.Analysis.AnalysisRule.PEAnalysisRuleEditor", hasChildren = FALSE, isConfigured = FALSE, isInitializing = FALSE, plugInName = "PerformanceEquation", variableMapping = "b||Attribute1;c||Attribute2")

}
