% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnalysisTemplate.r
\name{PIAnalysisTemplate}
\alias{PIAnalysisTemplate}
\title{Generate an instance of the PIAnalysisTemplate PI Web API class}
\usage{
PIAnalysisTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, analysisRulePlugInName = NULL,
  categoryNames = NULL, createEnabled = NULL, groupId = NULL,
  hasNotificationTemplate = NULL, hasTarget = NULL, outputTime = NULL,
  targetName = NULL, timeRulePlugInName = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{analysisRulePlugInName}{(string)}

\item{categoryNames}{(array)}

\item{createEnabled}{(boolean)}

\item{groupId}{(integer)}

\item{hasNotificationTemplate}{(boolean)}

\item{hasTarget}{(boolean)}

\item{outputTime}{(string)}

\item{targetName}{(string)}

\item{timeRulePlugInName}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAnalysisTemplate
}
\description{
Generate an instance of the PIAnalysisTemplate PI Web API class
}
\examples{
analysisTemplate <- PIAnalysisTemplate(webId = "I1XTG_auSSsvuECG8ad_p8b25QEZgtYQY_J06YnELl5cALiA", id = "612d9811-3f06-4e27-989c-42e5e5c00b88", name = "MyAnalysisTemplate", description = "", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\AnalysisTemplates[MyAnalysisTemplate]", analysisRulePlugInName = "PerformanceEquation", createEnabled = FALSE, groupId = 0, hasNotificationTemplate = FALSE, hasTarget = FALSE, outputTime = "", targetName = "MyElementTemplate", timeRulePlugInName = "Periodic")

}
