% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAnnotation.r
\name{PIAnnotation}
\alias{PIAnnotation}
\title{Generate an instance of the PIAnnotation PI Web API class}
\usage{
PIAnnotation(id = NULL, name = NULL, description = NULL, value = NULL,
  creator = NULL, creationDate = NULL, modifier = NULL,
  modifyDate = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{value}{(object)}

\item{creator}{(string)}

\item{creationDate}{(string)}

\item{modifier}{(string)}

\item{modifyDate}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAnnotation
}
\description{
Generate an instance of the PIAnnotation PI Web API class
}
\examples{
annotation <- PIAnnotation(id = "512B6616-CE39-4F70-9048-8C6A025FB592", name = "Temperature Annotation", description = "Signifies a spike in temperature.", creator = "MyDomain\\\\UserA", creationDate = "6/21/2016 2:45:50 PM", modifier = "MyDomain\\\\UserA", modifyDate = "6/21/2016 2:45:50 PM")

}
