% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAssetDatabase.r
\name{PIAssetDatabase}
\alias{PIAssetDatabase}
\title{Generate an instance of the PIAssetDatabase PI Web API class}
\usage{
PIAssetDatabase(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, extendedProperties = NULL,
  links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{extendedProperties}{(object)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAssetDatabase
}
\description{
Generate an instance of the PIAssetDatabase PI Web API class
}
\examples{
assetDatabase <- PIAssetDatabase(webId = "I1RDDqD5loBNH0erqeqJodtALAquulo6433EKdHra7fsmL0g", id = "a3a5ebaa-37ae-42dc-9d1e-b6bb7ec98bd2", name = "MyDatabase", description = "PI BI Project Asset Model", path = "\\\\\\\\MyAssetServer\\\\MyDatabase")

}
