% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAssetServer.r
\name{PIAssetServer}
\alias{PIAssetServer}
\title{Generate an instance of the PIAssetServer PI Web API class}
\usage{
PIAssetServer(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, isConnected = NULL, serverVersion = NULL,
  serverTime = NULL, extendedProperties = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{isConnected}{(boolean)}

\item{serverVersion}{(string)}

\item{serverTime}{(string)}

\item{extendedProperties}{(object)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAssetServer
}
\description{
Generate an instance of the PIAssetServer PI Web API class
}
\examples{
assetServer <- PIAssetServer(webId = "I1RSDqD5loBNH0erqeqJodtALA", id = "96f9a00e-4d80-471f-aba9-ea89a1db402c", name = "MyAssetServer", description = "Sample Description", path = "\\\\\\\\MyAssetServer", isConnected = FALSE, serverVersion = "2.5.0.5039", serverTime = "7/22/2014 2:00:00 PM")

}
