% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttribute.r
\name{PIAttribute}
\alias{PIAttribute}
\title{Generate an instance of the PIAttribute PI Web API class}
\usage{
PIAttribute(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, type = NULL, typeQualifier = NULL,
  defaultUnitsName = NULL, displayDigits = NULL,
  dataReferencePlugIn = NULL, configString = NULL,
  isConfigurationItem = NULL, isExcluded = NULL, isHidden = NULL,
  isManualDataEntry = NULL, hasChildren = NULL, categoryNames = NULL,
  step = NULL, traitName = NULL, defaultUnitsNameAbbreviation = NULL,
  span = NULL, zero = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{type}{(string)}

\item{typeQualifier}{(string)}

\item{defaultUnitsName}{(string)}

\item{displayDigits}{(integer)}

\item{dataReferencePlugIn}{(string)}

\item{configString}{(string)}

\item{isConfigurationItem}{(boolean)}

\item{isExcluded}{(boolean)}

\item{isHidden}{(boolean)}

\item{isManualDataEntry}{(boolean)}

\item{hasChildren}{(boolean)}

\item{categoryNames}{(array)}

\item{step}{(boolean)}

\item{traitName}{(string)}

\item{defaultUnitsNameAbbreviation}{(string)}

\item{span}{(number)}

\item{zero}{(number)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAttribute
}
\description{
Generate an instance of the PIAttribute PI Web API class
}
\examples{
attribute <- PIAttribute(webId = "I1AbEDqD5loBNH0erqeqJodtALAYIKyyz2F5BGAxQAVXYRDBAGyPedZG1sUmxOOclp3Flwg", id = "75de231b-b591-49b1-b138-e725a77165c2", name = "Water", description = "2008 Water Use", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\CityName\\\\EngineeringProcess\\\\Equipment\\\\MachineName|Water(2008)", type = "Int32", typeQualifier = "", defaultUnitsName = "liter", displayDigits = -5, dataReferencePlugIn = "Table Lookup", configString = "SELECT [Water Use] FROM [Energy Use 2008] WHERE [Asset ID] = '\%Element\%'", isConfigurationItem = FALSE, isExcluded = FALSE, isHidden = FALSE, isManualDataEntry = FALSE, hasChildren = FALSE, step = FALSE, traitName = "LimitLoLo", defaultUnitsNameAbbreviation = "L")

}
