% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeTemplate.r
\name{PIAttributeTemplate}
\alias{PIAttributeTemplate}
\title{Generate an instance of the PIAttributeTemplate PI Web API class}
\usage{
PIAttributeTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, type = NULL, typeQualifier = NULL,
  defaultUnitsName = NULL, defaultValue = NULL,
  dataReferencePlugIn = NULL, configString = NULL,
  isConfigurationItem = NULL, isExcluded = NULL, isHidden = NULL,
  isManualDataEntry = NULL, hasChildren = NULL, categoryNames = NULL,
  traitName = NULL, defaultUnitsNameAbbreviation = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{type}{(string)}

\item{typeQualifier}{(string)}

\item{defaultUnitsName}{(string)}

\item{defaultValue}{(object)}

\item{dataReferencePlugIn}{(string)}

\item{configString}{(string)}

\item{isConfigurationItem}{(boolean)}

\item{isExcluded}{(boolean)}

\item{isHidden}{(boolean)}

\item{isManualDataEntry}{(boolean)}

\item{hasChildren}{(boolean)}

\item{categoryNames}{(array)}

\item{traitName}{(string)}

\item{defaultUnitsNameAbbreviation}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAttributeTemplate
}
\description{
Generate an instance of the PIAttributeTemplate PI Web API class
}
\examples{
attributeTemplate <- PIAttributeTemplate(webId = "I1ATEG_auSSsvuECG8ad_p8b25QQkxqWDwIWU6zC4vmgpd4kgtSfQI9VdxUGA8fi1yf9DVg", id = "23d027b5-5dd5-41c5-80f1-f8b5c9ff4356", name = "Water(2008)", description = "2008 Water Use", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\ElementTemplates[MachineName]|Water(2008)", type = "Int32", typeQualifier = "", defaultUnitsName = "liter", dataReferencePlugIn = "Table Lookup", configString = "SELECT [Water Use] FROM [Energy Use 2008] WHERE [Asset ID] = '\%Element\%'", isConfigurationItem = FALSE, isExcluded = FALSE, isHidden = FALSE, isManualDataEntry = FALSE, hasChildren = FALSE, traitName = "LimitLoLo", defaultUnitsNameAbbreviation = "L")

}
