% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIAttributeTrait.r
\name{PIAttributeTrait}
\alias{PIAttributeTrait}
\title{Generate an instance of the PIAttributeTrait PI Web API class}
\usage{
PIAttributeTrait(name = NULL, abbreviation = NULL,
  allowChildAttributes = NULL, allowDuplicates = NULL,
  isAllowedOnRootAttribute = NULL, isTypeInherited = NULL,
  isUOMInherited = NULL, requireNumeric = NULL, requireString = NULL,
  links = NULL, webException = NULL)
}
\arguments{
\item{name}{(string)}

\item{abbreviation}{(string)}

\item{allowChildAttributes}{(boolean)}

\item{allowDuplicates}{(boolean)}

\item{isAllowedOnRootAttribute}{(boolean)}

\item{isTypeInherited}{(boolean)}

\item{isUOMInherited}{(boolean)}

\item{requireNumeric}{(boolean)}

\item{requireString}{(boolean)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIAttributeTrait
}
\description{
Generate an instance of the PIAttributeTrait PI Web API class
}
\examples{
attributeTrait <- PIAttributeTrait(name = "LimitLoLo", abbreviation = "LoLo", allowChildAttributes = FALSE, allowDuplicates = FALSE, isAllowedOnRootAttribute = FALSE, isTypeInherited = FALSE, isUOMInherited = FALSE, requireNumeric = FALSE, requireString = FALSE)

}
