% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PICacheInstance.r
\name{PICacheInstance}
\alias{PICacheInstance}
\title{Generate an instance of the PICacheInstance PI Web API class}
\usage{
PICacheInstance(id = NULL, lastRefreshTime = NULL,
  willRefreshAfter = NULL, scheduledExpirationTime = NULL, user = NULL,
  webException = NULL)
}
\arguments{
\item{id}{(string)}

\item{lastRefreshTime}{(string)}

\item{willRefreshAfter}{(string)}

\item{scheduledExpirationTime}{(string)}

\item{user}{(string)}

\item{webException}{()}
}
\value{
PICacheInstance
}
\description{
Generate an instance of the PICacheInstance PI Web API class
}
\examples{
cacheInstance <- PICacheInstance(id = "8a5033c9-b77f-431c-9820-c9f538d80dc6", lastRefreshTime = "7/20/2014 2:08:00 PM", willRefreshAfter = "7/20/2014 6:08:00 PM", scheduledExpirationTime = "7/20/2014 11:08:00 PM", user = "MyDomain\\\\Username")

}
