% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElementCategory.r
\name{PIElementCategory}
\alias{PIElementCategory}
\title{Generate an instance of the PIElementCategory PI Web API class}
\usage{
PIElementCategory(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIElementCategory
}
\description{
Generate an instance of the PIElementCategory PI Web API class
}
\examples{
elementCategory <- PIElementCategory(webId = "I1ECDqD5loBNH0erqeqJodtALAQ_lRME1-QUKrnEUKhMgEUA", id = "3051f943-7e4d-4241-ab9c-450a84c80450", name = "CategoryName", description = "Relative energy use per ton of process feed.", path = "\\\\\\\\MyAssetServer\\\\Database\\\\CategoriesElement[CategoryName]")

}
