% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piElementTemplate.r
\name{piElementTemplate}
\alias{piElementTemplate}
\title{Generate an instance of the piElementTemplate PI Web API class}
\usage{
piElementTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, allowElementToExtend = NULL,
  baseTemplate = NULL, instanceType = NULL, namingPattern = NULL,
  categoryNames = NULL, extendedProperties = NULL, severity = NULL,
  canBeAcknowledged = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{allowElementToExtend}{(boolean)}

\item{baseTemplate}{(string)}

\item{instanceType}{(string)}

\item{namingPattern}{(string)}

\item{categoryNames}{(array)}

\item{extendedProperties}{(object)}

\item{severity}{(string)}

\item{canBeAcknowledged}{(boolean)}

\item{links}{(object)}
}
\value{
piElementTemplate
}
\description{
Generate an instance of the piElementTemplate PI Web API class
}
\examples{
elementTemplate <- piElementTemplate(webId = "T0DqD5loBNH0erqeqJodtALAvdrykzvWo0SXjjgcl9gTqAkUVTVFVOSVRcU1VHQVJ5UQNcRUxFTUVOVFRFTVBMQVRFU1t1TFRFUlBSSVNFXQ", id = "74646B96-A130-40A3-AA25-49396C9D6B3C", name = "Boiler", description = "Boiler", path = "\\\\\\\\MyServer\\\\NuGreen\\\\ElementTemplates[Boiler]", allowElementToExtend = FALSE, baseTemplate = "Equipment", instanceType = "Element", namingPattern = "\%TEMPLATE\%", severity = "None", canBeAcknowledged = FALSE)

}
