% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIElementTemplate.r
\name{PIElementTemplate}
\alias{PIElementTemplate}
\title{Generate an instance of the PIElementTemplate PI Web API class}
\usage{
PIElementTemplate(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, allowElementToExtend = NULL,
  baseTemplate = NULL, instanceType = NULL, namingPattern = NULL,
  categoryNames = NULL, extendedProperties = NULL, severity = NULL,
  canBeAcknowledged = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{allowElementToExtend}{(boolean)}

\item{baseTemplate}{(string)}

\item{instanceType}{(string)}

\item{namingPattern}{(string)}

\item{categoryNames}{(array)}

\item{extendedProperties}{(object)}

\item{severity}{(string)}

\item{canBeAcknowledged}{(boolean)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIElementTemplate
}
\description{
Generate an instance of the PIElementTemplate PI Web API class
}
\examples{
elementTemplate <- PIElementTemplate(webId = "I1ETDqD5loBNH0erqeqJodtALAkpSYp6uykE2Ku0yChDU91g", id = "a7989492-b2ab-4d90-8abb-4c8284353dd6", name = "MachineName", description = "Manufacturing Machine MachineName", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\ElementTemplates[Boiler]", allowElementToExtend = FALSE, baseTemplate = "Equipment", instanceType = "Element", namingPattern = "\%TEMPLATE\%", severity = "None", canBeAcknowledged = FALSE)

}
