% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEnumerationSet.r
\name{PIEnumerationSet}
\alias{PIEnumerationSet}
\title{Generate an instance of the PIEnumerationSet PI Web API class}
\usage{
PIEnumerationSet(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, links = NULL,
  serializeDescription = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{links}{()}

\item{serializeDescription}{(boolean)}

\item{webException}{()}
}
\value{
PIEnumerationSet
}
\description{
Generate an instance of the PIEnumerationSet PI Web API class
}
\examples{
enumerationSet <- PIEnumerationSet(webId = "I1MSRDqD5loBNH0erqeqJodtALAT_x3jpGsKUCB1vtmvQHUMQ", id = "8e77fc4f-ac91-4029-81d6-fb66bd01d431", name = "Model Number", description = "Model numbers by brand of vehicle", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\EnumerationSets[Model Number]", serializeDescription = FALSE)

}
