% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIEnumerationValue.r
\name{PIEnumerationValue}
\alias{PIEnumerationValue}
\title{Generate an instance of the PIEnumerationValue PI Web API class}
\usage{
PIEnumerationValue(webId = NULL, id = NULL, name = NULL,
  description = NULL, value = NULL, path = NULL, parent = NULL,
  links = NULL, serializeWebId = NULL, serializeId = NULL,
  serializeDescription = NULL, serializePath = NULL,
  serializeLinks = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{value}{(integer)}

\item{path}{(string)}

\item{parent}{(string)}

\item{links}{()}

\item{serializeWebId}{(boolean)}

\item{serializeId}{(boolean)}

\item{serializeDescription}{(boolean)}

\item{serializePath}{(boolean)}

\item{serializeLinks}{(boolean)}

\item{webException}{()}
}
\value{
PIEnumerationValue
}
\description{
Generate an instance of the PIEnumerationValue PI Web API class
}
\examples{
enumerationValue <- PIEnumerationValue(webId = "I1MVRDqD5loBNH0erqeqJodtALAT_x3jpGsKUCB1vtmvQHUMQlIYqmOlvs0ygEQnSeGSe7w", id = "982a8694-6fe9-4cb3-a011-09d278649eef", name = "BMW|M3", description = "Model Number of BMW", value = 2005, path = "\\\\\\\\MyServer\\\\NuGreen\\\\EnumerationSets[Model Number]\\\\BMW|M3", parent = "BMW", serializeWebId = FALSE, serializeId = FALSE, serializeDescription = FALSE, serializePath = FALSE, serializeLinks = FALSE)

}
