% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIErrors.r
\name{PIErrors}
\alias{PIErrors}
\title{Generate an instance of the PIErrors PI Web API class}
\usage{
PIErrors(errors = NULL)
}
\arguments{
\item{errors}{(array)}
}
\value{
PIErrors
}
\description{
Generate an instance of the PIErrors PI Web API class
}
