% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemAttribute.r
\name{PIItemAttribute}
\alias{PIItemAttribute}
\title{Generate an instance of the PIItemAttribute PI Web API class}
\usage{
PIItemAttribute(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
PIItemAttribute
}
\description{
Generate an instance of the PIItemAttribute PI Web API class
}
\examples{
itemAttribute <- PIItemAttribute(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
