% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemElement.r
\name{PIItemElement}
\alias{PIItemElement}
\title{Generate an instance of the PIItemElement PI Web API class}
\usage{
PIItemElement(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
PIItemElement
}
\description{
Generate an instance of the PIItemElement PI Web API class
}
\examples{
itemElement <- PIItemElement(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
