% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemEventFrame.r
\name{PIItemEventFrame}
\alias{PIItemEventFrame}
\title{Generate an instance of the PIItemEventFrame PI Web API class}
\usage{
PIItemEventFrame(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
PIItemEventFrame
}
\description{
Generate an instance of the PIItemEventFrame PI Web API class
}
\examples{
itemEventFrame <- PIItemEventFrame(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
