% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemPoint.r
\name{PIItemPoint}
\alias{PIItemPoint}
\title{Generate an instance of the PIItemPoint PI Web API class}
\usage{
PIItemPoint(identifier = NULL, identifierType = NULL, object = NULL,
  exception = NULL)
}
\arguments{
\item{identifier}{(string)}

\item{identifierType}{(string)}

\item{object}{()}

\item{exception}{()}
}
\value{
PIItemPoint
}
\description{
Generate an instance of the PIItemPoint PI Web API class
}
\examples{
itemPoint <- PIItemPoint(identifier = "\\\\\\\\server\\\\path", identifierType = "Path")

}
