% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAnalysis.r
\name{PIItemsAnalysis}
\alias{PIItemsAnalysis}
\title{Generate an instance of the PIItemsAnalysis PI Web API class}
\usage{
PIItemsAnalysis(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAnalysis
}
\description{
Generate an instance of the PIItemsAnalysis PI Web API class
}
