% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAnalysisCategory.r
\name{PIItemsAnalysisCategory}
\alias{PIItemsAnalysisCategory}
\title{Generate an instance of the PIItemsAnalysisCategory PI Web API class}
\usage{
PIItemsAnalysisCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAnalysisCategory
}
\description{
Generate an instance of the PIItemsAnalysisCategory PI Web API class
}
