% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAnalysisRule.r
\name{PIItemsAnalysisRule}
\alias{PIItemsAnalysisRule}
\title{Generate an instance of the PIItemsAnalysisRule PI Web API class}
\usage{
PIItemsAnalysisRule(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAnalysisRule
}
\description{
Generate an instance of the PIItemsAnalysisRule PI Web API class
}
