% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAnalysisRulePlugIn.r
\name{PIItemsAnalysisRulePlugIn}
\alias{PIItemsAnalysisRulePlugIn}
\title{Generate an instance of the PIItemsAnalysisRulePlugIn PI Web API class}
\usage{
PIItemsAnalysisRulePlugIn(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAnalysisRulePlugIn
}
\description{
Generate an instance of the PIItemsAnalysisRulePlugIn PI Web API class
}
