% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAnalysisTemplate.r
\name{PIItemsAnalysisTemplate}
\alias{PIItemsAnalysisTemplate}
\title{Generate an instance of the PIItemsAnalysisTemplate PI Web API class}
\usage{
PIItemsAnalysisTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAnalysisTemplate
}
\description{
Generate an instance of the PIItemsAnalysisTemplate PI Web API class
}
