% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAnnotation.r
\name{PIItemsAnnotation}
\alias{PIItemsAnnotation}
\title{Generate an instance of the PIItemsAnnotation PI Web API class}
\usage{
PIItemsAnnotation(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAnnotation
}
\description{
Generate an instance of the PIItemsAnnotation PI Web API class
}
