% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAssetDatabase.r
\name{PIItemsAssetDatabase}
\alias{PIItemsAssetDatabase}
\title{Generate an instance of the PIItemsAssetDatabase PI Web API class}
\usage{
PIItemsAssetDatabase(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAssetDatabase
}
\description{
Generate an instance of the PIItemsAssetDatabase PI Web API class
}
