% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piItemsAssetServer.r
\name{piItemsAssetServer}
\alias{piItemsAssetServer}
\title{Generate an instance of the piItemsAssetServer PI Web API class}
\usage{
piItemsAssetServer(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{(object)}
}
\value{
piItemsAssetServer
}
\description{
Generate an instance of the piItemsAssetServer PI Web API class
}
