% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAssetServer.r
\name{PIItemsAssetServer}
\alias{PIItemsAssetServer}
\title{Generate an instance of the PIItemsAssetServer PI Web API class}
\usage{
PIItemsAssetServer(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAssetServer
}
\description{
Generate an instance of the PIItemsAssetServer PI Web API class
}
