% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAttribute.r
\name{PIItemsAttribute}
\alias{PIItemsAttribute}
\title{Generate an instance of the PIItemsAttribute PI Web API class}
\usage{
PIItemsAttribute(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAttribute
}
\description{
Generate an instance of the PIItemsAttribute PI Web API class
}
