% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAttributeCategory.r
\name{PIItemsAttributeCategory}
\alias{PIItemsAttributeCategory}
\title{Generate an instance of the PIItemsAttributeCategory PI Web API class}
\usage{
PIItemsAttributeCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAttributeCategory
}
\description{
Generate an instance of the PIItemsAttributeCategory PI Web API class
}
