% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAttributeTemplate.r
\name{PIItemsAttributeTemplate}
\alias{PIItemsAttributeTemplate}
\title{Generate an instance of the PIItemsAttributeTemplate PI Web API class}
\usage{
PIItemsAttributeTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAttributeTemplate
}
\description{
Generate an instance of the PIItemsAttributeTemplate PI Web API class
}
