% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsAttributeTrait.r
\name{PIItemsAttributeTrait}
\alias{PIItemsAttributeTrait}
\title{Generate an instance of the PIItemsAttributeTrait PI Web API class}
\usage{
PIItemsAttributeTrait(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsAttributeTrait
}
\description{
Generate an instance of the PIItemsAttributeTrait PI Web API class
}
