% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsCacheInstance.r
\name{PIItemsCacheInstance}
\alias{PIItemsCacheInstance}
\title{Generate an instance of the PIItemsCacheInstance PI Web API class}
\usage{
PIItemsCacheInstance(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsCacheInstance
}
\description{
Generate an instance of the PIItemsCacheInstance PI Web API class
}
