% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsDataServer.r
\name{PIItemsDataServer}
\alias{PIItemsDataServer}
\title{Generate an instance of the PIItemsDataServer PI Web API class}
\usage{
PIItemsDataServer(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsDataServer
}
\description{
Generate an instance of the PIItemsDataServer PI Web API class
}
