% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsElement.r
\name{PIItemsElement}
\alias{PIItemsElement}
\title{Generate an instance of the PIItemsElement PI Web API class}
\usage{
PIItemsElement(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsElement
}
\description{
Generate an instance of the PIItemsElement PI Web API class
}
