% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsElementCategory.r
\name{PIItemsElementCategory}
\alias{PIItemsElementCategory}
\title{Generate an instance of the PIItemsElementCategory PI Web API class}
\usage{
PIItemsElementCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsElementCategory
}
\description{
Generate an instance of the PIItemsElementCategory PI Web API class
}
