% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsElementTemplate.r
\name{PIItemsElementTemplate}
\alias{PIItemsElementTemplate}
\title{Generate an instance of the PIItemsElementTemplate PI Web API class}
\usage{
PIItemsElementTemplate(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsElementTemplate
}
\description{
Generate an instance of the PIItemsElementTemplate PI Web API class
}
