% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsEnumerationValue.r
\name{PIItemsEnumerationValue}
\alias{PIItemsEnumerationValue}
\title{Generate an instance of the PIItemsEnumerationValue PI Web API class}
\usage{
PIItemsEnumerationValue(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsEnumerationValue
}
\description{
Generate an instance of the PIItemsEnumerationValue PI Web API class
}
