% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsEventFrame.r
\name{PIItemsEventFrame}
\alias{PIItemsEventFrame}
\title{Generate an instance of the PIItemsEventFrame PI Web API class}
\usage{
PIItemsEventFrame(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsEventFrame
}
\description{
Generate an instance of the PIItemsEventFrame PI Web API class
}
