% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsItemAttribute.r
\name{PIItemsItemAttribute}
\alias{PIItemsItemAttribute}
\title{Generate an instance of the PIItemsItemAttribute PI Web API class}
\usage{
PIItemsItemAttribute(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsItemAttribute
}
\description{
Generate an instance of the PIItemsItemAttribute PI Web API class
}
