% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsItemElement.r
\name{PIItemsItemElement}
\alias{PIItemsItemElement}
\title{Generate an instance of the PIItemsItemElement PI Web API class}
\usage{
PIItemsItemElement(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsItemElement
}
\description{
Generate an instance of the PIItemsItemElement PI Web API class
}
