% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsItemEventFrame.r
\name{PIItemsItemEventFrame}
\alias{PIItemsItemEventFrame}
\title{Generate an instance of the PIItemsItemEventFrame PI Web API class}
\usage{
PIItemsItemEventFrame(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsItemEventFrame
}
\description{
Generate an instance of the PIItemsItemEventFrame PI Web API class
}
