% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsItemPoint.r
\name{PIItemsItemPoint}
\alias{PIItemsItemPoint}
\title{Generate an instance of the PIItemsItemPoint PI Web API class}
\usage{
PIItemsItemPoint(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsItemPoint
}
\description{
Generate an instance of the PIItemsItemPoint PI Web API class
}
