% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsItemsSubstatus.r
\name{PIItemsItemsSubstatus}
\alias{PIItemsItemsSubstatus}
\title{Generate an instance of the PIItemsItemsSubstatus PI Web API class}
\usage{
PIItemsItemsSubstatus(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsItemsSubstatus
}
\description{
Generate an instance of the PIItemsItemsSubstatus PI Web API class
}
