% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsPoint.r
\name{PIItemsPoint}
\alias{PIItemsPoint}
\title{Generate an instance of the PIItemsPoint PI Web API class}
\usage{
PIItemsPoint(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsPoint
}
\description{
Generate an instance of the PIItemsPoint PI Web API class
}
