% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsPointAttribute.r
\name{PIItemsPointAttribute}
\alias{PIItemsPointAttribute}
\title{Generate an instance of the PIItemsPointAttribute PI Web API class}
\usage{
PIItemsPointAttribute(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsPointAttribute
}
\description{
Generate an instance of the PIItemsPointAttribute PI Web API class
}
