% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsSecurityEntry.r
\name{PIItemsSecurityEntry}
\alias{PIItemsSecurityEntry}
\title{Generate an instance of the PIItemsSecurityEntry PI Web API class}
\usage{
PIItemsSecurityEntry(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsSecurityEntry
}
\description{
Generate an instance of the PIItemsSecurityEntry PI Web API class
}
