% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsSecurityIdentity.r
\name{PIItemsSecurityIdentity}
\alias{PIItemsSecurityIdentity}
\title{Generate an instance of the PIItemsSecurityIdentity PI Web API class}
\usage{
PIItemsSecurityIdentity(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsSecurityIdentity
}
\description{
Generate an instance of the PIItemsSecurityIdentity PI Web API class
}
