% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsSecurityMapping.r
\name{PIItemsSecurityMapping}
\alias{PIItemsSecurityMapping}
\title{Generate an instance of the PIItemsSecurityMapping PI Web API class}
\usage{
PIItemsSecurityMapping(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsSecurityMapping
}
\description{
Generate an instance of the PIItemsSecurityMapping PI Web API class
}
