% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsSecurityRights.r
\name{PIItemsSecurityRights}
\alias{PIItemsSecurityRights}
\title{Generate an instance of the PIItemsSecurityRights PI Web API class}
\usage{
PIItemsSecurityRights(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsSecurityRights
}
\description{
Generate an instance of the PIItemsSecurityRights PI Web API class
}
