% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsStreamSummaries.r
\name{PIItemsStreamSummaries}
\alias{PIItemsStreamSummaries}
\title{Generate an instance of the PIItemsStreamSummaries PI Web API class}
\usage{
PIItemsStreamSummaries(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsStreamSummaries
}
\description{
Generate an instance of the PIItemsStreamSummaries PI Web API class
}
