% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsStreamValue.r
\name{PIItemsStreamValue}
\alias{PIItemsStreamValue}
\title{Generate an instance of the PIItemsStreamValue PI Web API class}
\usage{
PIItemsStreamValue(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsStreamValue
}
\description{
Generate an instance of the PIItemsStreamValue PI Web API class
}
