% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsStreamValues.r
\name{PIItemsStreamValues}
\alias{PIItemsStreamValues}
\title{Generate an instance of the PIItemsStreamValues PI Web API class}
\usage{
PIItemsStreamValues(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsStreamValues
}
\description{
Generate an instance of the PIItemsStreamValues PI Web API class
}
