% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsSubstatus.r
\name{PIItemsSubstatus}
\alias{PIItemsSubstatus}
\title{Generate an instance of the PIItemsSubstatus PI Web API class}
\usage{
PIItemsSubstatus(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsSubstatus
}
\description{
Generate an instance of the PIItemsSubstatus PI Web API class
}
