% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsSummaryValue.r
\name{PIItemsSummaryValue}
\alias{PIItemsSummaryValue}
\title{Generate an instance of the PIItemsSummaryValue PI Web API class}
\usage{
PIItemsSummaryValue(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsSummaryValue
}
\description{
Generate an instance of the PIItemsSummaryValue PI Web API class
}
