% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsTable.r
\name{PIItemsTable}
\alias{PIItemsTable}
\title{Generate an instance of the PIItemsTable PI Web API class}
\usage{
PIItemsTable(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsTable
}
\description{
Generate an instance of the PIItemsTable PI Web API class
}
