% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsTableCategory.r
\name{PIItemsTableCategory}
\alias{PIItemsTableCategory}
\title{Generate an instance of the PIItemsTableCategory PI Web API class}
\usage{
PIItemsTableCategory(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsTableCategory
}
\description{
Generate an instance of the PIItemsTableCategory PI Web API class
}
