% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsTimeRulePlugIn.r
\name{PIItemsTimeRulePlugIn}
\alias{PIItemsTimeRulePlugIn}
\title{Generate an instance of the PIItemsTimeRulePlugIn PI Web API class}
\usage{
PIItemsTimeRulePlugIn(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsTimeRulePlugIn
}
\description{
Generate an instance of the PIItemsTimeRulePlugIn PI Web API class
}
