% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIItemsUnitClass.r
\name{PIItemsUnitClass}
\alias{PIItemsUnitClass}
\title{Generate an instance of the PIItemsUnitClass PI Web API class}
\usage{
PIItemsUnitClass(items = NULL, links = NULL)
}
\arguments{
\item{items}{(array)}

\item{links}{()}
}
\value{
PIItemsUnitClass
}
\description{
Generate an instance of the PIItemsUnitClass PI Web API class
}
