% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PILanding.r
\name{PILanding}
\alias{PILanding}
\title{Generate an instance of the PILanding PI Web API class}
\usage{
PILanding(links = NULL, webException = NULL)
}
\arguments{
\item{links}{()}

\item{webException}{()}
}
\value{
PILanding
}
\description{
Generate an instance of the PILanding PI Web API class
}
