% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piPoint.r
\name{piPoint}
\alias{piPoint}
\title{Generate an instance of the piPoint PI Web API class}
\usage{
piPoint(webId = NULL, id = NULL, name = NULL, path = NULL,
  descriptor = NULL, pointClass = NULL, pointType = NULL,
  digitalSetName = NULL, engineeringUnits = NULL, step = NULL,
  future = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(integer)}

\item{name}{(string)}

\item{path}{(string)}

\item{descriptor}{(string)}

\item{pointClass}{(string)}

\item{pointType}{(string)}

\item{digitalSetName}{(string)}

\item{engineeringUnits}{(string)}

\item{step}{(boolean)}

\item{future}{(boolean)}

\item{links}{(object)}
}
\value{
point
}
\description{
Generate an instance of the piPoint PI Web API class
}
\examples{
point(webId = "P0ZkdQMwx9oUCCzSWmnpgoEwTB8AAAUEhMVlpOXDAwR0FMDjBDRjAwMV9YUTAx", id = 8012, name = "PointName", path = "\\\\\\\\MyPIServer\\\\PointName", descriptor = "12 Hour Sine Wave", pointClass = "classic", pointType = "Float32", digitalSetName = "", engineeringUnits = "", step = FALSE, future = FALSE)

}
