% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPoint.r
\name{PIPoint}
\alias{PIPoint}
\title{Generate an instance of the PIPoint PI Web API class}
\usage{
PIPoint(webId = NULL, id = NULL, name = NULL, path = NULL,
  descriptor = NULL, pointClass = NULL, pointType = NULL,
  digitalSetName = NULL, span = NULL, zero = NULL,
  engineeringUnits = NULL, step = NULL, future = NULL,
  displayDigits = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(integer)}

\item{name}{(string)}

\item{path}{(string)}

\item{descriptor}{(string)}

\item{pointClass}{(string)}

\item{pointType}{(string)}

\item{digitalSetName}{(string)}

\item{span}{(number)}

\item{zero}{(number)}

\item{engineeringUnits}{(string)}

\item{step}{(boolean)}

\item{future}{(boolean)}

\item{displayDigits}{(integer)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIPoint
}
\description{
Generate an instance of the PIPoint PI Web API class
}
\examples{
point <- PIPoint(webId = "I1DPa70Wf0zBA06CLkV9ovNQgQCAAAAA", id = 82, name = "PointName", path = "\\\\\\\\MyPIServer\\\\PointName", descriptor = "12 Hour Sine Wave", pointClass = "classic", pointType = "Float32", digitalSetName = "", engineeringUnits = "", step = FALSE, future = FALSE, displayDigits = -5)

}
