% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIPointAttribute.r
\name{PIPointAttribute}
\alias{PIPointAttribute}
\title{Generate an instance of the PIPointAttribute PI Web API class}
\usage{
PIPointAttribute(name = NULL, value = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{name}{(string)}

\item{value}{(object)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIPointAttribute
}
\description{
Generate an instance of the PIPointAttribute PI Web API class
}
\examples{
pointAttribute <- PIPointAttribute(name = "typicalvalue")

}
