% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIRequest.r
\name{PIRequest}
\alias{PIRequest}
\title{Generate an instance of the PIRequest PI Web API class}
\usage{
PIRequest(method = NULL, resource = NULL, requestTemplate = NULL,
  parameters = NULL, headers = NULL, content = NULL, parentIds = NULL)
}
\arguments{
\item{method}{(string)}

\item{resource}{(string)}

\item{requestTemplate}{()}

\item{parameters}{(array)}

\item{headers}{(object)}

\item{content}{(string)}

\item{parentIds}{(array)}
}
\value{
PIRequest
}
\description{
Generate an instance of the PIRequest PI Web API class
}
\examples{
request <- PIRequest(method = "GET", resource = "", content = "")

}
