% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIRequestTemplate.r
\name{PIRequestTemplate}
\alias{PIRequestTemplate}
\title{Generate an instance of the PIRequestTemplate PI Web API class}
\usage{
PIRequestTemplate(resource = NULL)
}
\arguments{
\item{resource}{(string)}
}
\value{
PIRequestTemplate
}
\description{
Generate an instance of the PIRequestTemplate PI Web API class
}
\examples{
requestTemplate <- PIRequestTemplate(resource = "")

}
