% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIResponse.r
\name{PIResponse}
\alias{PIResponse}
\title{Generate an instance of the PIResponse PI Web API class}
\usage{
PIResponse(status = NULL, headers = NULL, content = NULL)
}
\arguments{
\item{status}{(integer)}

\item{headers}{(object)}

\item{content}{(object)}
}
\value{
PIResponse
}
\description{
Generate an instance of the PIResponse PI Web API class
}
\examples{
response <- PIResponse(status = 200)

}
