% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISearchByAttribute.r
\name{PISearchByAttribute}
\alias{PISearchByAttribute}
\title{Generate an instance of the PISearchByAttribute PI Web API class}
\usage{
PISearchByAttribute(searchRoot = NULL, elementTemplate = NULL,
  webException = NULL, valueQueries = NULL)
}
\arguments{
\item{searchRoot}{(string)}

\item{elementTemplate}{(string)}

\item{webException}{()}

\item{valueQueries}{(array)}
}
\value{
PISearchByAttribute
}
\description{
Generate an instance of the PISearchByAttribute PI Web API class
}
\examples{
searchByAttribute <- PISearchByAttribute(searchRoot = "I1RDDqD5loBNH0erqeqJodtALA8fbgUnyQW02v-gLGIxumSg", elementTemplate = "I1ETDqD5loBNH0erqeqJodtALAjFPVfUpY-02A8uioGDSgIg")

}
