% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityEntry.r
\name{PISecurityEntry}
\alias{PISecurityEntry}
\title{Generate an instance of the PISecurityEntry PI Web API class}
\usage{
PISecurityEntry(name = NULL, securityIdentityName = NULL,
  allowRights = NULL, denyRights = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{name}{(string)}

\item{securityIdentityName}{(string)}

\item{allowRights}{(array)}

\item{denyRights}{(array)}

\item{links}{()}

\item{webException}{()}
}
\value{
PISecurityEntry
}
\description{
Generate an instance of the PISecurityEntry PI Web API class
}
\examples{
securityEntry <- PISecurityEntry(name = "domain\\\\user1", securityIdentityName = "domain\\\\user1")

}
