% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityIdentity.r
\name{PISecurityIdentity}
\alias{PISecurityIdentity}
\title{Generate an instance of the PISecurityIdentity PI Web API class}
\usage{
PISecurityIdentity(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, isEnabled = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{isEnabled}{(boolean)}

\item{links}{()}

\item{webException}{()}
}
\value{
PISecurityIdentity
}
\description{
Generate an instance of the PISecurityIdentity PI Web API class
}
\examples{
securityIdentity <- PISecurityIdentity(webId = "I1SIDqD5loBNH0erqeqJodtALASe6l8zgYokqdeeFilFI9tw", id = "f3a5ee49-1838-4aa2-9d79-e16294523db7", name = "MySecurityIdentity", description = "", path = "\\\\\\\\MyAssetServer\\\\SecurityIdentities[MySecurityIdentity]", isEnabled = FALSE)

}
