% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSecurityMapping.r
\name{piSecurityMapping}
\alias{piSecurityMapping}
\title{Generate an instance of the piSecurityMapping PI Web API class}
\usage{
piSecurityMapping(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, account = NULL,
  securityIdentityWebId = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{account}{(string)}

\item{securityIdentityWebId}{(string)}

\item{links}{(object)}
}
\value{
securityMapping
}
\description{
Generate an instance of the piSecurityMapping PI Web API class
}
\examples{
securityMapping(webId = "M07v5To0Ux7Eus2sjMsdqq_AUEhMQUZTMDU", id = "a353feee-3145-4bec-acda-c8ccb1daaafc", name = "MySecurityMapping", description = "", path = "\\\\\\\\RESTUNIT\\\\SecurityMappings[MySecurityMapping]", account = "domain\\\\user", securityIdentityWebId = "I07v5To0Ux7Eus2sjMsdqq_AUEhMQUZTMDU")

}
