% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISecurityMapping.r
\name{PISecurityMapping}
\alias{PISecurityMapping}
\title{Generate an instance of the PISecurityMapping PI Web API class}
\usage{
PISecurityMapping(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, account = NULL,
  securityIdentityWebId = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{account}{(string)}

\item{securityIdentityWebId}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PISecurityMapping
}
\description{
Generate an instance of the PISecurityMapping PI Web API class
}
\examples{
securityMapping <- PISecurityMapping(webId = "I1SMDqD5loBNH0erqeqJodtALAgu8UrMAZB0qWp9H7C4TAXQ", id = "ac14ef82-19c0-4a07-96a7-d1fb0b84c05d", name = "MySecurityMapping", description = "", path = "\\\\\\\\MyAssetServer\\\\SecurityMappings[MySecurityMapping]", account = "domain\\\\user", securityIdentityWebId = "I1SIEDqD5loBNH0erqeqJodtALAYIKyyz2F5BGAxQAVXYRDBAGyPedZG1sUmxOOclp3Flwg")

}
