% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSecurityRights.r
\name{piSecurityRights}
\alias{piSecurityRights}
\title{Generate an instance of the piSecurityRights PI Web API class}
\usage{
piSecurityRights(ownerWebId = NULL, securityItem = NULL,
  userIdentity = NULL, links = NULL)
}
\arguments{
\item{ownerWebId}{(string)}

\item{securityItem}{(string)}

\item{userIdentity}{(string)}

\item{links}{(object)}
}
\value{
piSecurityRights
}
\description{
Generate an instance of the piSecurityRights PI Web API class
}
\examples{
securityRights <- piSecurityRights(ownerWebId = "E0DqD5loBNH0erqeqJodtALAvGEqktDl4xG-twAFmjx6UAUkVTVFVOSVRUc1VHQVJNQU5V1TcHUkVFTg", securityItem = "Default", userIdentity = "domain\\\\user")

}
