% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamSummaries.r
\name{PIStreamSummaries}
\alias{PIStreamSummaries}
\title{Generate an instance of the PIStreamSummaries PI Web API class}
\usage{
PIStreamSummaries(webId = NULL, name = NULL, path = NULL, items = NULL,
  links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{items}{(array)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIStreamSummaries
}
\description{
Generate an instance of the PIStreamSummaries PI Web API class
}
\examples{
streamSummaries <- PIStreamSummaries(webId = "I1AbEDqD5loBNH0erqeqJodtALAYIKyyz2F5BGAxQAVXYRDBAGyPedZG1sUmxOOclp3Flwg", name = "Water", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\MyElement|Water")

}
