% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piStreamValue.r
\name{piStreamValue}
\alias{piStreamValue}
\title{Generate an instance of the piStreamValue PI Web API class}
\usage{
piStreamValue(webId = NULL, name = NULL, path = NULL, value = NULL,
  links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{value}{()}

\item{links}{(object)}
}
\value{
piStreamValue
}
\description{
Generate an instance of the piStreamValue PI Web API class
}
\examples{
streamValue <- piStreamValue(webId = "A0EDqD5loBNH0erqeqJodtALAUvtkqIDl4xG-twAFmjx6AA79jxF7HiqFwpsgMBxwgWmgUkVTVFVOSVRcU1VNJQVHQU5cTlVHUkVZTlxIT1VTVE9f0FBMQU5UQ0lUWQ", name = "Water", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement|Water")

}
