% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIStreamValue.r
\name{PIStreamValue}
\alias{PIStreamValue}
\title{Generate an instance of the PIStreamValue PI Web API class}
\usage{
PIStreamValue(webId = NULL, name = NULL, path = NULL, value = NULL,
  links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{value}{()}

\item{links}{()}

\item{webException}{()}
}
\value{
PIStreamValue
}
\description{
Generate an instance of the PIStreamValue PI Web API class
}
\examples{
streamValue <- PIStreamValue(webId = "I1AbEDqD5loBNH0erqeqJodtALAYIKyyz2F5BGAxQAVXYRDBAGyPedZG1sUmxOOclp3Flwg", name = "Water", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\MyElement|Water")

}
