% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piStreamValues.r
\name{piStreamValues}
\alias{piStreamValues}
\title{Generate an instance of the piStreamValues PI Web API class}
\usage{
piStreamValues(webId = NULL, name = NULL, path = NULL, items = NULL,
  unitsAbbreviation = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{name}{(string)}

\item{path}{(string)}

\item{items}{(array)}

\item{unitsAbbreviation}{(string)}

\item{links}{(object)}
}
\value{
piStreamValues
}
\description{
Generate an instance of the piStreamValues PI Web API class
}
\examples{
streamValues <- piStreamValues(webId = "A0EDqD5loBNH0erqeqJodtALAUvtkqIDl4xG-twAFmjx6AA79jxF7HiqFwpsgMBxwgWmgUkVTVFVOSVRcU1VNJQVHQU5cTlVHUkVZTlxIT1VTVE9f0FBMQU5UQ0lUWQ", name = "Water", path = "\\\\\\\\MYSERVER\\\\MyDatabase\\\\MyElement|Water", unitsAbbreviation = "m")

}
