% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSubstatus.r
\name{piSubstatus}
\alias{piSubstatus}
\title{Generate an instance of the piSubstatus PI Web API class}
\usage{
piSubstatus(substatus = NULL, message = NULL)
}
\arguments{
\item{substatus}{(integer)}

\item{message}{(string)}
}
\value{
piSubstatus
}
\description{
Generate an instance of the piSubstatus PI Web API class
}
\examples{
substatus <- piSubstatus(substatus = 200, message = "")

}
