% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISubstatus.r
\name{PISubstatus}
\alias{PISubstatus}
\title{Generate an instance of the PISubstatus PI Web API class}
\usage{
PISubstatus(substatus = NULL, message = NULL, webException = NULL)
}
\arguments{
\item{substatus}{(integer)}

\item{message}{(string)}

\item{webException}{()}
}
\value{
PISubstatus
}
\description{
Generate an instance of the PISubstatus PI Web API class
}
\examples{
substatus <- PISubstatus(substatus = 200, message = "")

}
