% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISummaryValue.r
\name{PISummaryValue}
\alias{PISummaryValue}
\title{Generate an instance of the PISummaryValue PI Web API class}
\usage{
PISummaryValue(type = NULL, value = NULL, webException = NULL)
}
\arguments{
\item{type}{(string)}

\item{value}{()}

\item{webException}{()}
}
\value{
PISummaryValue
}
\description{
Generate an instance of the PISummaryValue PI Web API class
}
\examples{
summaryValue <- PISummaryValue(type = "Total")

}
