% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISystemLanding.r
\name{PISystemLanding}
\alias{PISystemLanding}
\title{Generate an instance of the PISystemLanding PI Web API class}
\usage{
PISystemLanding(productTitle = NULL, productVersion = NULL, links = NULL,
  webException = NULL)
}
\arguments{
\item{productTitle}{(string)}

\item{productVersion}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PISystemLanding
}
\description{
Generate an instance of the PISystemLanding PI Web API class
}
\examples{
systemLanding <- PISystemLanding(productTitle = "PI Web API 2016 R2", productVersion = "1.8.0.201")

}
