% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piSystemStatus.r
\name{piSystemStatus}
\alias{piSystemStatus}
\title{Generate an instance of the piSystemStatus PI Web API class}
\usage{
piSystemStatus(upTimeInMinutes = NULL, state = NULL,
  cacheInstances = NULL)
}
\arguments{
\item{upTimeInMinutes}{(number)}

\item{state}{(string)}

\item{cacheInstances}{(integer)}
}
\value{
piSystemStatus
}
\description{
Generate an instance of the piSystemStatus PI Web API class
}
\examples{
systemStatus <- piSystemStatus(state = "Running", cacheInstances = 0)

}
