% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PISystemStatus.r
\name{PISystemStatus}
\alias{PISystemStatus}
\title{Generate an instance of the PISystemStatus PI Web API class}
\usage{
PISystemStatus(upTimeInMinutes = NULL, state = NULL,
  cacheInstances = NULL, serverTime = NULL, webException = NULL)
}
\arguments{
\item{upTimeInMinutes}{(number)}

\item{state}{(string)}

\item{cacheInstances}{(integer)}

\item{serverTime}{(string)}

\item{webException}{()}
}
\value{
PISystemStatus
}
\description{
Generate an instance of the PISystemStatus PI Web API class
}
\examples{
systemStatus <- PISystemStatus(state = "Running", cacheInstances = 0, serverTime = "7/22/2014 2:00:00 PM")

}
