% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITable.r
\name{PITable}
\alias{PITable}
\title{Generate an instance of the PITable PI Web API class}
\usage{
PITable(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, categoryNames = NULL, timeZone = NULL,
  convertToLocalTime = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{categoryNames}{(array)}

\item{timeZone}{(string)}

\item{convertToLocalTime}{(boolean)}

\item{links}{()}

\item{webException}{()}
}
\value{
PITable
}
\description{
Generate an instance of the PITable PI Web API class
}
\examples{
table <- PITable(webId = "I1BlDqD5loBNH0erqeqJodtALAmLr4X86Jmkeynt3QVwlqXw", id = "5ff8ba98-89ce-479a-b29e-ddd057096a5f", name = "CarInfo", description = "Table of car info.", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\Tables[CarInfo]", timeZone = "Eastern Standard Time", convertToLocalTime = FALSE)

}
