% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITableData.r
\name{PITableData}
\alias{PITableData}
\title{Generate an instance of the PITableData PI Web API class}
\usage{
PITableData(columns = NULL, rows = NULL, webException = NULL)
}
\arguments{
\item{columns}{(object)}

\item{rows}{(array)}

\item{webException}{()}
}
\value{
PITableData
}
\description{
Generate an instance of the PITableData PI Web API class
}
