% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piTimedValue.r
\name{piTimedValue}
\alias{piTimedValue}
\title{Generate an instance of the piTimedValue PI Web API class}
\usage{
piTimedValue(timestamp = NULL, unitsAbbreviation = NULL, good = NULL,
  questionable = NULL, substituted = NULL, value = NULL,
  exception = NULL)
}
\arguments{
\item{timestamp}{(string)}

\item{unitsAbbreviation}{(string)}

\item{good}{(boolean)}

\item{questionable}{(boolean)}

\item{substituted}{(boolean)}

\item{value}{(object)}

\item{exception}{()}
}
\value{
timedValue
}
\description{
Generate an instance of the piTimedValue PI Web API class
}
\examples{
timedValue(timestamp = "22/07/2014 14:00:00", unitsAbbreviation = "m", good = FALSE, questionable = FALSE, substituted = FALSE)

}
