% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piTimedValues.r
\name{piTimedValues}
\alias{piTimedValues}
\title{Generate an instance of the piTimedValues PI Web API class}
\usage{
piTimedValues(items = NULL, unitsAbbreviation = NULL)
}
\arguments{
\item{items}{(array)}

\item{unitsAbbreviation}{(string)}
}
\value{
timedValues
}
\description{
Generate an instance of the piTimedValues PI Web API class
}
\examples{
timedValues(unitsAbbreviation = "")

}
