% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PITimeRule.r
\name{PITimeRule}
\alias{PITimeRule}
\title{Generate an instance of the PITimeRule PI Web API class}
\usage{
PITimeRule(webId = NULL, id = NULL, name = NULL, description = NULL,
  path = NULL, configString = NULL, configStringStored = NULL,
  displayString = NULL, editorType = NULL, isConfigured = NULL,
  isInitializing = NULL, mergeDuplicatedItems = NULL, plugInName = NULL,
  links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{configString}{(string)}

\item{configStringStored}{(string)}

\item{displayString}{(string)}

\item{editorType}{(string)}

\item{isConfigured}{(boolean)}

\item{isInitializing}{(boolean)}

\item{mergeDuplicatedItems}{(boolean)}

\item{plugInName}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PITimeRule
}
\description{
Generate an instance of the PITimeRule PI Web API class
}
\examples{
timeRule <- PITimeRule(webId = "I1TRXDqD5loBNH0erqeqJodtALAROsUFcWU5xGEQwAVXYTCAAROsUFcWU5xGEQwAVXYTCAA", id = "1514eb44-94c5-11e7-8443-00155d84c200", name = "Periodic", description = "Creates regular periodic time periods.", path = "\\\\\\\\MyAssetServer\\\\MyDatabase\\\\MyElement\\\\Analyses[MyAnalysis]\\\\TimeRule", configString = "Frequency=300", configStringStored = "", displayString = "Frequency=300", editorType = "OSIsoft.AF.Time.TimeRule.PeriodicConfig", isConfigured = FALSE, isInitializing = FALSE, mergeDuplicatedItems = FALSE, plugInName = "Periodic")

}
