% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piTimeRulePlugIn.r
\name{piTimeRulePlugIn}
\alias{piTimeRulePlugIn}
\title{Generate an instance of the piTimeRulePlugIn PI Web API class}
\usage{
piTimeRulePlugIn(webId = NULL, id = NULL, name = NULL,
  description = NULL, path = NULL, assemblyFileName = NULL,
  assemblyID = NULL, assemblyLoadProperties = NULL, assemblyTime = NULL,
  compatibilityVersion = NULL, isBrowsable = NULL,
  isNonEditableConfig = NULL, loadedAssemblyTime = NULL,
  loadedVersion = NULL, version = NULL, links = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{assemblyFileName}{(string)}

\item{assemblyID}{(string)}

\item{assemblyLoadProperties}{(array)}

\item{assemblyTime}{(string)}

\item{compatibilityVersion}{(integer)}

\item{isBrowsable}{(boolean)}

\item{isNonEditableConfig}{(boolean)}

\item{loadedAssemblyTime}{(string)}

\item{loadedVersion}{(string)}

\item{version}{(string)}

\item{links}{(object)}
}
\value{
timeRulePlugIn
}
\description{
Generate an instance of the piTimeRulePlugIn PI Web API class
}
\examples{
timeRulePlugIn(webId = "W0NxzXSxtlKkGzAhZfHOB-KAjzvCrVcGmUePCcTgk3_HzAUEhMQUZTMDRcUExVR0lOU1RJTUVSVUxFW1BFUklPRElDXQ", id = "adc23b8f-0657-4799-8f09-c4e0937fc7cc", name = "Periodic", description = "Creates regular periodic time periods.", path = "\\\\\\\\MYSERVER\\\\PlugInsTimeRule[Periodic]", assemblyFileName = "OSIsoft.AF.Time.TimeRule.Periodic.dll", assemblyID = "7a635b27-3c3c-4a5f-bba9-01f06e9e525b", assemblyTime = "25/05/2016 14:01:07", compatibilityVersion = 2, isBrowsable = FALSE, isNonEditableConfig = FALSE, loadedAssemblyTime = "01/01/1970 00:00:00", loadedVersion = "", version = "2.8.5.7552")

}
