% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUnit.r
\name{PIUnit}
\alias{PIUnit}
\title{Generate an instance of the PIUnit PI Web API class}
\usage{
PIUnit(webId = NULL, id = NULL, name = NULL, abbreviation = NULL,
  description = NULL, path = NULL, factor = NULL, offset = NULL,
  referenceFactor = NULL, referenceOffset = NULL,
  referenceUnitAbbreviation = NULL, links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{abbreviation}{(string)}

\item{description}{(string)}

\item{path}{(string)}

\item{factor}{(number)}

\item{offset}{(number)}

\item{referenceFactor}{(number)}

\item{referenceOffset}{(number)}

\item{referenceUnitAbbreviation}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIUnit
}
\description{
Generate an instance of the PIUnit PI Web API class
}
\examples{
unit <- PIUnit(webId = "I1UtDqD5loBNH0erqeqJodtALAjqwhgeI8lEeV4xeD1db0_A", id = "8121ac8e-3ce2-4794-95e3-1783d5d6f4fc", name = "Hertz", abbreviation = "Hz", description = "Hertz Unit", path = "\\\\\\\\MyAssetServer\\\\UOMDatabase\\\\Hertz", referenceUnitAbbreviation = "")

}
