% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUnitClass.r
\name{PIUnitClass}
\alias{PIUnitClass}
\title{Generate an instance of the PIUnitClass PI Web API class}
\usage{
PIUnitClass(webId = NULL, id = NULL, name = NULL, description = NULL,
  canonicalUnitName = NULL, canonicalUnitAbbreviation = NULL, path = NULL,
  links = NULL, webException = NULL)
}
\arguments{
\item{webId}{(string)}

\item{id}{(string)}

\item{name}{(string)}

\item{description}{(string)}

\item{canonicalUnitName}{(string)}

\item{canonicalUnitAbbreviation}{(string)}

\item{path}{(string)}

\item{links}{()}

\item{webException}{()}
}
\value{
PIUnitClass
}
\description{
Generate an instance of the PIUnitClass PI Web API class
}
\examples{
unitClass <- PIUnitClass(webId = "I1UCDqD5loBNH0erqeqJodtALATbkl-fxulEulDQAVw5HySQ", id = "f925b94d-6efc-4b94-a50d-0015c391f249", name = "Power", description = "Power Unit Class", canonicalUnitName = "watt", canonicalUnitAbbreviation = "W", path = "\\\\\\\\MyAssetServer\\\\UOMDatabase\\\\UOMClasses[Power]")

}
