% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIUserInfo.r
\name{PIUserInfo}
\alias{PIUserInfo}
\title{Generate an instance of the PIUserInfo PI Web API class}
\usage{
PIUserInfo(identityType = NULL, name = NULL, isAuthenticated = NULL,
  sID = NULL, impersonationLevel = NULL, webException = NULL)
}
\arguments{
\item{identityType}{(string)}

\item{name}{(string)}

\item{isAuthenticated}{(boolean)}

\item{sID}{(string)}

\item{impersonationLevel}{(string)}

\item{webException}{()}
}
\value{
PIUserInfo
}
\description{
Generate an instance of the PIUserInfo PI Web API class
}
\examples{
userInfo <- PIUserInfo(identityType = "WindowsIdentity", name = "MyDomain\\\\Username", isAuthenticated = FALSE, sID = "S-1-5-21-2083424470-1893017325-1469799231-71366", impersonationLevel = "None")

}
