% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIValue.r
\name{PIValue}
\alias{PIValue}
\title{Generate an instance of the PIValue PI Web API class}
\usage{
PIValue(value = NULL, exception = NULL, webException = NULL)
}
\arguments{
\item{value}{(object)}

\item{exception}{()}

\item{webException}{()}
}
\value{
PIValue
}
\description{
Generate an instance of the PIValue PI Web API class
}
