% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PIVersion.r
\name{PIVersion}
\alias{PIVersion}
\title{Generate an instance of the PIVersion PI Web API class}
\usage{
PIVersion(fullVersion = NULL, majorMinorRevision = NULL, build = NULL,
  webException = NULL)
}
\arguments{
\item{fullVersion}{(string)}

\item{majorMinorRevision}{(string)}

\item{build}{(string)}

\item{webException}{()}
}
\value{
PIVersion
}
\description{
Generate an instance of the PIVersion PI Web API class
}
\examples{
version <- PIVersion(fullVersion = "1.10.0.201", majorMinorRevision = "1.10.0", build = "201")

}
