% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pointApi.r
\name{point$update}
\alias{point$update}
\title{Update a point.}
\arguments{
\item{webId}{The ID of the point.}

\item{PIPoint}{A partial point containing the desired changes.}
}
\value{
The point was updated.
}
\description{
Update a point.
}
