% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/streamSetApi.r
\name{streamSet$getChannelAdHoc}
\alias{streamSet$getChannelAdHoc}
\title{Opens a channel that will send messages about any value changes for the specified streams.}
\arguments{
\item{webId}{The ID of a stream. Multiple streams may be specified with multiple instances of the parameter.}

\item{heartbeatRate}{Specifies the maximum number of consecutive empty messages that can be elapsed with no new data updates from the PI System, after which the client receives an empty payload. It helps to check if the connection is still alive. Zero/negative values correspond to no heartbeat, and the default value is no heartbeat.}

\item{includeInitialValues}{Specified if the channel should send a message with the current values of all the streams after the connection is opened. The default is 'false'.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
Switches to the Web Socket protocol.
}
\description{
Opens a channel that will send messages about any value changes for the specified streams.
}
