% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemApi.r
\name{system$landing}
\alias{system$landing}
\title{Get system links for this PI System Web API instance.}
\value{
Links to top level collections and system information.
}
\description{
Get system links for this PI System Web API instance.
}
