% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemApi.r
\name{system$status}
\alias{system$status}
\title{Get information about this PI Web API instance. Examples of information returned include the system uptime, the number of cache instances for this PI System Web API instance, and the system run state.}
\value{
System Status.
}
\description{
Get information about this PI Web API instance. Examples of information returned include the system uptime, the number of cache instances for this PI System Web API instance, and the system run state.
}
