% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$getByPath}
\alias{table$getByPath}
\title{Retrieve a table by path.}
\arguments{
\item{path}{The path to the table.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The specified table.
}
\description{
Retrieve a table by path.
}
