% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tableApi.r
\name{table$update}
\alias{table$update}
\title{Update a table by replacing items in its definition.}
\arguments{
\item{webId}{The ID of the table to update.}

\item{PITable}{A partial table containing the desired changes.}
}
\value{
The table was updated.
}
\description{
Update a table by replacing items in its definition.
}
