% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitApi.r
\name{unit$getByPath}
\alias{unit$getByPath}
\title{Retrieve a unit by path.}
\arguments{
\item{path}{The path to the unit.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
The specified unit.
}
\description{
Retrieve a unit by path.
}
