% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$getCanonicalUnit}
\alias{unitClass$getCanonicalUnit}
\title{Get the canonical unit of a unit class.}
\arguments{
\item{webId}{The ID of unit class.}

\item{selectedFields}{List of fields to be returned in the response, separated by semicolons (;). If this parameter is not specified, all available fields will be returned.}

\item{webIdType}{Optional parameter. Used to specify the type of WebID. Useful for URL brevity and other special cases. Default is the value of the configuration item "WebIDType".}
}
\value{
the canonical unit of the specified unit class.
}
\description{
Get the canonical unit of a unit class.
}
