% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unitClassApi.r
\name{unitClass$update}
\alias{unitClass$update}
\title{Update a unit class.}
\arguments{
\item{webId}{The ID of the unit class.}

\item{PIUnitClass}{A partial unit class containing the desired changes.}
}
\value{
The unit class was updated.
}
\description{
Update a unit class.
}
